Crate polars_core

Modules

chunked_array

The typed heart of every Series column.

datatypes

Data types supported by Polars.

doc

Other documentation

error

export

frame

DataFrame module.

functions

Functions

prelude

Everything you need to get started with Polars.

schema

series

Type agnostic columnar data structure.

testing

Testing utilities.

utils

Macros

apply_method_all_arrow_series

apply_method_physical_integer

apply_method_physical_numeric

df

match_arrow_data_type_apply_macro_ca

Apply a macro on the Downcasted ChunkedArray’s

match_arrow_data_type_apply_macro_ca_logical_num

Apply a macro on the Downcasted ChunkedArray’s of DataTypes that are logical numerics. So no logical.

match_dtype_to_logical_apply_macro

Apply a macro on the Series

match_dtype_to_physical_apply_macro

Apply a macro on the Series

Structs

StrHashGlobal

Statics

POOL

SINGLE_LOCK

Functions

reset_string_cache

Reset the global string cache used for the Categorical Types.

toggle_string_cache

Use a global string cache for the Categorical Types.

with_string_cache