Module polars_core::chunked_array::iterator

Structs

BoolIterNoNull

The no null iterator for a BooleanArray

ListIterNoNull

SomeIterator

Wrapper struct to convert an iterator of type T into one of type Option<T>. It is useful to make the IntoIterator trait, in which every iterator shall return an Option<T>.

Utf8IterNoNull

Traits

IntoNoNullIterator

Trait for ChunkedArrays that don’t have null values. The result is the most efficient implementation Iterator, according to the number of chunks.

PolarsIterator

A PolarsIterator is an iterator over a ChunkedArray which contains polars types. A PolarsIterator must implement ExactSizeIterator and DoubleEndedIterator.