polars_core::chunked_array

Module object

Available on crate feature object only.

Modules

builder

registry

This is a heap allocated utility that can be used to register an object type.

Structs

ObjectArray

Traits

PolarsObject

Values need to implement this so that they can be stored into a Series and DataFrame

PolarsObjectSafe

Trimmed down object safe polars object

Functions

set_polars_allow_extension

Control whether extension types may be created.

Type Aliases

ObjectValueIter