polars_core::chunked_array::object

Module registry

Available on crate feature object only.

This is a heap allocated utility that can be used to register an object type.

That object type will know its own generic type parameter T and callers can simply send &Any values and don’t have to know the generic type themselves.

Structs

ObjectRegistry

Traits

AnonymousObjectBuilder

This trait can be registered, after which that global registration can be used to materialize object types

Functions

get_object_builder

get_object_converter

get_object_physical_type

get_pyobject_converter

register_object_builder

Type Aliases

BuilderConstructor

Takes a name and capacity and constructs a new builder.

ObjectConverter

PyObjectConverter