polars_core::chunked_array::ops::search_sorted

Function lower_bound_chunks

pub fn lower_bound_chunks<'a, T, F>(
    chunks: &[&'a T::Array],
    search_values: impl Iterator<Item = Option<T::Physical<'a>>>,
    null_idx: IdxSize,
    f: F,
) -> Vec<IdxSize> 
where
    T: PolarsDataType,
    F: Fn(&'a T::Array, usize, &T::Physical<'a>) -> bool,

Search through a series of chunks for the first position where f(x) is true, assuming it is first always false and then always true.

It repeats this for each value in search_values. If the search value is null null_idx is returned.

Assumes the chunks are non-empty.