Trait polars_core::chunked_array::ops::ChunkCast

pub trait ChunkCast {
    // Required methods
    fn cast(&self, data_type: &DataType) -> PolarsResult<Series>;
    unsafe fn cast_unchecked(
        &self,
        data_type: &DataType
    ) -> PolarsResult<Series>;
}

Cast ChunkedArray<T> to ChunkedArray<N>

Required Methods

fn cast(&self, data_type: &DataType) -> PolarsResult<Series>

Cast a ChunkedArray to DataType

unsafe fn cast_unchecked(&self, data_type: &DataType) -> PolarsResult<Series>

Does not check if the cast is a valid one and may over/underflow

Safety

This doesn’t do utf8 validation checking when casting from binary

This doesn’t do categorical bound checking when casting from UInt32