Trait polars_core::chunked_array::ops::ChunkPeaks

pub trait ChunkPeaks {
    // Provided methods
    fn peak_max(&self) -> BooleanChunked { ... }
    fn peak_min(&self) -> BooleanChunked { ... }
}

Find local minima/ maxima

Provided Methods

fn peak_max(&self) -> BooleanChunked

Get a boolean mask of the local maximum peaks.

fn peak_min(&self) -> BooleanChunked

Get a boolean mask of the local minimum peaks.