polars_core::chunked_array::ops

Trait ChunkZip

pub trait ChunkZip<T: PolarsDataType> {
    // Required method
    fn zip_with(
        &self,
        mask: &BooleanChunked,
        other: &ChunkedArray<T>,
    ) -> PolarsResult<ChunkedArray<T>>;
}

Combine two ChunkedArray based on some predicate.

Required Methods

fn zip_with( &self, mask: &BooleanChunked, other: &ChunkedArray<T>, ) -> PolarsResult<ChunkedArray<T>>

Create a new ChunkedArray with values from self where the mask evaluates true and values from other where the mask evaluates false