Module polars_core::datatypes

Data types supported by Polars.

At the moment Polars doesn’t include all data types available by Arrow. The goal is to incrementally support more data types and prioritize these by usability.

See the AnyValue variants for the data types that are currently supported.

Structs

BooleanType

CategoricalType

DateType

DatetimeType

DurationType

Field

Characterizes the name and the DataType of a column.

Float32Type

Float64Type

Int8Type

Int16Type

Int32Type

Int64Type

ListType

Logical

Maps a logical type to a a chunked array implementation of the physical type. This saves a lot of compiler bloat and allows us to reuse functionality.

TimeType

UInt8Type

UInt16Type

UInt32Type

UInt64Type

Utf8Type

Enums

AnyValue

ArrowDataType

The set of supported logical types in this crate.

ArrowTimeUnit

The time units defined in Arrow.

DataType

TimeUnit

Constants

IDX_DTYPE

NULL_DTYPE

Traits

InitHashMaps

LogicalType

NumericNative

PolarsDataType

PolarsFloatType

PolarsIntegerType

PolarsNumericType

PolarsSingleType

Any type that is not nested

Functions

merge_dtypes

Type Definitions

BooleanChunked

DateChunked

DatetimeChunked

DurationChunked

Float32Chunked

Float64Chunked

IdxArr

IdxCa

IdxSize

The type used by polars to index data.

IdxType

Int8Chunked

Int16Chunked

Int32Chunked

Int64Chunked

ListChunked

PlHashMap

PlHashSet

PlIndexMap

PlIndexSet

TimeChunked

TimeZone

UInt8Chunked

UInt16Chunked

UInt32Chunked

UInt64Chunked

Utf8Chunked