polars_core::datatypes::categorical

Type Alias CategoricalChunked

pub type CategoricalChunked<T> = Logical<T, <T as PolarsCategoricalType>::PolarsPhysical>;

Available on crate feature dtype-categorical only.

Aliased Type

pub struct CategoricalChunked<T> {
    pub phys: ChunkedArray<<T as PolarsCategoricalType>::PolarsPhysical>,
    pub dtype: DataType,
    /* private fields */
}

Fields
phys: ChunkedArray<<T as PolarsCategoricalType>::PolarsPhysical>dtype: DataType
Implementations

impl<T: PolarsCategoricalType> CategoricalChunked<T>

pub fn sort_with(&self, options: SortOptions) -> CategoricalChunked<T>

pub fn sort(&self, descending: bool) -> CategoricalChunked<T>

Returned a sorted ChunkedArray.

pub fn arg_sort(&self, options: SortOptions) -> IdxCa

Retrieve the indexes needed to sort this array.

impl<T: PolarsCategoricalType> CategoricalChunked<T>

pub fn is_enum(&self) -> bool

pub fn uses_lexical_ordering(&self) -> bool

Return whether or not the CategoricalChunked uses the lexical order of the string values when sorting.

pub fn full_null_with_dtype( name: PlSmallStr, length: usize, dtype: DataType, ) -> Self

pub fn from_cats_and_dtype( cat_ids: ChunkedArray<T::PolarsPhysical>, dtype: DataType, ) -> Self

Create a CategoricalChunked from a physical array and dtype.

Checks that all the category ids are valid, mapping invalid ones to nulls.

pub unsafe fn from_cats_and_dtype_unchecked( cat_ids: ChunkedArray<T::PolarsPhysical>, dtype: DataType, ) -> Self

Create a CategoricalChunked from a physical array and dtype.

Safety

It’s not checked that the indices are in-bounds or that the dtype is correct.

pub fn get_mapping(&self) -> &Arc<CategoricalMapping>

Get a reference to the mapping of categorical types to the string values.

pub fn iter_str(&self) -> impl PolarsIterator<Item = Option<&str>>

Create an Iterator that iterates over the &str values of the CategoricalChunked.

pub fn from_str_iter<'a, I: IntoIterator<Item = Option<&'a str>>>( name: PlSmallStr, dtype: DataType, strings: I, ) -> PolarsResult<Self>

Converts from strings to this CategoricalChunked.

If this dtype is an Enum any non-existing strings get mapped to null.

pub fn to_arrow(&self, compat_level: CompatLevel) -> DictionaryArray<T::Native>