polars_core::datatypes

Enum AnyValue

pub enum AnyValue<'a> {

    Null,
    Boolean(bool),
    String(&'a str),
    UInt8(u8),
    UInt16(u16),
    UInt32(u32),
    UInt64(u64),
    Int8(i8),
    Int16(i16),
    Int32(i32),
    Int64(i64),
    Int128(i128),
    Float32(f32),
    Float64(f64),
    Categorical(CatSize, &'a Arc<CategoricalMapping>),
    CategoricalOwned(CatSize, Arc<CategoricalMapping>),
    Enum(CatSize, &'a Arc<CategoricalMapping>),
    EnumOwned(CatSize, Arc<CategoricalMapping>),
    List(Series),
    Array(Series, usize),
    Object(&'a dyn PolarsObjectSafe),
    ObjectOwned(OwnedObject),
    StringOwned(PlSmallStr),
    Binary(&'a [u8]),
    BinaryOwned(Vec<u8>),
    Decimal(i128, usize),

}

Variants

Null

Boolean(bool)

A binary true or false.

String(&'a str)

A UTF8 encoded string type.

UInt8(u8)

An unsigned 8-bit integer number.

UInt16(u16)

An unsigned 16-bit integer number.

UInt32(u32)

An unsigned 32-bit integer number.

UInt64(u64)

An unsigned 64-bit integer number.

Int8(i8)

An 8-bit integer number.

Int16(i16)

A 16-bit integer number.

Int32(i32)

A 32-bit integer number.

Int64(i64)

A 64-bit integer number.

Int128(i128)

A 128-bit integer number.

Float32(f32)

A 32-bit floating point number.

Float64(f64)

A 64-bit floating point number.

Categorical(CatSize, &'a Arc<CategoricalMapping>)

Available on crate feature dtype-categorical only.

CategoricalOwned(CatSize, Arc<CategoricalMapping>)

Available on crate feature dtype-categorical only.

Enum(CatSize, &'a Arc<CategoricalMapping>)

Available on crate feature dtype-categorical only.

EnumOwned(CatSize, Arc<CategoricalMapping>)

Available on crate feature dtype-categorical only.

List(Series)

Nested type, contains arrays that are filled with one of the datatypes.

Array(Series, usize)

Available on crate feature dtype-array only.

Object(&'a dyn PolarsObjectSafe)

Available on crate feature object only.

Can be used to fmt and implements Any, so can be downcasted to the proper value type.

ObjectOwned(OwnedObject)

Available on crate feature object only.

StringOwned(PlSmallStr)

An UTF8 encoded string type.

Binary(&'a [u8])

BinaryOwned(Vec<u8>)

Decimal(i128, usize)

Available on crate feature dtype-decimal only.

A 128-bit fixed point decimal number with a scale.

Implementations

impl AnyValue<'static>

pub fn zero_sum(dtype: &DataType) -> Self

pub fn can_have_dtype(&self, dtype: &DataType) -> bool

Can the AnyValue exist as having dtype as its DataType.

impl<'a> AnyValue<'a>

pub fn dtype(&self) -> DataType

Get the matching DataType for this AnyValue`.

Note: For Categorical and Enum values, the exact mapping information is not preserved in the result for performance reasons.

pub fn try_extract<T: NumCast>(&self) -> PolarsResult<T>

pub fn is_boolean(&self) -> bool

pub fn is_primitive_numeric(&self) -> bool

pub fn is_float(&self) -> bool

pub fn is_integer(&self) -> bool

pub fn is_signed_integer(&self) -> bool

pub fn is_unsigned_integer(&self) -> bool

pub fn is_nan(&self) -> bool

pub fn is_null(&self) -> bool

pub fn is_nested_null(&self) -> bool

pub fn strict_cast(&self, dtype: &'a DataType) -> Option<AnyValue<'a>>

Cast AnyValue to the provided data type and return a new AnyValue with type dtype, if possible.

pub fn try_strict_cast(&self, dtype: &'a DataType) -> PolarsResult<AnyValue<'a>>

Cast AnyValue to the provided data type and return a new AnyValue with type dtype, if possible.

pub fn cast(&self, dtype: &'a DataType) -> AnyValue<'a>

pub fn idx(&self) -> IdxSize

pub fn str_value(&self) -> Cow<'a, str>

pub fn to_physical(self) -> Self

pub fn extract_bool(&self) -> Option<bool>

pub fn extract_str(&self) -> Option<&str>

pub fn extract_bytes(&self) -> Option<&[u8]>

impl AnyValue<'_>

pub fn hash_impl<H: Hasher>(&self, state: &mut H, cheap: bool)

impl<'a> AnyValue<'a>

pub fn add(&self, rhs: &AnyValue<'_>) -> AnyValue<'static>

pub fn as_borrowed(&self) -> AnyValue<'_>

pub fn into_static(self) -> AnyValue<'static>

Try to coerce to an AnyValue with static lifetime. This can be done if it does not borrow any values.

pub fn get_str(&self) -> Option<&str>

Get a reference to the &str contained within AnyValue.

impl AnyValue<'_>

pub fn eq_missing(&self, other: &Self, null_equal: bool) -> bool