polars_core::datatypes
Struct CategoricalChunked

pub struct CategoricalChunked { /* private fields */ }

Available on crate feature dtype-categorical only.

Implementations

impl CategoricalChunked

pub fn sort_with(&self, options: SortOptions) -> CategoricalChunked

pub fn sort(&self, descending: bool) -> CategoricalChunked

Returned a sorted ChunkedArray.

pub fn arg_sort(&self, options: SortOptions) -> IdxCa

Retrieve the indexes needed to sort this array.

impl CategoricalChunked

pub unsafe fn from_global_indices_unchecked( cats: UInt32Chunked, ordering: CategoricalOrdering, ) -> CategoricalChunked

Create a CategoricalChunked from a categorical indices. The indices will probe the global string cache.

Safety

This does not do any bound checks

pub fn from_string_to_enum( values: &StringChunked, categories: &Utf8ViewArray, ordering: CategoricalOrdering, ) -> PolarsResult<CategoricalChunked>

Create a CategoricalChunked from a fixed list of categories and a List of strings. This will error if a string is not in the fixed list of categories

impl CategoricalChunked

pub fn to_arrow(&self, compat_level: CompatLevel, as_i64: bool) -> ArrayRef

impl CategoricalChunked

pub fn append(&mut self, other: &Self) -> PolarsResult<()>

impl CategoricalChunked

pub fn full_null( name: PlSmallStr, is_enum: bool, length: usize, ordering: CategoricalOrdering, ) -> CategoricalChunked

impl CategoricalChunked

pub fn unique(&self) -> PolarsResult<Self>

Available on crate feature algorithm_group_by only.

pub fn n_unique(&self) -> PolarsResult<usize>

Available on crate feature algorithm_group_by only.

pub fn value_counts(&self) -> PolarsResult<DataFrame>

Available on crate feature algorithm_group_by only.

impl CategoricalChunked

pub fn is_empty(&self) -> bool

pub fn len(&self) -> usize

pub fn null_count(&self) -> usize

pub fn name(&self) -> &PlSmallStr

pub fn physical(&self) -> &UInt32Chunked

Get a reference to the physical array (the categories).

pub fn is_enum(&self) -> bool

pub fn to_local(&self) -> Self

Convert a categorical column to its local representation.

pub fn to_global(&self) -> PolarsResult<Self>

pub fn to_enum(&self, categories: &Utf8ViewArray, hash: u128) -> Self

pub fn uses_lexical_ordering(&self) -> bool

Return whether or not the CategoricalChunked uses the lexical order of the string values when sorting.

pub unsafe fn from_cats_and_rev_map_unchecked( idx: UInt32Chunked, rev_map: Arc<RevMapping>, is_enum: bool, ordering: CategoricalOrdering, ) -> Self

Create a CategoricalChunked from an array of idx and an existing RevMapping: rev_map.

Safety

Invariant in v < rev_map.len() for v in idx must hold.

pub fn _can_fast_unique(&self) -> bool

True if all categories are represented in this array. When this is the case, the unique values of the array are the categories.

pub fn _with_fast_unique(self, toggle: bool) -> Self

pub fn get_rev_map(&self) -> &Arc<RevMapping>

Get a reference to the mapping of categorical types to the string values.

pub fn iter_str(&self) -> CatIter<'_>

Create an [Iterator] that iterates over the &str values of the [CategoricalChunked].