Struct polars_core::datatypes::Logical

pub struct Logical<Logical: PolarsDataType, Physical: PolarsDataType>(pub ChunkedArray<Physical>, _, pub Option<DataType>);

Maps a logical type to a a chunked array implementation of the physical type. This saves a lot of compiler bloat and allows us to reuse functionality.

Tuple Fields
0: ChunkedArray<Physical>2: Option<DataType>
Implementations

impl<K: PolarsDataType, T: PolarsDataType> Logical<K, T>

pub fn new_logical<J: PolarsDataType>(ca: ChunkedArray<T>) -> Logical<J, T>

impl<K: PolarsDataType, T: PolarsDataType> Logical<K, T>where Self: LogicalType,

pub fn field(&self) -> Field

impl Logical<DateType, Int32Type>

pub fn as_date_iter( &self ) -> impl Iterator<Item = Option<NaiveDate>> + TrustedLen + '_

pub fn from_naive_date<I: IntoIterator<Item = NaiveDate>>( name: &str, v: I ) -> Self

Construct a new DateChunked from an iterator over NaiveDate.

pub fn to_string(&self, format: &str) -> Utf8Chunked

Convert from Date into Utf8 with the given format. See chrono strftime/strptime.

pub fn strftime(&self, format: &str) -> Utf8Chunked

Convert from Date into Utf8 with the given format. See chrono strftime/strptime.

Alias for to_string.

pub fn from_naive_date_options<I: IntoIterator<Item = Option<NaiveDate>>>( name: &str, v: I ) -> Self

Construct a new DateChunked from an iterator over optional NaiveDate.

impl Logical<DatetimeType, Int64Type>

pub fn as_datetime_iter( &self ) -> impl Iterator<Item = Option<NaiveDateTime>> + TrustedLen + '_

pub fn time_unit(&self) -> TimeUnit

pub fn time_zone(&self) -> &Option<TimeZone>

pub fn to_string(&self, format: &str) -> PolarsResult<Utf8Chunked>

Convert from Datetime into Utf8 with the given format. See chrono strftime/strptime.

pub fn strftime(&self, format: &str) -> PolarsResult<Utf8Chunked>

Convert from Datetime into Utf8 with the given format. See chrono strftime/strptime.

Alias for to_string.

pub fn from_naive_datetime<I: IntoIterator<Item = NaiveDateTime>>( name: &str, v: I, tu: TimeUnit ) -> Self

Construct a new DatetimeChunked from an iterator over NaiveDateTime.

pub fn from_naive_datetime_options<I: IntoIterator<Item = Option<NaiveDateTime>>>( name: &str, v: I, tu: TimeUnit ) -> Self

pub fn cast_time_unit(&self, tu: TimeUnit) -> Self

Change the underlying TimeUnit. And update the data accordingly.

pub fn set_time_unit(&mut self, tu: TimeUnit)

Change the underlying TimeUnit. This does not modify the data.

impl Logical<DurationType, Int64Type>

pub fn time_unit(&self) -> TimeUnit

pub fn cast_time_unit(&self, tu: TimeUnit) -> Self

Change the underlying TimeUnit. And update the data accordingly.

pub fn set_time_unit(&mut self, tu: TimeUnit)

Change the underlying TimeUnit. This does not modify the data.

pub fn from_duration<I: IntoIterator<Item = ChronoDuration>>( name: &str, v: I, tu: TimeUnit ) -> Self

Construct a new DurationChunked from an iterator over ChronoDuration.

pub fn from_duration_options<I: IntoIterator<Item = Option<ChronoDuration>>>( name: &str, v: I, tu: TimeUnit ) -> Self

Construct a new DurationChunked from an iterator over optional ChronoDuration.

impl Logical<TimeType, Int64Type>

pub fn to_string(&self, format: &str) -> Utf8Chunked

Convert from Time into Utf8 with the given format. See chrono strftime/strptime.

pub fn strftime(&self, format: &str) -> Utf8Chunked

Convert from Time into Utf8 with the given format. See chrono strftime/strptime.

Alias for to_string.

pub fn as_time_iter( &self ) -> impl Iterator<Item = Option<NaiveTime>> + TrustedLen + '_

pub fn from_naive_time<I: IntoIterator<Item = NaiveTime>>( name: &str, v: I ) -> Self

Construct a new TimeChunked from an iterator over NaiveTime.

pub fn from_naive_time_options<I: IntoIterator<Item = Option<NaiveTime>>>( name: &str, v: I ) -> Self

Construct a new TimeChunked from an iterator over optional NaiveTime.