Type Definition polars_core::datatypes::ArrayChunked

pub type ArrayChunked = ChunkedArray<FixedSizeListType>;

Available on crate feature dtype-array only.

Implementations

impl ArrayChunked

pub fn full_null_with_dtype( name: &str, length: usize, inner_dtype: &DataType, width: usize ) -> ArrayChunked

impl ArrayChunked

pub fn amortized_iter( &self ) -> AmortizedListIter<'_, impl Iterator<Item = Option<ArrayBox>> + '_>

Available on crate feature private only.

This is an iterator over a ListChunked that save allocations. A Series is: 1. Arc ChunkedArray is: 2. Vec< 3. ArrayRef>

The ArrayRef we indicated with 3. will be updated during iteration. The Series will be pinned in memory, saving an allocation for

Arc<..>

Vec<…>

Warning

Though memory safe in the sense that it will not read unowned memory, UB, or memory leaks this function still needs precautions. The returned should never be cloned or taken longer than a single iteration, as every call on next of the iterator will change the contents of that Series.

pub fn amortized_iter_with_name( &self, name: &str ) -> AmortizedListIter<'_, impl Iterator<Item = Option<ArrayBox>> + '_>

Available on crate feature private only.

impl ArrayChunked

pub fn inner_dtype(&self) -> DataType

Get the inner data type of the fixed size list.

pub fn width(&self) -> usize

pub fn to_physical(&mut self, inner_dtype: DataType)

pub fn get_inner(&self) -> Series

Get the inner values as Series

pub fn apply_to_inner( &self, func: &dyn Fn(Series) -> PolarsResult<Series> ) -> PolarsResult<ArrayChunked>

Ignore the list indices and apply func to the inner type as Series.