polars_core::datatypes

Type Alias DateChunked

pub type DateChunked = Logical<DateType, Int32Type>;

Aliased Type

struct DateChunked {
    pub phys: ChunkedArray<Int32Type>,
    pub dtype: DataType,
    /* private fields */
}

Fields
phys: ChunkedArray<Int32Type>dtype: DataType
Implementations

impl DateChunked

pub fn as_date_iter(&self) -> impl TrustedLen<Item = Option<NaiveDate>> + '_

pub fn from_naive_date<I: IntoIterator<Item = NaiveDate>>( name: PlSmallStr, v: I, ) -> Self

Construct a new DateChunked from an iterator over NaiveDate.

pub fn to_string(&self, format: &str) -> PolarsResult<StringChunked>

Convert from Date into String with the given format. See chrono strftime/strptime.

pub fn strftime(&self, format: &str) -> PolarsResult<StringChunked>

Convert from Date into String with the given format. See chrono strftime/strptime.

Alias for to_string.

pub fn from_naive_date_options<I: IntoIterator<Item = Option<NaiveDate>>>( name: PlSmallStr, v: I, ) -> Self

Construct a new DateChunked from an iterator over optional NaiveDate.