polars_core::datatypes

Type Alias DatetimeChunked

pub type DatetimeChunked = Logical<DatetimeType, Int64Type>;

Aliased Type

struct DatetimeChunked {
    pub phys: ChunkedArray<Int64Type>,
    pub dtype: DataType,
    /* private fields */
}

Fields
phys: ChunkedArray<Int64Type>dtype: DataType
Implementations

impl DatetimeChunked

pub fn as_datetime_iter( &self, ) -> impl TrustedLen<Item = Option<NaiveDateTime>> + '_

pub fn time_unit(&self) -> TimeUnit

pub fn time_zone(&self) -> &Option<TimeZone>

pub fn to_string(&self, format: &str) -> PolarsResult<StringChunked>

Convert from Datetime into String with the given format. See chrono strftime/strptime.

pub fn strftime(&self, format: &str) -> PolarsResult<StringChunked>

Convert from Datetime into String with the given format. See chrono strftime/strptime.

Alias for to_string.

pub fn from_naive_datetime<I: IntoIterator<Item = NaiveDateTime>>( name: PlSmallStr, v: I, tu: TimeUnit, ) -> Self

Construct a new DatetimeChunked from an iterator over NaiveDateTime.

pub fn from_naive_datetime_options<I: IntoIterator<Item = Option<NaiveDateTime>>>( name: PlSmallStr, v: I, tu: TimeUnit, ) -> Self

pub fn cast_time_unit(&self, tu: TimeUnit) -> Self

Change the underlying TimeUnit. And update the data accordingly.

pub fn set_time_unit(&mut self, time_unit: TimeUnit)

Change the underlying TimeUnit. This does not modify the data.