Type Definition polars_core::datatypes::ObjectChunked

pub type ObjectChunked<T> = ChunkedArray<ObjectType<T>>;

Implementations

impl<'a, T> ObjectChunked<T> where
T: PolarsObject,

pub fn downcast_iter(
&self
) -> impl Iterator<Item = &ObjectArray<T>> + DoubleEndedIterator

pub fn downcast_chunks(&self) -> Chunks<'_, ObjectArray<T>>

impl<T: PolarsObject> ObjectChunked<T>

pub fn into_no_null_iter(
&self
) -> impl Iterator<Item = &T> + '_ + Send + Sync + ExactSizeIterator + DoubleEndedIterator + TrustedLen

impl<T> ObjectChunked<T> where
T: PolarsObject,

pub fn new_from_vec(name: &str, v: Vec<T>) -> Self

impl<T> ObjectChunked<T> where
T: PolarsObject,

pub unsafe fn get_object_unchecked(
&self,
index: usize
) -> Option<&dyn PolarsObjectSafe>

Get a hold to an object that can be formatted or downcasted via the Any trait.

Safety

No bounds checks

pub fn get_object(&self, index: usize) -> Option<&dyn PolarsObjectSafe>

Get a hold to an object that can be formatted or downcasted via the Any trait.