polars_core::datatypes

Type Alias StringChunked

pub type StringChunked = ChunkedArray<StringType>;

Aliased Type

pub struct StringChunked { /* private fields */ }

Implementations

impl StringChunked

pub fn apply_mut<'a, F>(&'a self, f: F) -> Self
where F: FnMut(&'a str) -> &'a str,

impl StringChunked

pub unsafe fn apply_views<F: FnMut(View, &str) -> View + Copy>( &self, update_view: F, ) -> Self

Safety

Update the views. All invariants of the views apply.

impl StringChunked

pub fn to_decimal(&self, infer_length: usize) -> PolarsResult<Series>

Available on crate feature dtype-decimal only.

Convert an StringChunked to a Series of DataType::Decimal. Scale needed for the decimal type are inferred. Parsing is not strict.
Scale inference assumes that all tested strings are well-formed numbers, and may produce unexpected results for scale if this is not the case.

If the decimal precision and scale are already known, consider using the cast method.

impl StringChunked

pub fn as_binary(&self) -> BinaryChunked

impl StringChunked

pub fn par_iter_indexed( &self, ) -> impl IndexedParallelIterator<Item = Option<&str>>

pub fn par_iter(&self) -> impl ParallelIterator<Item = Option<&str>> + '_