Type Definition polars_core::datatypes::Utf8Chunked

pub type Utf8Chunked = ChunkedArray<Utf8Type>;

Implementations

impl Utf8Chunked

pub fn apply_mut<'a, F>(&'a self, f: F) -> Selfwhere F: FnMut(&'a str) -> &'a str,

pub fn apply_to_buffer<'a, F>(&'a self, f: F) -> Selfwhere F: FnMut(&'a str, &mut String),

Utility that reuses an string buffer to amortize allocations. Prefer this over an apply that returns an owned String.

impl Utf8Chunked

pub fn as_binary(&self) -> BinaryChunked

impl Utf8Chunked

pub fn par_iter_indexed( &self ) -> impl IndexedParallelIterator<Item = Option<&str>>

pub fn par_iter(&self) -> impl ParallelIterator<Item = Option<&str>> + '_