Function polars_core::error::set_warning_function

pub unsafe fn set_warning_function(function: fn(_: &str, _: PolarsWarning))

Set the function that will be called by the polars_warn! macro. You can use this to set logging in polars.

Safety

The caller must ensure there is no other thread accessing this function or calling polars_warn!.