polars_core::export
Module cast

Defines different casting operators such as cast or [primitive_to_binary].

Modules

temporal

Structs

CastOptionsImpl

options defining how Cast kernels behave

Traits

SerPrimitive

Functions

binary_large_to_binary

Conversion of binary

binary_to_binview

binary_to_dictionary

Cast [BinaryArray] to [DictionaryArray], also known as packing.

binary_to_large_binary

Conversion of binary

binary_to_list

Conversion of binary

binary_to_utf8

Conversion to utf8

binview_to_decimaldtype-decimal

boolean_to_binaryview

Casts the [BooleanArray] to a [BinaryArray], casting trues to "1" and falses to "0"

boolean_to_primitive

Casts the [BooleanArray] to a [PrimitiveArray].

boolean_to_utf8view

cast

Cast array to the provided data type and return a new [Array] with type to_type, if possible.

cast_default

cast_unchecked

date32_to_date64

Conversion of dates

date64_to_date32

Conversion of dates

decimal_to_decimal

Returns a [PrimitiveArray<i128>] with the casted values. Values are None on overflow

decimal_to_float

Returns a [PrimitiveArray<i128>] with the casted values. Values are None on overflow

decimal_to_integer

Returns a [PrimitiveArray<i128>] with the casted values. Values are None on overflow

f16_to_f32

Casts f16 into f32

fixed_size_binary_binary

Conversion of FixedSizeBinary to Binary.

fixed_size_binary_to_binview

float_to_decimal

Returns a [PrimitiveArray<i128>] with the casted values. Values are None on overflow

int32_to_time32ms

Conversion of Int32 to Time32(TimeUnit::Millisecond)

int32_to_time32s

Conversion of Int32 to Time32(TimeUnit::Second)

int64_to_time64ns

Conversion of Int64 to Time32(TimeUnit::Nanosecond)

int64_to_time64us

Conversion of Int64 to Time32(TimeUnit::Microsecond)

integer_to_decimal

Returns a [PrimitiveArray<i128>] with the casted values. Values are None on overflow

primitive_as_primitive

Cast [PrimitiveArray] as a [PrimitiveArray] Same as number as to_number_type in rust

primitive_map_is_valid⚠

Safety

primitive_to_boolean

Returns a [BooleanArray] where every element is different from zero. Validity is preserved.

primitive_to_dictionary

Cast [PrimitiveArray] to [DictionaryArray]. Also known as packing.

primitive_to_primitive

Cast [PrimitiveArray] to a [PrimitiveArray] of another physical type via numeric conversion.

primitive_to_same_primitive

Cast [PrimitiveArray] to a [PrimitiveArray] of the same physical type. This is O(1).

time32_to_time64

Conversion of time

time32ms_to_time32s

Conversion of times

time32s_to_time32ms

Conversion of times

time64_to_time32

Conversion of time

time64ns_to_time64us

Conversion of times

time64us_to_time64ns

Conversion of times

timestamp_to_date32

Conversion of timestamp

timestamp_to_date64

Conversion of timestamp

timestamp_to_timestamp

Conversion of timestamp

utf8_large_to_utf8

Conversion of utf8

utf8_to_binary

Conversion to binary

utf8_to_dictionary

Cast [Utf8Array] to [DictionaryArray], also known as packing.

utf8_to_large_utf8

Conversion of utf8

utf8_to_utf8view

utf8view_to_utf8