polars_core::frame::column
Struct ScalarColumn

pub struct ScalarColumn { /* private fields */ }

A Column that consists of a repeated Scalar

This is lazily materialized into a Series.

Implementations

impl ScalarColumn

pub fn new(name: PlSmallStr, scalar: Scalar, length: usize) -> Self

pub fn new_empty(name: PlSmallStr, dtype: DataType) -> Self

pub fn name(&self) -> &PlSmallStr

pub fn scalar(&self) -> &Scalar

pub fn dtype(&self) -> &DataType

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn to_series(&self) -> Series

Materialize the ScalarColumn into a Series.

pub fn lazy_as_materialized_series(&self) -> Option<&Series>

Get the ScalarColumn as Series if it was already materialized.

pub fn as_materialized_series(&self) -> &Series

Get the ScalarColumn as Series

This needs to materialize upon the first call. Afterwards, this is cached.

pub fn take_materialized_series(self) -> Series

Take the ScalarColumn and materialize as a Series if not already done.

pub fn as_single_value_series(&self) -> Series

Take the ScalarColumn as a series with a single value.

If the ScalarColumn has length=0 the resulting Series will also have length=0.

pub fn unit_scalar_from_series(series: Series) -> Self

Create a new ScalarColumn from a length=1 Series and expand it length.

This will panic if the value cannot be made static or if the series has length 0.

pub fn from_single_value_series(series: Series, length: usize) -> Self

Create a new ScalarColumn from a length=1 Series and expand it length.

This will panic if the value cannot be made static or if the series has length 0.

pub fn resize(&self, length: usize) -> ScalarColumn

Resize the ScalarColumn to new length.

This reuses the materialized Series, if length <= self.length.

pub fn cast_with_options( &self, dtype: &DataType, options: CastOptions, ) -> PolarsResult<Self>

pub fn strict_cast(&self, dtype: &DataType) -> PolarsResult<Self>

pub fn cast(&self, dtype: &DataType) -> PolarsResult<Self>

pub unsafe fn cast_unchecked(&self, dtype: &DataType) -> PolarsResult<Self>

Safety

This can lead to invalid memory access in downstream code.

pub fn rename(&mut self, name: PlSmallStr) -> &mut Self

pub fn has_nulls(&self) -> bool