Module polars_core::frame::groupby

Modules

aggregations

Structs

GroupBy

Returned by a groupby operation on a DataFrame. This struct supports several aggregations.

GroupsIdx

Indexes of the groups, the first index is stored separately. this make sorting fast.

GroupsProxyIter

GroupsProxyParIter

SlicedGroups

Enums

GroupByMethod

GroupsIndicator

GroupsProxy

PivotAgg

Traits

IntoGroupsProxy

Used to create the tuples for a groupby operation.

Functions

fmt_groupby_column

Type Definitions

BorrowIdxItem

GroupsSlice

Every group is indicated by an array where the

IdxItem