Trait polars_core::frame::groupby::IntoGroupsProxy

pub trait IntoGroupsProxy {
    fn group_tuples(
        &self,
        _multithreaded: bool,
        _sorted: bool
    ) -> PolarsResult<GroupsProxy> { ... }
}

Used to create the tuples for a groupby operation.

Provided Methods

fn group_tuples(
&self,
_multithreaded: bool,
_sorted: bool
) -> PolarsResult<GroupsProxy>

Create the tuples need for a groupby operation. * The first value in the tuple is the first index of the group. * The second value in the tuple is are the indexes of the groups including the first value.