polars_core::frame

Module row

Available on crate features rows or object only.

Structs

Row

Enums

AnyValueBuffer

AnyValueBufferTrusted

An AnyValueBuffer that should be used when we trust the builder

Functions

coerce_dtype

Coerces a slice of datatypes into a single supertype.

infer_schema

rows_to_schema_first_non_null

Infer schema from rows and set the first no null type as column data type.

rows_to_schema_supertypes

Infer the schema of rows by determining the supertype of the values.

rows_to_supertypes

Infer the schema data types of rows by determining the supertype of the values.