Module polars_core::functions

Functions

Functions that might be useful.

Functions

argsort_by

Find the indexes that would sort these series in order of appearance. That means that the first Series will be used to determine the ordering until duplicates are found. Once duplicates are found, the next Series will be used and so on.

cov_f

Compute the covariance between two columns.

cov_i

Compute the covariance between two columns.

pearson_corr_f

Compute the pearson correlation between two columns.

pearson_corr_i

Compute the pearson correlation between two columns.