Module polars_core::hashing

Structs

IdHasher

IdxHash

Contains an idx of a row in a DataFrame and the precomputed hash of that row. That hash still needs to be used to create another hash to be able to resize hashmaps without accidental quadratic behavior. So do not use an Identity function!

Constants

_HASHMAP_INIT_SIZE

Traits

VecHash

Functions

_boost_hash_combine

_df_rows_to_hashes_threaded_vertical

_hash_binary_array

populate_multiple_key_hashmap

Populate a multiple key hashmap with row indexes. Instead of the keys (which could be very large), the row indexes are stored. To check if a row is equal the original DataFrame is also passed as ref. When a hash collision occurs the indexes are ptrs to the rows and the rows are compared on equality.

Type Aliases

IdBuildHasher