Macro polars_core::assert_df_eq

macro_rules! assert_df_eq {
    ($a:expr, $b:expr $(,)?) => { ... };
}

Asserts that two expressions of type DataFrame are equal according to DataFrame::equals at runtime. If the expression are not equal, the program will panic with a message that displays both dataframes.