Struct polars_core::prelude::ArrowSchema

pub struct ArrowSchema {
    pub fields: Vec<Field, Global>,
    pub metadata: BTreeMap<String, String, Global>,
}

An ordered sequence of Fields with associated Metadata.

Schema is an abstration used to read from, and write to, Arrow IPC format, Apache Parquet, and Apache Avro. All these formats have a concept of a schema with fields and metadata.

Fields
fields: Vec<Field, Global>

The fields composing this schema.

metadata: BTreeMap<String, String, Global>

Optional metadata.

Implementations

impl Schema

pub fn with_metadata(self, metadata: BTreeMap<String, String, Global>) -> Schema

Attaches a Metadata to Schema

pub fn filter<F>(self, predicate: F) -> Schema where
F: Fn(usize, &Field) -> bool,

Returns a new Schema with a subset of all fields whose predicate evaluates to true.