Type Alias polars_core::prelude::ArrowSchema

pub type ArrowSchema = Schema<Field>;

An ordered sequence of Fields

ArrowSchema is an abstraction used to read from, and write to, Arrow IPC format, Apache Parquet, and Apache Avro. All these formats have a concept of a schema with fields and metadata.

Aliased Type

struct ArrowSchema { /* private fields */ }