Module polars_core::series::implementations::dates

This module exists to reduce compilation times. All the data types are backed by a physical type in memory e.g. Date32 -> i32, Date64 -> i64.

Series lead to code implementations of all traits. Whereas there are a lot of duplicates due to data types being backed by the same physical type. In this module we reduce compile times by opting for a little more run time cost. We cast to the physical type -> apply the operation and (depending on the result) cast back to the original type

Macros

cast_and_apply

Same as physical dispatch, but doesnt care about return type

impl_dyn_series

impl_dyn_series_numeric

opt_physical_dispatch

physical_dispatch

Dispatch the method call to the physical type and coerce back to logical type

try_physical_dispatch