Module polars_core::utils

Re-exports

pub use crate::chunked_array::ops::sort::arg_sort_no_nulls;

pub use arrow;

pub use rayon;

Modules

flatten

Structs

NoNull

Just a wrapper structure. Useful for certain impl specializations This is for instance use to implement impl<T> FromIterator<T::Native> for NoNull<ChunkedArray<T>> as Option<T::Native> was already implemented: impl<T> FromIterator<Option<T::Native>> for ChunkedArray<T>

TrustMyLength

Wrap

Traits

Arg

CustomIterTools

CustomIterToolsSized

FromTrustedLenIterator

IntoVec

Functions

_set_partition_size

_to_physical_and_bit_repr

Transform to physical type and coerce floating point and similar sized integer to a bit representation to reduce compiler bloat

accumulate_dataframes_horizontal

accumulate_dataframes_vertical

This takes ownership of the DataFrame so that drop is called earlier.

accumulate_dataframes_vertical_unchecked

This takes ownership of the DataFrame so that drop is called earlier. Does not check if schema is correct

align_chunks_binary

align_chunks_ternary

coalesce_nulls

ensure that nulls are propagated to both arrays

coalesce_nulls_series

combine_validities_and

combine_validities_or

concat_df

Concat the DataFrames to a single DataFrame.

concat_df_unchecked

Concat the DataFrames to a single DataFrame.

ensure_sorted_arg

get_casting_failures

get_supertype

Given two datatypes, determine the supertype that both types can safely be cast to

get_time_units

slice_slice

split_ca

split_df_as_ref

try_get_supertype

Given two datatypes, determine the supertype that both types can safely be cast to

with_unstable_series

A utility that allocates an UnstableSeries. The applied function can then use that series container to save heap allocations and swap arrow arrays.