Function polars_core::utils::parallel_op_series

pub fn parallel_op_series<F>(
    f: F,
    s: Series,
    n_threads: Option<usize>
) -> PolarsResult<Series>where
    F: Fn(Series) -> PolarsResult<Series> + Send + Sync,

Simple wrapper to parallelize functions that can be divided over threads aggregated and finally aggregated in the main thread. This can be done for sum, min, max, etc.