Function polars_core::utils::with_unstable_series

pub fn with_unstable_series<F, T>(dtype: &DataType, f: F) -> T where
    F: Fn(&mut UnstableSeries<'_>) -> T, 

A utility that allocates an UnstableSeries. The applied function can then use that series container to save heap allocations and swap arrow arrays.