Crate polars_core

Re-exports

pub use hashing::IdBuildHasher;

Modules

chunked_array

The typed heart of every Series column.

config

datatypes

Data types supported by Polars.

error

export

fmt

frame

DataFrame module.

functions

Functions

hashing

prelude

Everything you need to get started with Polars.

randomrandom

schema

series

Type agnostic columnar data structure.

testing

Testing utilities.

utils

Macros

apply_method_all_arrow_series

apply_method_physical_integer

apply_method_physical_numeric

assert_df_eq

Asserts that two expressions of type DataFrame are equal according to DataFrame::equals at runtime. If the expression are not equal, the program will panic with a message that displays both dataframes.

df

downcast_as_macro_arg_physical

Apply a macro on the Downcasted ChunkedArray’s of DataTypes that are logical numerics. So no logical.

downcast_as_macro_arg_physical_mut

Apply a macro on the Downcasted ChunkedArray’s of DataTypes that are logical numerics. So no logical.

match_arrow_data_type_apply_macro_ca

Apply a macro on the Downcasted ChunkedArray’s

match_dtype_to_logical_apply_macro

Apply a macro on the Series

match_dtype_to_physical_apply_macro

Apply a macro on the Series

with_match_physical_float_polars_type

with_match_physical_float_type

with_match_physical_integer_polars_type

with_match_physical_integer_type

with_match_physical_numeric_polars_type

with_match_physical_numeric_type

Structs

StringCacheHolderdtype-categorical

Enable the global string cache as long as the object is alive (RAII).

Statics

POOLNon-target_family="wasm"

PROCESS_ID

SINGLE_LOCK

Functions

disable_string_cachedtype-categorical

Disable and clear the global string cache.

enable_string_cachedtype-categorical

Enable the global string cache.

using_string_cachedtype-categorical

Check whether the global string cache is enabled.