Crate polars_core

Modules

chunked_array

The typed heart of every Series column.

cloud

config

datatypes

Data types supported by Polars.

docdocs

Other documentation

error

export

fmt

frame

DataFrame module.

functions

Functions

prelude

Everything you need to get started with Polars.

schema

series

Type agnostic columnar data structure.

testing

Testing utilities.

utils

Macros

apply_method_all_arrow_series

apply_method_physical_integer

apply_method_physical_numeric

df

downcast_as_macro_arg_physical

Apply a macro on the Downcasted ChunkedArray’s of DataTypes that are logical numerics. So no logical.

downcast_as_macro_arg_physical_mut

Apply a macro on the Downcasted ChunkedArray’s of DataTypes that are logical numerics. So no logical.

match_arrow_data_type_apply_macro_ca

Apply a macro on the Downcasted ChunkedArray’s

match_dtype_to_logical_apply_macro

Apply a macro on the Series

match_dtype_to_physical_apply_macro

Apply a macro on the Series

with_match_physical_integer_polars_type

with_match_physical_numeric_polars_type

with_match_physical_numeric_type

Structs

IUseStringCachedtype-categorical

RAII for the string cache

Statics

POOLNon-target_family="wasm"

PROCESS_ID

SINGLE_LOCK

Functions

reset_string_cachedtype-categorical

Reset the global string cache used for the Categorical Types.

toggle_string_cachedtype-categorical

Use a global string cache for the Categorical Types.

using_string_cachedtype-categorical

Check if string cache is set.

with_string_cachedtype-categorical