polars_core::chunked_array
Module iterator

Modules

par

Structs

BoolIterNoNull

The no null iterator for a [BooleanArray]

FixedSizeListIterNoNulldtype-array

SomeIterator

Wrapper struct to convert an iterator of type T into one of type Option<T>. It is useful to make the IntoIterator trait, in which every iterator shall return an Option<T>.

Traits

PolarsIterator

A PolarsIterator is an iterator over a ChunkedArray which contains polars types. A PolarsIterator must implement ExactSizeIterator and DoubleEndedIterator.