Module polars_core::chunked_array::object

Available on crate feature object only.

Modules

builder

registry

This is a heap allocated utility that can be used to register an object type. That object type will know its own generic type parameter T and callers can simply send &Any values and don’t have to know the generic type themselves.

Structs

ObjectArray

Traits

PolarsObject

Values need to implement this so that they can be stored into a Series and DataFrame

PolarsObjectSafe

Trimmed down object safe polars object

Type Aliases

ObjectValueIter