Trait polars_core::chunked_array::object::registry::AnonymousObjectBuilder

pub trait AnonymousObjectBuilder {
    // Required methods
    fn append_null(&mut self);
    fn append_value(&mut self, value: &dyn Any);
    fn to_series(&mut self) -> Series;

    // Provided method
    fn append_option(&mut self, value: Option<&dyn Any>) { ... }
}

Available on crate feature object only.

This trait can be registered, after which that global registration can be used to materialize object types

Required Methods

fn append_null(&mut self)

Append a null value.

fn append_value(&mut self, value: &dyn Any)

Append a T of ObjectChunked<T> made generic via the Any trait.

fn to_series(&mut self) -> Series

Take the current state and materialize as a Series the builder should not be used after that.

Provided Methods

fn append_option(&mut self, value: Option<&dyn Any>)