Module polars_core::chunked_array::ops

Traits for miscellaneous operations on ChunkedArray

Enums

FillNoneStrategy

TakeIdx

Traits

ChunkAgg
Aggregation operations

ChunkApply
Fastest way to do elementwise operations on a ChunkedArray when the operation is cheaper than branching due to null checking

ChunkApplyKernel
Apply kernels on the arrow array chunks in a ChunkedArray.

ChunkBytes

ChunkCast
Cast ChunkedArray<T> to ChunkedArray<N>

ChunkCompare
Compare Series and ChunkedArray's and get a boolean mask that can be used to filter rows.

ChunkCumAgg

ChunkExpandAtIndex
Create a new ChunkedArray filled with values at that index.

ChunkExplode

ChunkFillNone
Replace None values with various strategies

ChunkFillNoneValue
Replace None values with a value

ChunkFilter
Filter values by a boolean mask.

ChunkFull
Fill a ChunkedArray with one value.

ChunkFullNull

ChunkReverse
Reverse a ChunkedArray

ChunkSet
Create a ChunkedArray with new values by index or by boolean mask. Note that these operations clone data. This is however the only way we can modify at mask or index level as the underlying Arrow arrays are immutable.

ChunkShift

ChunkShiftFill
Shift the values of a ChunkedArray by a number of periods.

ChunkSort
Sort operations on ChunkedArray.

ChunkTake
Fast access by index.

ChunkTakeEvery

ChunkUnique
Get unique values in a ChunkedArray

ChunkVar
Variance and standard deviation aggregation.

ChunkWindow

ChunkWindowCustom

ChunkZip
Combine 2 ChunkedArrays based on some predicate.

TakeRandom
Random access

TakeRandomUtf8

ToDummies

Type Definitions

Dummy

TakeIdxIter

TakeIdxIterNull