Module polars_core::chunked_array::ops

Traits for miscellaneous operations on ChunkedArray

Enums

FillNoneStrategy

TakeIdx

One of the three arguments allowed in unchecked_take

Traits

ArgAgg

Argmin/ Argmax

ChunkAgg

Aggregation operations

ChunkAnyValue

ChunkApply

Fastest way to do elementwise operations on a ChunkedArray when the operation is cheaper than branching due to null checking

ChunkApplyKernel

Apply kernels on the arrow array chunks in a ChunkedArray.

ChunkBytes

ChunkCast

Cast ChunkedArray<T> to ChunkedArray<N>

ChunkCompare

Compare Series and ChunkedArray’s and get a boolean mask that can be used to filter rows.

ChunkCumAgg

ChunkExpandAtIndex

Create a new ChunkedArray filled with values at that index.

ChunkExplode

Explode/ flatten a

ChunkFillNone

Replace None values with various strategies

ChunkFillNoneValue

Replace None values with a value

ChunkFilter

Filter values by a boolean mask.

ChunkFull

Fill a ChunkedArray with one value.

ChunkFullNull

ChunkPeaks

Find local minima/ maxima

ChunkReverse

Reverse a ChunkedArray

ChunkSet

Create a ChunkedArray with new values by index or by boolean mask. Note that these operations clone data. This is however the only way we can modify at mask or index level as the underlying Arrow arrays are immutable.

ChunkShift

ChunkShiftFill

Shift the values of a ChunkedArray by a number of periods.

ChunkSort

Sort operations on ChunkedArray.

ChunkTake

Fast access by index.

ChunkTakeEvery

Traverse and collect every nth element

ChunkUnique

Get unique values in a ChunkedArray

ChunkVar

Variance and standard deviation aggregation.

ChunkWindow

Rolling window functions

ChunkWindowCustom

Custom rolling window functions

ChunkZip

Combine 2 ChunkedArrays based on some predicate.

Interpolate

IsFirst

IsIn

Check if element is member of list array

IsLast

Reinterpret

RepeatBy

Repeat the values n times.

TakeIterator

TakeIteratorNulls

TakeRandom

Random access

TakeRandomUtf8

ToDummies

Type Definitions

Dummy

Dummy type, we need to instantiate all generic types, so we fill one with a dummy.