Trait polars_core::chunked_array::ops::ChunkAgg

pub trait ChunkAgg<T> {
    fn sum(&self) -> Option<T>;

    fn min(&self) -> Option<T>;

    fn max(&self) -> Option<T>;

    fn mean(&self) -> Option<T>;

    fn median(&self) -> Option<T>;

    fn quantile(&self, quantile: f64) -> Result<Option<T>>;
}

Aggregation operations

Required methods

fn sum(&self) -> Option<T>

Aggregate the sum of the ChunkedArray. Returns None if the array is empty or only contains null values.

fn min(&self) -> Option<T>

fn max(&self) -> Option<T>

Returns the maximum value in the array, according to the natural order. Returns None if the array is empty or only contains null values.

fn mean(&self) -> Option<T>

Returns the mean value in the array. Returns None if the array is empty or only contains null values.

fn median(&self) -> Option<T>

Returns the mean value in the array. Returns None if the array is empty or only contains null values.

fn quantile(&self, quantile: f64) -> Result<Option<T>>

Aggregate a given quantile of the ChunkedArray. Returns None if the array is empty or only contains null values.

Loading content...