Trait polars_core::chunked_array::ops::ChunkApply

pub trait ChunkApply<'a, T> {
    type FuncRet;

    // Required methods
    fn apply_values<F>(&'a self, f: F) -> Self
       where F: Fn(T) -> Self::FuncRet + Copy;
    fn apply<F>(&'a self, f: F) -> Self
       where F: Fn(Option<T>) -> Option<Self::FuncRet> + Copy;
    fn apply_to_slice<F, S>(&'a self, f: F, slice: &mut [S])
       where F: Fn(Option<T>, &S) -> S;
}

Fastest way to do elementwise operations on a ChunkedArray<T> when the operation is cheaper than branching due to null checking.

Required Associated Types

type FuncRet

Required Methods

fn apply_values<F>(&'a self, f: F) -> Self
where F: Fn(T) -> Self::FuncRet + Copy,

Apply a closure elementwise. This is fastest when the null check branching is more expensive than the closure application. Often it is.

Null values remain null.

Example

use polars_core::prelude::*;
fn double(ca: &UInt32Chunked) -> UInt32Chunked {
    ca.apply_values(|v| v * 2)
}

fn apply<F>(&'a self, f: F) -> Self
where F: Fn(Option<T>) -> Option<Self::FuncRet> + Copy,

Apply a closure elementwise including null values.

fn apply_to_slice<F, S>(&'a self, f: F, slice: &mut [S])
where F: Fn(Option<T>, &S) -> S,

Apply a closure elementwise and write results to a mutable slice.

Object Safety

This trait is not object safe.