Trait polars_core::chunked_array::ops::ChunkApplyKernel

pub trait ChunkApplyKernel<A: Array> {
    fn apply_kernel(&self, f: &dyn Fn(&A) -> ArrayRef) -> Self;
    fn apply_kernel_cast<S>(&self, f: &dyn Fn(&A) -> ArrayRef) -> ChunkedArray<S>
    where
        S: PolarsDataType;
}

Apply kernels on the arrow array chunks in a ChunkedArray.

Required Methods

fn apply_kernel(&self, f: &dyn Fn(&A) -> ArrayRef) -> Self

Apply kernel and return result as a new ChunkedArray.

fn apply_kernel_cast<S>(&self, f: &dyn Fn(&A) -> ArrayRef) -> ChunkedArray<S>where
S: PolarsDataType,

Apply a kernel that outputs an array of different type.