polars_core::chunked_array::ops

Trait ChunkApplyKernel

pub trait ChunkApplyKernel<A: Array> {
    // Required methods
    fn apply_kernel(&self, f: &dyn Fn(&A) -> ArrayRef) -> Self;
    fn apply_kernel_cast<S>(
        &self,
        f: &dyn Fn(&A) -> ArrayRef,
    ) -> ChunkedArray<S>
       where S: PolarsDataType;
}

Apply kernels on the arrow array chunks in a ChunkedArray.

Required Methods

fn apply_kernel(&self, f: &dyn Fn(&A) -> ArrayRef) -> Self

Apply kernel and return result as a new ChunkedArray.

fn apply_kernel_cast<S>(&self, f: &dyn Fn(&A) -> ArrayRef) -> ChunkedArray<S>
where S: PolarsDataType,

Apply a kernel that outputs an array of different type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.