Trait polars_core::chunked_array::ops::ChunkQuantile

pub trait ChunkQuantile<T> {
    fn median(&self) -> Option<T> { ... }

    fn quantile(
        &self, 
        _quantile: f64, 
        _interpol: QuantileInterpolOptions
    ) -> Result<Option<T>> { ... }
}

Quantile and median aggregation

Provided methods

fn median(&self) -> Option<T>

Returns the mean value in the array. Returns None if the array is empty or only contains null values.

fn quantile(
&self,
_quantile: f64,
_interpol: QuantileInterpolOptions
) -> Result<Option<T>>

Aggregate a given quantile of the ChunkedArray. Returns None if the array is empty or only contains null values.