polars_core::chunked_array::ops
Trait ChunkShiftFill

pub trait ChunkShiftFill<T: PolarsDataType, V> {
    // Required method
    fn shift_and_fill(&self, periods: i64, fill_value: V) -> ChunkedArray<T>;
}

Shift the values of a ChunkedArray by a number of periods.

Required Methods

fn shift_and_fill(&self, periods: i64, fill_value: V) -> ChunkedArray<T>

Shift the values by a given period and fill the parts that will be empty due to this operation with fill_value.