Struct polars_core::chunked_array::ChunkedArray

pub struct ChunkedArray<T> { /* private fields */ }

ChunkedArray

Every Series contains a ChunkedArray<T>. Unlike Series, ChunkedArray’s are typed. This allows us to apply closures to the data and collect the results to a ChunkedArray of the same type T. Below we use an apply to use the cosine function to the values of a ChunkedArray.

fn apply_cosine(ca: &Float32Chunked) -> Float32Chunked {
    ca.apply(|v| v.cos())
}

If we would like to cast the result we could use a Rust Iterator instead of an apply method. Note that Iterators are slightly slower as the null values aren’t ignored implicitly.

fn apply_cosine_and_cast(ca: &Float32Chunked) -> Float64Chunked {
    ca.into_iter()
        .map(|opt_v| {
        opt_v.map(|v| v.cos() as f64)
    }).collect()
}

Another option is to first cast and then use an apply.

fn apply_cosine_and_cast(ca: &Float32Chunked) -> Float64Chunked {
    ca.apply_cast_numeric(|v| v.cos() as f64)
}

Conversion between Series and ChunkedArray’s

Conversion from a Series to a ChunkedArray is effortless.

fn to_chunked_array(series: &Series) -> Result<&Int32Chunked>{
    series.i32()
}

fn to_series(ca: Int32Chunked) -> Series {
    ca.into_series()
}

Iterators

ChunkedArrays fully support Rust native Iterator and DoubleEndedIterator traits, thereby giving access to all the excellent methods available for Iterators.

fn iter_forward(ca: &Float32Chunked) {
    ca.into_iter()
        .for_each(|opt_v| println!("{:?}", opt_v))
}

fn iter_backward(ca: &Float32Chunked) {
    ca.into_iter()
        .rev()
        .for_each(|opt_v| println!("{:?}", opt_v))
}

Memory layout

ChunkedArray’s use Apache Arrow as backend for the memory layout. Arrows memory is immutable which makes it possible to make multiple zero copy (sub)-views from a single array.

To be able to append data, Polars uses chunks to append new memory locations, hence the ChunkedArray<T> data structure. Appends are cheap, because it will not lead to a full reallocation of the whole array (as could be the case with a Rust Vec).

However, multiple chunks in a ChunkArray will slow down many operations that need random access because we have an extra indirection and indexes need to be mapped to the proper chunk. Arithmetic may also be slowed down by this. When multiplying two ChunkArray's with different chunk sizes they cannot utilize SIMD for instance.

If you want to have predictable performance (no unexpected re-allocation of memory), it is advised to call the rechunk after multiple append operations.

See also ChunkedArray::extend for appends within a chunk.

Implementations

impl<T> ChunkedArray<T> where
T: PolarsNumericType,

pub fn append(&mut self, other: &Self)

Append in place. This is done by adding the chunks of other to this ChunkedArray.

See also extend for appends to the underlying memory

impl<T: PolarsNumericType> ChunkedArray<T>

pub fn cast_and_apply_in_place<F, S>(&self, f: F) -> ChunkedArray<S> where
F: Fn(S::Native) -> S::Native + Copy,
S: PolarsNumericType,

Cast a numeric array to another numeric data type and apply a function in place. This saves an allocation.

impl<T> ChunkedArray<T> where
T: PolarsNumericType,

pub fn extend(&mut self, other: &Self)

Extend the memory backed by this array with the values from other.

Different from ChunkedArray::append which adds chunks to this ChunkedArray extent appends the data from other to the underlying PrimitiveArray and thus may cause a reallocation.

However if this does not cause a reallocation, the resulting data structure will not have any extra chunks and thus will yield faster queries.

Prefer extend over append when you want to do a query after a single append. For instance during online operations where you add n rows and rerun a query.

Prefer append over extend when you want to append many times before doing a query. For instance when you read in multiple files and when to store them in a single DataFrame. In the latter case finish the sequence of append operations with a rechunk.

impl ChunkedArray<ListType>

pub fn full_null_with_dtype(
name: &str,
length: usize,
inner_dtype: &DataType
) -> ListChunked

impl ChunkedArray<BooleanType>

pub fn arg_true(&self) -> IdxCa

impl ChunkedArray<BooleanType>

pub fn all(&self) -> bool

Check if all values are true

pub fn any(&self) -> bool

Check if any value is true

impl<T> ChunkedArray<T> where
T: PolarsFloatType,
T::Native: Float,

pub fn is_nan(&self) -> BooleanChunked

pub fn is_not_nan(&self) -> BooleanChunked

pub fn is_finite(&self) -> BooleanChunked

pub fn is_infinite(&self) -> BooleanChunked

pub fn none_to_nan(&self) -> Self

Convert missing values to NaN values.

impl ChunkedArray<ListType>

pub fn par_iter(&self) -> impl ParallelIterator<Item = Option<Series>> + '_

impl ChunkedArray<Utf8Type>

pub fn par_iter_indexed(
&self
) -> impl IndexedParallelIterator<Item = Option<&str>>

pub fn par_iter(&self) -> impl ParallelIterator<Item = Option<&str>> + '_

impl ChunkedArray<ListType>

pub fn amortized_iter(
&self
) -> AmortizedListIter<'_, impl Iterator<Item = Option<ArrayBox>> + '_>

This is an iterator over a ListChunked that save allocations. A Series is: 1. Arc ChunkedArray is: 2. Vec< 3. ArrayRef>

The ArrayRef we indicated with 3. will be updated during iteration. The Series will be pinned in memory, saving an allocation for

Arc<..>

Vec<…>

Warning

Though memory safe in the sense that it will not read unowned memory, UB, or memory leaks this function still needs precautions. The returned should never be cloned or taken longer than a single iteration, as every call on next of the iterator will change the contents of that Series.

pub fn apply_amortized<'a, F>(&'a self, f: F) -> Self where
F: FnMut(UnstableSeries<'a>) -> Series,

Apply a closure F elementwise.

pub fn try_apply_amortized<'a, F>(&'a self, f: F) -> Result<Self> where
F: FnMut(UnstableSeries<'a>) -> Result<Series>,

impl ChunkedArray<ListType>

pub fn set_fast_explode(&mut self)

pub fn to_logical(&mut self, inner_dtype: DataType)

impl ChunkedArray<Int32Type>

pub fn into_date(self) -> DateChunked

impl ChunkedArray<Int64Type>

pub fn into_datetime(
self,
timeunit: TimeUnit,
tz: Option<TimeZone>
) -> DatetimeChunked

impl ChunkedArray<Int64Type>

pub fn into_duration(self, timeunit: TimeUnit) -> DurationChunked

impl ChunkedArray<Int64Type>

pub fn into_time(self) -> TimeChunked

impl<T> ChunkedArray<T> where
T: PolarsNumericType,
Standard: Distribution<T::Native>,

pub fn init_rand(size: usize, null_density: f32, seed: u64) -> Self

impl<T> ChunkedArray<T> where
ChunkedArray<T>: ChunkTake,

pub fn sample_n(
&self,
n: usize,
with_replacement: bool,
shuffle: bool,
seed: Option<u64>
) -> Result<Self>

Sample n datapoints from this ChunkedArray.

pub fn sample_frac(
&self,
frac: f64,
with_replacement: bool,
shuffle: bool,
seed: Option<u64>
) -> Result<Self>

Sample a fraction between 0.0-1.0 of this ChunkedArray.

impl<T> ChunkedArray<T> where
T: PolarsNumericType,
T::Native: Float,

pub fn rand_normal(
name: &str,
length: usize,
mean: f64,
std_dev: f64
) -> Result<Self>

Create ChunkedArray with samples from a Normal distribution.

pub fn rand_standard_normal(name: &str, length: usize) -> Self

Create ChunkedArray with samples from a Standard Normal distribution.

pub fn rand_uniform(name: &str, length: usize, low: f64, high: f64) -> Self

Create ChunkedArray with samples from a Uniform distribution.

impl ChunkedArray<BooleanType>

pub fn rand_bernoulli(name: &str, length: usize, p: f64) -> Result<Self>

Create ChunkedArray with samples from a Bernoulli distribution.

impl<T> ChunkedArray<T>

pub fn set_sorted(&mut self, reverse: bool)

Set the ‘sorted’ bit meta info.

pub fn set_sorted2(&mut self, sorted: IsSorted)

pub fn first_non_null(&self) -> Option<usize>

Get the index of the first non null value in this ChunkedArray.

pub fn iter_validities(&self) -> impl Iterator<Item = Option<&Bitmap>> + '_

Get the buffer of bits representing null values

pub fn has_validity(&self) -> bool

Return if any the chunks in this [ChunkedArray] have a validity bitmap. no bitmap means no null values.

pub fn shrink_to_fit(&mut self)

Shrink the capacity of this array to fit it’s length.

pub fn unpack_series_matching_type(
&self,
series: &Series
) -> Result<&ChunkedArray<T>>

Series to ChunkedArray

pub fn chunk_id(&self) -> ChunkIdIter<'_>

Unique id representing the number of chunks

pub fn chunks(&self) -> &Vec<ArrayRef>

A reference to the chunks

pub fn is_optimal_aligned(&self) -> bool

Returns true if contains a single chunk and has no null values

pub fn null_count(&self) -> usize

Count the null values.

pub fn append_array(&mut self, other: ArrayRef) -> Result<()>

Append arrow array in place.

let mut array = Int32Chunked::new("array", &[1, 2]);
let array_2 = Int32Chunked::new("2nd", &[3]);

array.append(&array_2);
assert_eq!(Vec::from(&array), [Some(1), Some(2), Some(3)])

pub fn is_null(&self) -> BooleanChunked

Get a mask of the null values.

pub fn is_not_null(&self) -> BooleanChunked

Get a mask of the valid values.

pub fn dtype(&self) -> &DataType

Get data type of ChunkedArray.

pub fn name(&self) -> &str

Name of the ChunkedArray.

pub fn ref_field(&self) -> &Field

Get a reference to the field.

pub fn rename(&mut self, name: &str)

Rename this ChunkedArray.

impl<T> ChunkedArray<T> where
T: PolarsDataType,

pub fn from_chunks(name: &str, chunks: Vec<ArrayRef>) -> Self

Create a new ChunkedArray from existing chunks.

impl<T> ChunkedArray<T> where
T: PolarsNumericType,

pub fn from_vec(name: &str, v: Vec<T::Native>) -> Self

Create a new ChunkedArray by taking ownership of the Vec. This operation is zero copy.

pub fn new_from_owned_with_null_bitmap(
name: &str,
values: Vec<T::Native>,
buffer: Option<Bitmap>
) -> Self

Nullify values in slice with an existing null bitmap

impl<T> ChunkedArray<T> where
T: PolarsNumericType,

pub fn cont_slice(&self) -> Result<&[T::Native]>

Contiguous slice

pub fn data_views(
&self
) -> impl Iterator<Item = &[T::Native]> + DoubleEndedIterator

Get slices of the underlying arrow data. NOTE: null values should be taken into account by the user of these slices as they are handled separately

pub fn into_no_null_iter(
&self
) -> impl Iterator<Item = T::Native> + '_ + Send + Sync + ExactSizeIterator + DoubleEndedIterator + TrustedLen

impl ChunkedArray<ListType>

pub fn inner_dtype(&self) -> DataType

Get the inner data type of the list.

impl<T: PolarsNumericType> ChunkedArray<T>

pub fn new_vec(name: &str, v: Vec<T::Native>) -> Self

Specialization that prevents an allocation prefer this over ChunkedArray::new when you have a Vec<T::Native> and no null values.

impl<T> ChunkedArray<T> where
T: PolarsNumericType,
ChunkedArray<T>: IntoSeries,

We cannot override the left hand side behaviour. So we create a trait LhsNumOps. This allows for 1.add(&Series)

pub fn lhs_sub<N: Num + NumCast>(&self, lhs: N) -> Self

Apply lhs - self

pub fn lhs_div<N: Num + NumCast>(&self, lhs: N) -> Self

Apply lhs / self

pub fn lhs_rem<N: Num + NumCast>(&self, lhs: N) -> Self

Apply lhs % self