polars_core::chunked_array

Struct ChunkedArray

pub struct ChunkedArray<T: PolarsDataType> { /* private fields */ }

ChunkedArray

Every Series contains a ChunkedArray<T>. Unlike Series, ChunkedArrays are typed. This allows us to apply closures to the data and collect the results to a ChunkedArray of the same type T. Below we use an apply to use the cosine function to the values of a ChunkedArray.

fn apply_cosine_and_cast(ca: &Float32Chunked) -> Float32Chunked {
    ca.apply_values(|v| v.cos())
}

Conversion between Series and ChunkedArrays

Conversion from a Series to a ChunkedArray is effortless.

fn to_chunked_array(series: &Series) -> PolarsResult<&Int32Chunked>{
    series.i32()
}

fn to_series(ca: Int32Chunked) -> Series {
    ca.into_series()
}

Iterators

ChunkedArrays fully support Rust native Iterator and DoubleEndedIterator traits, thereby giving access to all the excellent methods available for Iterators.

fn iter_forward(ca: &Float32Chunked) {
    ca.iter()
        .for_each(|opt_v| println!("{:?}", opt_v))
}

fn iter_backward(ca: &Float32Chunked) {
    ca.iter()
        .rev()
        .for_each(|opt_v| println!("{:?}", opt_v))
}

Memory layout

ChunkedArrays use Apache Arrow as backend for the memory layout. Arrows memory is immutable which makes it possible to make multiple zero copy (sub)-views from a single array.

To be able to append data, Polars uses chunks to append new memory locations, hence the ChunkedArray<T> data structure. Appends are cheap, because it will not lead to a full reallocation of the whole array (as could be the case with a Rust Vec).

However, multiple chunks in a ChunkedArray will slow down many operations that need random access because we have an extra indirection and indexes need to be mapped to the proper chunk. Arithmetic may also be slowed down by this. When multiplying two ChunkedArrays with different chunk sizes they cannot utilize SIMD for instance.

If you want to have predictable performance (no unexpected re-allocation of memory), it is advised to call the ChunkedArray::rechunk after multiple append operations.

See also ChunkedArray::extend for appends within a chunk.

Invariants

A ChunkedArray should always have at least a single ArrayRef.

The PolarsDataType T should always map to the correct ArrowDataType in the ArrayRef chunks.

Nested datatypes such as List and [Array] store the physical types instead of the logical type given by the datatype.

Implementations

impl ChunkedArray<BooleanType>

Booleans are cast to 1 or 0.

pub fn sum(&self) -> Option<IdxSize>

pub fn min(&self) -> Option<bool>

pub fn max(&self) -> Option<bool>

pub fn mean(&self) -> Option<f64>

impl ChunkedArray<BinaryType>

pub fn max_binary(&self) -> Option<&[u8]>

pub fn min_binary(&self) -> Option<&[u8]>

impl<T> ChunkedArray<T>
where T: PolarsDataType<IsNested = FalseT, IsObject = FalseT>, for<'a> T::Physical<'a>: TotalOrd,

pub fn append(&mut self, other: &Self) -> PolarsResult<()>

Append in place. This is done by adding the chunks of other to this ChunkedArray.

See also extend for appends to the underlying memory

pub fn append_owned(&mut self, other: Self) -> PolarsResult<()>

Append in place. This is done by adding the chunks of other to this ChunkedArray.

See also extend for appends to the underlying memory

impl<T> ChunkedArray<T>
where T: PolarsDataType,

pub fn apply_nonnull_values_generic<'a, U, K, F>( &'a self, dtype: DataType, op: F, ) -> ChunkedArray<U>
where U: PolarsDataType, F: FnMut(T::Physical<'a>) -> K, U::Array: ArrayFromIterDtype<K> + ArrayFromIterDtype<Option<K>>,

Applies a function only to the non-null elements, propagating nulls.

pub fn try_apply_nonnull_values_generic<'a, U, K, F, E>( &'a self, op: F, ) -> Result<ChunkedArray<U>, E>
where U: PolarsDataType, F: FnMut(T::Physical<'a>) -> Result<K, E>, U::Array: ArrayFromIter<K> + ArrayFromIter<Option<K>>,

Applies a function only to the non-null elements, propagating nulls.

pub fn apply_into_string_amortized<'a, F>(&'a self, f: F) -> StringChunked
where F: FnMut(T::Physical<'a>, &mut String),

pub fn try_apply_into_string_amortized<'a, F, E>( &'a self, f: F, ) -> Result<StringChunked, E>
where F: FnMut(T::Physical<'a>, &mut String) -> Result<(), E>,

impl<T: PolarsNumericType> ChunkedArray<T>

pub fn cast_and_apply_in_place<F, S>(&self, f: F) -> ChunkedArray<S>
where F: Fn(S::Native) -> S::Native + Copy, S: PolarsNumericType,

Cast a numeric array to another numeric data type and apply a function in place. This saves an allocation.

pub fn apply_in_place<F>(self, f: F) -> Self
where F: Fn(T::Native) -> T::Native + Copy,

Cast a numeric array to another numeric data type and apply a function in place. This saves an allocation.

impl<T: PolarsNumericType> ChunkedArray<T>

pub fn apply_mut<F>(&mut self, f: F)
where F: Fn(T::Native) -> T::Native + Copy,

impl ChunkedArray<StringType>

pub fn apply_mut<'a, F>(&'a self, f: F) -> Self
where F: FnMut(&'a str) -> &'a str,

impl ChunkedArray<BinaryType>

pub fn apply_mut<'a, F>(&'a self, f: F) -> Self
where F: FnMut(&'a [u8]) -> &'a [u8],

impl ChunkedArray<StringType>

pub unsafe fn apply_views<F: FnMut(View, &str) -> View + Copy>( &self, update_view: F, ) -> Self

Safety

Update the views. All invariants of the views apply.

impl ChunkedArray<Float32Type>

Used to save compilation paths. Use carefully. Although this is safe, if misused it can lead to incorrect results.

pub fn apply_as_ints<F>(&self, f: F) -> Series
where F: Fn(&Series) -> Series,

impl ChunkedArray<Float64Type>

pub fn apply_as_ints<F>(&self, f: F) -> Series
where F: Fn(&Series) -> Series,

impl ChunkedArray<BooleanType>

pub fn num_trues(&self) -> usize

pub fn num_falses(&self) -> usize

pub fn first_true_idx(&self) -> Option<usize>

pub fn first_false_idx(&self) -> Option<usize>

impl<T: PolarsDataType> ChunkedArray<T>

pub fn len(&self) -> usize

Get the length of the ChunkedArray

pub fn null_count(&self) -> usize

Return the number of null values in the ChunkedArray.

pub unsafe fn set_null_count(&mut self, null_count: usize)

Set the null count directly.

This can be useful after mutably adjusting the validity of the underlying arrays.

Safety

The new null count must match the total null count of the underlying arrays.

pub fn is_empty(&self) -> bool

Check if ChunkedArray is empty.

pub fn rechunk(&self) -> Cow<'_, Self>

Rechunks this ChunkedArray, returning a new Cow::Owned ChunkedArray if it was rechunked or simply a Cow::Borrowed of itself if it was already a single chunk.

pub fn rechunk_mut(&mut self)

Rechunks this ChunkedArray in-place.

pub fn rechunk_validity(&self) -> Option<Bitmap>

pub fn with_validities(&mut self, validities: &[Option<Bitmap>])

pub fn split_at(&self, offset: i64) -> (Self, Self)

Split the array. The chunks are reallocated the underlying data slices are zero copy.

When offset is negative it will be counted from the end of the array. This method will never error, and will slice the best match when offset, or length is out of bounds

pub fn slice(&self, offset: i64, length: usize) -> Self

Slice the array. The chunks are reallocated the underlying data slices are zero copy.

When offset is negative it will be counted from the end of the array. This method will never error, and will slice the best match when offset, or length is out of bounds

pub fn limit(&self, num_elements: usize) -> Self
where Self: Sized,

Take a view of top n elements

pub fn head(&self, length: Option<usize>) -> Self
where Self: Sized,

Get the head of the ChunkedArray

pub fn tail(&self, length: Option<usize>) -> Self
where Self: Sized,

Get the tail of the ChunkedArray

pub fn prune_empty_chunks(&mut self)

Remove empty chunks.

impl ChunkedArray<StringType>

pub fn to_decimal(&self, infer_length: usize) -> PolarsResult<Series>

Available on crate feature dtype-decimal only.

Convert an StringChunked to a Series of DataType::Decimal. Scale needed for the decimal type are inferred. Parsing is not strict.
Scale inference assumes that all tested strings are well-formed numbers, and may produce unexpected results for scale if this is not the case.

If the decimal precision and scale are already known, consider using the cast method.

impl<T> ChunkedArray<T>
where T: PolarsNumericType,

pub fn extend(&mut self, other: &Self) -> PolarsResult<()>

Extend the memory backed by this array with the values from other.

Different from ChunkedArray::append which adds chunks to this ChunkedArray extend appends the data from other to the underlying PrimitiveArray and thus may cause a reallocation.

However if this does not cause a reallocation, the resulting data structure will not have any extra chunks and thus will yield faster queries.

Prefer extend over append when you want to do a query after a single append. For instance during online operations where you add n rows and rerun a query.

Prefer append over extend when you want to append many times before doing a query. For instance when you read in multiple files and when to store them in a single DataFrame. In the latter case finish the sequence of append operations with a rechunk.

impl<T> ChunkedArray<T>
where T: PolarsDataType,

pub fn for_each<'a, F>(&'a self, op: F)
where F: FnMut(Option<T::Physical<'a>>),

impl ChunkedArray<FixedSizeListType>

pub fn full_null_with_dtype( name: PlSmallStr, length: usize, inner_dtype: &DataType, width: usize, ) -> ArrayChunked

Available on crate feature dtype-array only.

impl ChunkedArray<ListType>

pub fn full_null_with_dtype( name: PlSmallStr, length: usize, inner_dtype: &DataType, ) -> ListChunked

impl ChunkedArray<UInt32Type>

pub fn with_nullable_idx<T, F: FnOnce(&IdxCa) -> T>( idx: &[NullableIdxSize], f: F, ) -> T

impl<T: PolarsDataType> ChunkedArray<T>

pub fn is_null(&self) -> BooleanChunked

Get a mask of the null values.

pub fn is_not_null(&self) -> BooleanChunked

Get a mask of the valid values.

impl<T> ChunkedArray<T>
where T: PolarsFloatType, T::Native: Float + IsFloat + SubAssign + Pow<T::Native, Output = T::Native>,

pub fn rolling_map_float<F>( &self, window_size: usize, f: F, ) -> PolarsResult<Self>
where F: FnMut(&mut ChunkedArray<T>) -> Option<T::Native>,

Available on crate feature rolling_window only.

Apply a rolling custom function. This is pretty slow because of dynamic dispatch.

impl ChunkedArray<BinaryType>

pub unsafe fn to_string_unchecked(&self) -> StringChunked

Safety

String is not validated

impl ChunkedArray<StringType>

pub fn as_binary(&self) -> BinaryChunked

impl ChunkedArray<BooleanType>

pub fn any(&self) -> bool

Returns whether any of the values in the column are true.

Null values are ignored.

pub fn all(&self) -> bool

Returns whether all values in the array are true.

Null values are ignored.

pub fn any_kleene(&self) -> Option<bool>

Returns whether any of the values in the column are true.

The output is unknown (None) if the array contains any null values and no true values.

pub fn all_kleene(&self) -> Option<bool>

Returns whether all values in the column are true.

The output is unknown (None) if the array contains any null values and no false values.

impl<T> ChunkedArray<T>
where T: PolarsFloatType, T::Native: Float,

pub fn is_nan(&self) -> BooleanChunked

pub fn is_not_nan(&self) -> BooleanChunked

pub fn is_finite(&self) -> BooleanChunked

pub fn is_infinite(&self) -> BooleanChunked

pub fn none_to_nan(&self) -> Self

Convert missing values to NaN values.

impl<T> ChunkedArray<T>
where T: PolarsFloatType, T::Native: Float + Canonical,

pub fn to_canonical(&self) -> Self

impl ChunkedArray<ListType>

pub fn par_iter(&self) -> impl ParallelIterator<Item = Option<Series>> + '_

pub fn par_iter_indexed( &mut self, ) -> impl IndexedParallelIterator<Item = Option<Series>> + '_

impl ChunkedArray<StringType>

pub fn par_iter_indexed( &self, ) -> impl IndexedParallelIterator<Item = Option<&str>>

pub fn par_iter(&self) -> impl ParallelIterator<Item = Option<&str>> + '_

impl<T> ChunkedArray<T>
where T: PolarsDataType,

pub fn iter(&self) -> impl PolarsIterator<Item = Option<T::Physical<'_>>>

impl<T> ChunkedArray<T>
where T: PolarsNumericType,

pub fn to_ndarray(&self) -> PolarsResult<ArrayView1<'_, T::Native>>

Available on crate feature ndarray only.

If data is aligned in a single chunk and has no Null values a zero copy view is returned as an ndarray

impl ChunkedArray<ListType>

pub fn to_ndarray<N>(&self) -> PolarsResult<Array2<N::Native>>
where N: PolarsNumericType,

Available on crate feature ndarray only.

If all nested Series have the same length, a 2 dimensional ndarray::Array is returned.

impl ChunkedArray<FixedSizeListType>

pub fn amortized_iter( &self, ) -> AmortizedListIter<'_, impl Iterator<Item = Option<ArrayBox>> + '_>

Available on crate feature dtype-array only.

This is an iterator over a ArrayChunked that save allocations. A Series is: 1. Arc<ChunkedArray> ChunkedArray is: 2. Vec< 3. ArrayRef>

The ArrayRef we indicated with 3. will be updated during iteration. The Series will be pinned in memory, saving an allocation for

Arc<..>

Vec<…>

Warning

Though memory safe in the sense that it will not read unowned memory, UB, or memory leaks this function still needs precautions. The returned should never be cloned or taken longer than a single iteration, as every call on next of the iterator will change the contents of that Series.

Safety

The lifetime of AmortSeries is bound to the iterator. Keeping it alive longer than the iterator is UB.

pub fn amortized_iter_with_name( &self, name: PlSmallStr, ) -> AmortizedListIter<'_, impl Iterator<Item = Option<ArrayBox>> + '_>

Available on crate feature dtype-array only.

This is an iterator over a ArrayChunked that save allocations. A Series is: 1. Arc<ChunkedArray> ChunkedArray is: 2. Vec< 3. ArrayRef>

The ArrayRef we indicated with 3. will be updated during iteration. The Series will be pinned in memory, saving an allocation for

Arc<..>

Vec<…>

If the returned AmortSeries is cloned, the local copy will be replaced and a new container will be set.

pub fn try_apply_amortized_to_list<F>(&self, f: F) -> PolarsResult<ListChunked>
where F: FnMut(AmortSeries) -> PolarsResult<Series>,

Available on crate feature dtype-array only.

pub unsafe fn apply_amortized_same_type<F>(&self, f: F) -> Self
where F: FnMut(AmortSeries) -> Series,

Available on crate feature dtype-array only.

Apply a closure F to each array.

Safety

Return series of F must has the same dtype and number of elements as input.

pub unsafe fn try_apply_amortized_same_type<F>( &self, f: F, ) -> PolarsResult<Self>
where F: FnMut(AmortSeries) -> PolarsResult<Series>,

Available on crate feature dtype-array only.

Try apply a closure F to each array.

Safety

Return series of F must has the same dtype and number of elements as input if it is Ok.

pub unsafe fn zip_and_apply_amortized_same_type<'a, T, F>( &'a self, ca: &'a ChunkedArray<T>, f: F, ) -> Self
where T: PolarsDataType, F: FnMut(Option<AmortSeries>, Option<T::Physical<'a>>) -> Option<Series>,

Available on crate feature dtype-array only.

Zip with a ChunkedArray then apply a binary function F elementwise.

Safety

pub fn apply_amortized_generic<F, K, V>(&self, f: F) -> ChunkedArray<V>
where V: PolarsDataType, F: FnMut(Option<AmortSeries>) -> Option<K> + Copy, V::Array: ArrayFromIter<Option<K>>,

Available on crate feature dtype-array only.

Apply a closure F elementwise.

pub fn try_apply_amortized_generic<F, K, V>( &self, f: F, ) -> PolarsResult<ChunkedArray<V>>
where V: PolarsDataType, F: FnMut(Option<AmortSeries>) -> PolarsResult<Option<K>> + Copy, V::Array: ArrayFromIter<Option<K>>,

Available on crate feature dtype-array only.

Try apply a closure F elementwise.

pub fn for_each_amortized<F>(&self, f: F)
where F: FnMut(Option<AmortSeries>),

Available on crate feature dtype-array only.

impl ChunkedArray<FixedSizeListType>

pub fn inner_dtype(&self) -> &DataType

Available on crate feature dtype-array only.

Get the inner data type of the fixed size list.

pub fn set_inner_dtype(&mut self, dtype: DataType)

Available on crate feature dtype-array only.

Panics

Panics if the physical representation of dtype differs the physical representation of the existing inner dtype.

pub fn width(&self) -> usize

Available on crate feature dtype-array only.

pub unsafe fn to_logical(&mut self, inner_dtype: DataType)

Available on crate feature dtype-array only.

Safety

The caller must ensure that the logical type given fits the physical type of the array.

pub fn to_physical_repr(&self) -> Cow<'_, ArrayChunked>

Available on crate feature dtype-array only.

Convert the datatype of the array into the physical datatype.

pub unsafe fn from_physical_unchecked( &self, to_inner_dtype: DataType, ) -> PolarsResult<Self>

Available on crate feature dtype-array only.

Convert a non-logical ArrayChunked back into a logical ArrayChunked without casting.

Safety

This can lead to invalid memory access in downstream code.

pub fn get_inner(&self) -> Series

Available on crate feature dtype-array only.

Get the inner values as Series

pub fn apply_to_inner( &self, func: &dyn Fn(Series) -> PolarsResult<Series>, ) -> PolarsResult<ArrayChunked>

Available on crate feature dtype-array only.

Ignore the list indices and apply func to the inner type as Series.

pub fn get_leaf_array(&self) -> Series

Available on crate feature dtype-array only.

Recurse nested types until we are at the leaf array.

impl<T> ChunkedArray<T>
where T: PolarsDataType, for<'a> <T::Array as StaticArray>::ValueT<'a>: AsRef<[u8]>,

pub fn to_bytes_hashes<'a>( &'a self, multithreaded: bool, hb: PlRandomState, ) -> Vec<Vec<BytesHash<'a>>>

impl<T> ChunkedArray<T>
where T: PolarsDataType,

pub fn with_chunk<A>(name: PlSmallStr, arr: A) -> Self
where A: Array, T: PolarsDataType<Array = A>,

pub fn with_chunk_like<A>(ca: &Self, arr: A) -> Self
where A: Array, T: PolarsDataType<Array = A>,

pub fn from_chunk_iter<I>(name: PlSmallStr, iter: I) -> Self
where I: IntoIterator, T: PolarsDataType<Array = <I as IntoIterator>::Item>, <I as IntoIterator>::Item: Array,

pub fn from_chunk_iter_like<I>(ca: &Self, iter: I) -> Self
where I: IntoIterator, T: PolarsDataType<Array = <I as IntoIterator>::Item>, <I as IntoIterator>::Item: Array,

pub fn try_from_chunk_iter<I, A, E>( name: PlSmallStr, iter: I, ) -> Result<Self, E>
where I: IntoIterator<Item = Result<A, E>>, T: PolarsDataType<Array = A>, A: Array,

pub unsafe fn from_chunks(name: PlSmallStr, chunks: Vec<ArrayRef>) -> Self

Create a new ChunkedArray from existing chunks.

Safety

The Arrow datatype of all chunks must match the PolarsDataType T.

pub unsafe fn with_chunks(&self, chunks: Vec<ArrayRef>) -> Self

Safety

The Arrow datatype of all chunks must match the PolarsDataType T.

pub unsafe fn from_chunks_and_dtype( name: PlSmallStr, chunks: Vec<ArrayRef>, dtype: DataType, ) -> Self

Create a new ChunkedArray from existing chunks.

Safety

The Arrow datatype of all chunks must match the PolarsDataType T.

pub fn full_null_like(ca: &Self, length: usize) -> Self

impl<T> ChunkedArray<T>
where T: PolarsNumericType,

pub fn from_vec(name: PlSmallStr, v: Vec<T::Native>) -> Self

Create a new ChunkedArray by taking ownership of the Vec. This operation is zero copy.

pub fn from_vec_validity( name: PlSmallStr, values: Vec<T::Native>, buffer: Option<Bitmap>, ) -> Self

Create a new ChunkedArray from a Vec and a validity mask.

pub unsafe fn mmap_slice(name: PlSmallStr, values: &[T::Native]) -> Self

Create a temporary ChunkedArray from a slice.

Safety

The lifetime will be bound to the lifetime of the slice. This will not be checked by the borrowchecker.

impl ChunkedArray<BooleanType>

pub unsafe fn mmap_slice( name: PlSmallStr, values: &[u8], offset: usize, len: usize, ) -> Self

Create a temporary ChunkedArray from a slice.

Safety

The lifetime will be bound to the lifetime of the slice. This will not be checked by the borrowchecker.

pub fn from_bitmap(name: PlSmallStr, bitmap: Bitmap) -> Self

impl ChunkedArray<ListType>

pub fn amortized_iter( &self, ) -> AmortizedListIter<'_, impl Iterator<Item = Option<ArrayBox>> + '_>

This is an iterator over a ListChunked that saves allocations. A Series is: 1. Arc<ChunkedArray> ChunkedArray is: 2. Vec< 3. ArrayRef>

The ArrayRef we indicated with 3. will be updated during iteration. The Series will be pinned in memory, saving an allocation for

Arc<..>

Vec<…>

If the returned AmortSeries is cloned, the local copy will be replaced and a new container will be set.

pub fn amortized_iter_with_name( &self, name: PlSmallStr, ) -> AmortizedListIter<'_, impl Iterator<Item = Option<ArrayBox>> + '_>

See amortized_iter.

pub fn apply_amortized_generic<F, K, V>(&self, f: F) -> ChunkedArray<V>
where V: PolarsDataType, F: FnMut(Option<AmortSeries>) -> Option<K> + Copy, V::Array: ArrayFromIter<Option<K>>,

Apply a closure F elementwise.

pub fn try_apply_amortized_generic<F, K, V>( &self, f: F, ) -> PolarsResult<ChunkedArray<V>>
where V: PolarsDataType, F: FnMut(Option<AmortSeries>) -> PolarsResult<Option<K>> + Copy, V::Array: ArrayFromIter<Option<K>>,

pub fn for_each_amortized<F>(&self, f: F)
where F: FnMut(Option<AmortSeries>),

pub fn zip_and_apply_amortized<'a, T, I, F>( &'a self, ca: &'a ChunkedArray<T>, f: F, ) -> Self
where T: PolarsDataType, &'a ChunkedArray<T>: IntoIterator<IntoIter = I>, I: TrustedLen<Item = Option<T::Physical<'a>>>, F: FnMut(Option<AmortSeries>, Option<T::Physical<'a>>) -> Option<Series>,

Zip with a ChunkedArray then apply a binary function F elementwise.

pub fn binary_zip_and_apply_amortized<'a, T, U, F>( &'a self, ca1: &'a ChunkedArray<T>, ca2: &'a ChunkedArray<U>, f: F, ) -> Self
where T: PolarsDataType, U: PolarsDataType, F: FnMut(Option<AmortSeries>, Option<T::Physical<'a>>, Option<U::Physical<'a>>) -> Option<Series>,

pub fn try_binary_zip_and_apply_amortized<'a, T, U, F>( &'a self, ca1: &'a ChunkedArray<T>, ca2: &'a ChunkedArray<U>, f: F, ) -> PolarsResult<Self>
where T: PolarsDataType, U: PolarsDataType, F: FnMut(Option<AmortSeries>, Option<T::Physical<'a>>, Option<U::Physical<'a>>) -> PolarsResult<Option<Series>>,

pub fn try_zip_and_apply_amortized<'a, T, I, F>( &'a self, ca: &'a ChunkedArray<T>, f: F, ) -> PolarsResult<Self>
where T: PolarsDataType, &'a ChunkedArray<T>: IntoIterator<IntoIter = I>, I: TrustedLen<Item = Option<T::Physical<'a>>>, F: FnMut(Option<AmortSeries>, Option<T::Physical<'a>>) -> PolarsResult<Option<Series>>,

pub fn apply_amortized<F>(&self, f: F) -> Self
where F: FnMut(AmortSeries) -> Series,

Apply a closure F elementwise.

pub fn try_apply_amortized<F>(&self, f: F) -> PolarsResult<Self>
where F: FnMut(AmortSeries) -> PolarsResult<Series>,

impl ChunkedArray<ListType>

pub fn inner_dtype(&self) -> &DataType

Get the inner data type of the list.

pub fn set_inner_dtype(&mut self, dtype: DataType)

Panics

Panics if the physical representation of dtype differs the physical representation of the existing inner dtype.

pub fn set_fast_explode(&mut self)

pub fn _can_fast_explode(&self) -> bool

pub unsafe fn to_logical(&mut self, inner_dtype: DataType)

Set the logical type of the ListChunked.

Safety

The caller must ensure that the logical type given fits the physical type of the array.

pub fn to_physical_repr(&self) -> Cow<'_, ListChunked>

Convert the datatype of the list into the physical datatype.

pub unsafe fn from_physical_unchecked( &self, to_inner_dtype: DataType, ) -> PolarsResult<ListChunked>

Convert a non-logical ListChunked back into a logical ListChunked without casting.

Safety

This can lead to invalid memory access in downstream code.

pub fn get_inner(&self) -> Series

Get the inner values as Series, ignoring the list offsets.

pub fn apply_to_inner( &self, func: &dyn Fn(Series) -> PolarsResult<Series>, ) -> PolarsResult<ListChunked>

Ignore the list indices and apply func to the inner type as Series.

impl ChunkedArray<Int128Type>

pub fn into_decimal_unchecked( self, precision: Option<usize>, scale: usize, ) -> DecimalChunked

Available on crate feature dtype-decimal only.

pub fn into_decimal( self, precision: Option<usize>, scale: usize, ) -> PolarsResult<DecimalChunked>

Available on crate feature dtype-decimal only.

impl<T> ChunkedArray<ObjectType<T>>
where T: PolarsObject,

pub fn new_from_vec(name: PlSmallStr, v: Vec<T>) -> Self

Available on crate feature object only.

pub fn new_from_vec_and_validity( name: PlSmallStr, v: Vec<T>, validity: Option<Bitmap>, ) -> Self

Available on crate feature object only.

pub fn new_empty(name: PlSmallStr) -> Self

Available on crate feature object only.

impl<T> ChunkedArray<ObjectType<T>>
where T: PolarsObject,

pub unsafe fn get_object_unchecked( &self, index: usize, ) -> Option<&dyn PolarsObjectSafe>

Available on crate feature object only.

Get a hold to an object that can be formatted or downcasted via the Any trait.

Safety

No bounds checks

pub fn get_object(&self, index: usize) -> Option<&dyn PolarsObjectSafe>

Available on crate feature object only.

Get a hold to an object that can be formatted or downcasted via the Any trait.

impl<T> ChunkedArray<T>
where T: PolarsNumericType, StandardUniform: Distribution<T::Native>,

pub fn init_rand(size: usize, null_density: f32, seed: Option<u64>) -> Self

Available on crate feature random only.

impl<T> ChunkedArray<T>
where T: PolarsDataType, ChunkedArray<T>: ChunkTake<IdxCa>,

pub fn sample_n( &self, n: usize, with_replacement: bool, shuffle: bool, seed: Option<u64>, ) -> PolarsResult<Self>

Available on crate feature random only.

Sample n datapoints from this ChunkedArray.

pub fn sample_frac( &self, frac: f64, with_replacement: bool, shuffle: bool, seed: Option<u64>, ) -> PolarsResult<Self>

Available on crate feature random only.

Sample a fraction between 0.0-1.0 of this ChunkedArray.

impl<T> ChunkedArray<T>
where T: PolarsNumericType, T::Native: Float,

pub fn rand_normal( name: PlSmallStr, length: usize, mean: f64, std_dev: f64, ) -> PolarsResult<Self>

Available on crate feature random only.

Create ChunkedArray with samples from a Normal distribution.

pub fn rand_standard_normal(name: PlSmallStr, length: usize) -> Self

Available on crate feature random only.

Create ChunkedArray with samples from a Standard Normal distribution.

pub fn rand_uniform( name: PlSmallStr, length: usize, low: f64, high: f64, ) -> Self

Available on crate feature random only.

Create ChunkedArray with samples from a Uniform distribution.

impl ChunkedArray<BooleanType>

pub fn rand_bernoulli( name: PlSmallStr, length: usize, p: f64, ) -> PolarsResult<Self>

Available on crate feature random only.

Create ChunkedArray with samples from a Bernoulli distribution.

impl<T: PolarsNumericType> ChunkedArray<T>

pub fn to_vec(&self) -> Vec<Option<T::Native>>

Convert to a Vec of Option<T::Native>.

pub fn to_vec_null_aware( &self, ) -> Either<Vec<T::Native>, Vec<Option<T::Native>>>

Convert to a Vec but don’t return Option<T::Native> if there are no null values

impl<T: PolarsDataType> ChunkedArray<T>

pub fn unpack_series_matching_type<'a>( &self, series: &'a Series, ) -> PolarsResult<&'a ChunkedArray<T>>

Series to ChunkedArray<T>

pub unsafe fn new_with_dims( field: Arc<Field>, chunks: Vec<ArrayRef>, length: usize, null_count: usize, ) -> Self

Create a new ChunkedArray and explicitly set its length and null_count.

Safety

The length and null_count must be correct.

pub fn unset_fast_explode_list(&mut self)

pub fn set_fast_explode_list(&mut self, value: bool)

pub fn get_fast_explode_list(&self) -> bool

pub fn get_flags(&self) -> StatisticsFlags

pub fn set_flags(&mut self, flags: StatisticsFlags)

Set flags for the ChunkedArray

pub fn is_sorted_flag(&self) -> IsSorted

pub fn retain_flags_from<U: PolarsDataType>( &mut self, from: &ChunkedArray<U>, retain_flags: StatisticsFlags, )

pub fn set_sorted_flag(&mut self, sorted: IsSorted)

Set the ‘sorted’ bit meta info.

pub fn with_sorted_flag(&self, sorted: IsSorted) -> Self

Set the ‘sorted’ bit meta info.

pub fn first_non_null(&self) -> Option<usize>

Get the index of the first non null value in this ChunkedArray.

pub fn last_non_null(&self) -> Option<usize>

Get the index of the last non null value in this ChunkedArray.

pub fn drop_nulls(&self) -> Self

pub fn iter_validities( &self, ) -> Map<Iter<'_, ArrayRef>, fn(&ArrayRef) -> Option<&Bitmap>>

Get the buffer of bits representing null values

pub fn has_nulls(&self) -> bool

Return if any the chunks in this ChunkedArray have nulls.

pub fn shrink_to_fit(&mut self)

Shrink the capacity of this array to fit its length.

pub fn clear(&self) -> Self

pub fn chunk_lengths(&self) -> ChunkLenIter<'_>

Returns an iterator over the lengths of the chunks of the array.

pub fn chunks(&self) -> &Vec<ArrayRef>

A reference to the chunks

pub unsafe fn chunks_mut(&mut self) -> &mut Vec<ArrayRef>

A mutable reference to the chunks

Safety

The caller must ensure to not change the DataType or length of any of the chunks. And the null_count remains correct.

pub fn is_optimal_aligned(&self) -> bool

Returns true if contains a single chunk and has no null values

pub fn dtype(&self) -> &DataType

Get data type of ChunkedArray.

pub fn name(&self) -> &PlSmallStr

Name of the ChunkedArray.

pub fn ref_field(&self) -> &Field

Get a reference to the field.

pub fn rename(&mut self, name: PlSmallStr)

Rename this ChunkedArray.

pub fn with_name(self, name: PlSmallStr) -> Self

Return this ChunkedArray with a new name.

impl<T> ChunkedArray<T>
where T: PolarsDataType,

pub fn get(&self, idx: usize) -> Option<T::Physical<'_>>

Get a single value from this ChunkedArray. If the return values is None this indicates a NULL value.

Panics

This function will panic if idx is out of bounds.

pub unsafe fn get_unchecked(&self, idx: usize) -> Option<T::Physical<'_>>

Get a single value from this ChunkedArray. If the return values is None this indicates a NULL value.

Safety

It is the callers responsibility that the idx < self.len().

pub unsafe fn value_unchecked(&self, idx: usize) -> T::Physical<'_>

Get a single value from this ChunkedArray. Null values are ignored and the returned value could be garbage if it was masked out by NULL. Note that the value always is initialized.

Safety

It is the callers responsibility that the idx < self.len().

pub fn first(&self) -> Option<T::Physical<'_>>

pub fn last(&self) -> Option<T::Physical<'_>>

impl ChunkedArray<ListType>

pub fn get_as_series(&self, idx: usize) -> Option<Series>

impl ChunkedArray<FixedSizeListType>

pub fn get_as_series(&self, idx: usize) -> Option<Series>

Available on crate feature dtype-array only.

impl<T> ChunkedArray<T>
where T: PolarsDataType,

pub fn match_chunks<I>(&self, chunk_id: I) -> Self
where I: Iterator<Item = usize>,

Should be used to match the chunk_id of another ChunkedArray.

Panics

It is the callers responsibility to ensure that this ChunkedArray has a single chunk.

impl<T> ChunkedArray<T>
where T: PolarsDataType,

pub fn layout(&self) -> ChunkedArrayLayout<'_, T>

impl<T> ChunkedArray<T>
where T: PolarsNumericType,

pub fn cont_slice(&self) -> PolarsResult<&[T::Native]>

Returns the values of the array as a contiguous slice.

pub fn data_views(&self) -> impl DoubleEndedIterator<Item = &[T::Native]>

Get slices of the underlying arrow data. NOTE: null values should be taken into account by the user of these slices as they are handled separately

pub fn into_no_null_iter( &self, ) -> impl '_ + Send + Sync + ExactSizeIterator<Item = T::Native> + DoubleEndedIterator + TrustedLen

impl<T: PolarsNumericType> ChunkedArray<T>

pub fn new_vec(name: PlSmallStr, v: Vec<T::Native>) -> Self

Specialization that prevents an allocation prefer this over ChunkedArray::new when you have a Vec<T::Native> and no null values.

impl<T: PolarsNumericType> ChunkedArray<T>

We cannot override the left hand side behaviour. So we create a trait LhsNumOps. This allows for 1.add(&Series)

pub fn lhs_sub<N: Num + NumCast>(&self, lhs: N) -> Self

Apply lhs - self

pub fn lhs_div<N: Num + NumCast>(&self, lhs: N) -> Self

Apply lhs / self

pub fn lhs_rem<N: Num + NumCast>(&self, lhs: N) -> Self

Apply lhs % self