polars_core
Module datatypes

Data types supported by Polars.

At the moment Polars doesn’t include all data types available by Arrow. The goal is to incrementally support more data types and prioritize these by usability.

See the AnyValue variants for the data types that are currently supported.

Modules

categoricaldtype-categorical

enum_dtype-categorical

revmapdtype-categorical

string_cachedtype-categorical

Structs

BinaryOffsetType

BinaryType

BooleanType

CatIterdtype-categorical

CategoricalChunkeddtype-categorical

CategoricalChunkedBuilderdtype-categorical

CategoricalType

CompatLevel

DateType

DatetimeType

DecimalType

Dimension

DurationType

FalseT

Field

Characterizes the name and the DataType of a column.

Float32Type

Float64Type

GlobalRevMapMergerdtype-categorical

Int8Type

Int16Type

Int32Type

Int64Type

Int128Typedtype-decimal

ListType

Logical

Maps a logical type to a chunked array implementation of the physical type. This saves a lot of compiler bloat and allows us to reuse functionality.

ObjectTypeobject

OwnedObjectobject

PlSmallStr

String type that inlines small strings.

StringType

TimeType

TrueT

UInt8Type

UInt16Type

UInt32Type

UInt64Type

Enums

AnyValue

ArrowDataType

The set of supported logical types in this crate.

ArrowTimeUnit

The time units defined in Arrow.

CategoricalOrderingdtype-categorical

DataType

ReshapeDimension

A dimension in a reshape.

RevMappingdtype-categorical

TimeUnit

UnknownKind

Constants

IDX_DTYPENon-bigidx

Statics

DTYPE_ENUM_KEY

DTYPE_ENUM_VALUE

Traits

ArrayCollectIterExt

ArrayFromIter

ArrayFromIterDtype

AsRefDataType

CategoricalMergeOperationdtype-categorical

GetAnyValue

InitHashMaps

InitHashMaps2

IntoScalar

LogicalType

NumericNative

PolarsDataType

Safety

PolarsFloatType

PolarsIntegerType

PolarsNumericType

StaticArray

Functions

call_categorical_merge_operationdtype-categorical

create_enum_dtypedtype-categorical

make_categoricals_compatibledtype-categorical

make_list_categoricals_compatibledtype-categorical

merge_dtypes

Type Aliases

BinaryChunked

BinaryOffsetChunked

BooleanChunked

DecimalChunkeddtype-decimal

FieldRef

Float32Chunked

Float64Chunked

IdxArr

IdxCaNon-bigidx

IdxTypeNon-bigidx

Int8Chunked

Int16Chunked

Int32Chunked

Int64Chunked

Int128Chunkeddtype-decimal

ListChunked

ObjectChunkedobject

PlHashMap

PlHashSet

PlIdHashMap

This hashmap uses an IdHasher

PlIndexMap

PlIndexSet

PlRandomState

StringChunked

TimeZone

UInt8Chunked

UInt16Chunked

UInt32Chunked

UInt64Chunked