Module polars_core::datatypes

Data types supported by Polars.

At the moment Polars doesn’t include all data types available by Arrow. The goal is to incrementally support more data types and prioritize these by usability.

See the AnyValue variants for the data types that are currently supported.

Modules

categoricaldtype-categorical

stringcachedtype-categorical

Structs

BinaryTypedtype-binary

BooleanType

CatIterdtype-categorical

CategoricalChunkeddtype-categorical

CategoricalChunkedBuilderdtype-categorical

CategoricalType

DateType

DatetimeType

DecimalType

DurationType

Field

Characterizes the name and the DataType of a column.

Float32Type

Float64Type

Int8Type

Int16Type

Int32Type

Int64Type

Int128Typedtype-i128

ListType

Logical

Maps a logical type to a a chunked array implementation of the physical type. This saves a lot of compiler bloat and allows us to reuse functionality.

ObjectTypeobject

OwnedObjectobject

StrHashLocaldtype-categorical

StructChunkeddtype-struct

This is logical type StructChunked that dispatches most logic to the fields implementations

TimeType

UInt8Type

UInt16Type

UInt32Type

UInt64Type

Utf8Type

Enums

AnyValue

ArrowDataType

The set of supported logical types in this crate.

ArrowTimeUnit

The time units defined in Arrow.

DataType

RevMappingdtype-categorical

RevMappingBuilderdtype-categorical

TimeUnit

Constants

IDX_DTYPENon-bigidx

NULL_DTYPE

Traits

GetAnyValue

InitHashMaps

LogicalType

NumericNative

PolarsDataType

PolarsFloatType

PolarsIntegerType

PolarsNumericType

PolarsSingleType

Any type that is not nested

Functions

merge_dtypes

Type Definitions

BinaryChunkeddtype-binary

BooleanChunked

CloudOptions

DateChunkeddtype-date

DatetimeChunkeddtype-datetime

DecimalChunkeddtype-i128

DurationChunkeddtype-duration

Float32Chunked

Float64Chunked

IdxArrNon-bigidx

IdxCaNon-bigidx

IdxSizeNon-bigidx

The type used by polars to index data.

IdxTypeNon-bigidx

Int8Chunked

Int16Chunked

Int32Chunked

Int64Chunked

Int128Chunkeddtype-i128

ListChunked

ObjectChunkedobject

PlHashMapprivate

PlHashSetprivate

PlIdHashMapprivate

This hashmap has the uses an IdHasher

PlIndexMapprivate

PlIndexSetprivate

TimeChunkeddtype-time

TimeZone

UInt8Chunked

UInt16Chunked

UInt32Chunked

UInt64Chunked

Utf8Chunked