polars_core::datatypes
Struct Logical

pub struct Logical<Logical: PolarsDataType, Physical: PolarsDataType>(pub ChunkedArray<Physical>, _, pub Option<DataType>);

Maps a logical type to a chunked array implementation of the physical type. This saves a lot of compiler bloat and allows us to reuse functionality.

Tuple Fields
0: ChunkedArray<Physical>2: Option<DataType>
Implementations

impl Logical<DecimalType, Int128Type>

pub fn precision(&self) -> Option<usize>

Available on crate feature dtype-decimal only.

pub fn scale(&self) -> usize

Available on crate feature dtype-decimal only.

pub fn to_scale(&self, scale: usize) -> PolarsResult<Cow<'_, Self>>

Available on crate feature dtype-decimal only.

impl<K: PolarsDataType, T: PolarsDataType> Logical<K, T>

pub fn new_logical<J: PolarsDataType>(ca: ChunkedArray<T>) -> Logical<J, T>

impl<K: PolarsDataType, T: PolarsDataType> Logical<K, T>
where Self: LogicalType,

pub fn physical(&self) -> &ChunkedArray<T>

pub fn field(&self) -> Field