polars_core::datatypes

Struct Logical

pub struct Logical<Logical: PolarsDataType, Physical: PolarsDataType> {
    pub phys: ChunkedArray<Physical>,
    pub dtype: DataType,
    /* private fields */
}

Maps a logical type to a chunked array implementation of the physical type. This saves a lot of compiler bloat and allows us to reuse functionality.

Fields
phys: ChunkedArray<Physical>dtype: DataType
Implementations

impl Logical<DecimalType, Int128Type>

pub fn precision(&self) -> Option<usize>

Available on crate feature dtype-decimal only.

pub fn scale(&self) -> usize

Available on crate feature dtype-decimal only.

pub fn to_scale(&self, scale: usize) -> PolarsResult<Cow<'_, Self>>

Available on crate feature dtype-decimal only.

impl<K: PolarsDataType, T: PolarsDataType> Logical<K, T>

pub fn new_logical(phys: ChunkedArray<T>, dtype: DataType) -> Logical<K, T>

impl<K: PolarsDataType, T: PolarsDataType> Logical<K, T>
where Self: LogicalType,

pub fn physical(&self) -> &ChunkedArray<T>

pub fn field(&self) -> Field