polars_core::datatypes

Struct Logical

pub struct Logical<Logical: PolarsDataType, Physical: PolarsDataType> {
    pub phys: ChunkedArray<Physical>,
    pub dtype: DataType,
    /* private fields */
}

Maps a logical type to a chunked array implementation of the physical type. This saves a lot of compiler bloat and allows us to reuse functionality.

Fields
phys: ChunkedArray<Physical>dtype: DataType
Implementations

impl<T: PolarsCategoricalType> Logical<T, <T as PolarsCategoricalType>::PolarsPhysical>

pub fn sort_with(&self, options: SortOptions) -> CategoricalChunked<T>

Available on crate feature dtype-categorical only.

pub fn sort(&self, descending: bool) -> CategoricalChunked<T>

Available on crate feature dtype-categorical only.

Returned a sorted ChunkedArray.

pub fn arg_sort(&self, options: SortOptions) -> IdxCa

Available on crate feature dtype-categorical only.

Retrieve the indexes needed to sort this array.

impl Logical<DecimalType, Int128Type>

pub fn precision(&self) -> Option<usize>

Available on crate feature dtype-decimal only.

pub fn scale(&self) -> usize

Available on crate feature dtype-decimal only.

pub fn to_scale(&self, scale: usize) -> PolarsResult<Cow<'_, Self>>

Available on crate feature dtype-decimal only.

impl<T: PolarsCategoricalType> Logical<T, <T as PolarsCategoricalType>::PolarsPhysical>

pub fn is_enum(&self) -> bool

Available on crate feature dtype-categorical only.

pub fn uses_lexical_ordering(&self) -> bool

Available on crate feature dtype-categorical only.

Return whether or not the CategoricalChunked uses the lexical order of the string values when sorting.

pub fn full_null_with_dtype( name: PlSmallStr, length: usize, dtype: DataType, ) -> Self

Available on crate feature dtype-categorical only.

pub fn from_cats_and_dtype( cat_ids: ChunkedArray<T::PolarsPhysical>, dtype: DataType, ) -> Self

Available on crate feature dtype-categorical only.

Create a CategoricalChunked from a physical array and dtype.

Checks that all the category ids are valid, mapping invalid ones to nulls.

pub unsafe fn from_cats_and_dtype_unchecked( cat_ids: ChunkedArray<T::PolarsPhysical>, dtype: DataType, ) -> Self

Available on crate feature dtype-categorical only.

Create a CategoricalChunked from a physical array and dtype.

Safety

It’s not checked that the indices are in-bounds or that the dtype is correct.

pub fn get_mapping(&self) -> &Arc<CategoricalMapping>

Available on crate feature dtype-categorical only.

Get a reference to the mapping of categorical types to the string values.

pub fn iter_str(&self) -> impl PolarsIterator<Item = Option<&str>>

Available on crate feature dtype-categorical only.

Create an Iterator that iterates over the &str values of the CategoricalChunked.

pub fn from_str_iter<'a, I: IntoIterator<Item = Option<&'a str>>>( name: PlSmallStr, dtype: DataType, strings: I, ) -> PolarsResult<Self>

Available on crate feature dtype-categorical only.

Converts from strings to this CategoricalChunked.

If this dtype is an Enum any non-existing strings get mapped to null.

pub fn to_arrow(&self, compat_level: CompatLevel) -> DictionaryArray<T::Native>

Available on crate feature dtype-categorical only.

impl<K: PolarsDataType, T: PolarsDataType> Logical<K, T>

pub unsafe fn new_logical( phys: ChunkedArray<T>, dtype: DataType, ) -> Logical<K, T>

Safety

You must uphold the logical types’ invariants.

impl<K: PolarsDataType, T: PolarsDataType> Logical<K, T>
where Self: LogicalType,

pub fn name(&self) -> &PlSmallStr

pub fn rename(&mut self, name: PlSmallStr)

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn null_count(&self) -> usize

pub fn has_nulls(&self) -> bool

pub fn is_null(&self) -> BooleanChunked

pub fn is_not_null(&self) -> BooleanChunked

pub fn split_at(&self, offset: i64) -> (Self, Self)

pub fn slice(&self, offset: i64, length: usize) -> Self

pub fn field(&self) -> Field

pub fn physical(&self) -> &ChunkedArray<T>

pub fn physical_mut(&mut self) -> &mut ChunkedArray<T>

pub fn into_physical(self) -> ChunkedArray<T>