Struct polars_core::datatypes::Logical

pub struct Logical<K: PolarsDataType, T: PolarsDataType>(pub ChunkedArray<T>, _);

Maps a logical type to a a chunked array implementation of the physical type. This saves a lot of compiler bloat and allows us to reuse functionality.

Tuple Fields
0: ChunkedArray<T>
Implementations

impl<K: PolarsDataType, T: PolarsDataType> Logical<K, T>

pub fn new<J: PolarsDataType>(ca: ChunkedArray<T>) -> Logical<J, T>

impl<K: PolarsDataType, T: PolarsDataType> Logical<K, T> where
Self: LogicalType,

pub fn field(&self) -> Field

impl Logical<DateType, Int32Type>

pub fn as_date_iter(
&self
) -> impl Iterator<Item = Option<NaiveDate>> + TrustedLen + '_

pub fn year(&self) -> Int32Chunked

Extract month from underlying NaiveDate representation. Returns the year number in the calendar date.

pub fn month(&self) -> UInt32Chunked

Extract month from underlying NaiveDateTime representation. Returns the month number starting from 1.

The return value ranges from 1 to 12.

pub fn weekday(&self) -> UInt32Chunked

Extract weekday from underlying NaiveDate representation. Returns the weekday number where monday = 0 and sunday = 6

pub fn week(&self) -> UInt32Chunked

Returns the ISO week number starting from 1. The return value ranges from 1 to 53. (The last week of year differs by years.)

pub fn day(&self) -> UInt32Chunked

Extract day from underlying NaiveDate representation. Returns the day of month starting from 1.

The return value ranges from 1 to 31. (The last day of month differs by months.)

pub fn ordinal(&self) -> UInt32Chunked

Returns the day of year starting from 1.

The return value ranges from 1 to 366. (The last day of year differs by years.)

pub fn strftime(&self, fmt: &str) -> Utf8Chunked

Format Date with a fmt rule. See chrono strftime/strptime.

pub fn new_from_naive_date(name: &str, v: &[NaiveDate]) -> Self

pub fn parse_from_str_slice(name: &str, v: &[&str], fmt: &str) -> Self

impl Logical<DatetimeType, Int64Type>

pub fn as_datetime_iter(
&self
) -> impl Iterator<Item = Option<NaiveDateTime>> + TrustedLen + '_

pub fn year(&self) -> Int32Chunked

Extract month from underlying NaiveDateTime representation. Returns the year number in the calendar date.

pub fn month(&self) -> UInt32Chunked

Extract month from underlying NaiveDateTime representation. Returns the month number starting from 1.

The return value ranges from 1 to 12.

pub fn weekday(&self) -> UInt32Chunked

Extract weekday from underlying NaiveDateTime representation. Returns the weekday number where monday = 0 and sunday = 6

pub fn week(&self) -> UInt32Chunked

Returns the ISO week number starting from 1. The return value ranges from 1 to 53. (The last week of year differs by years.)

pub fn day(&self) -> UInt32Chunked

Extract day from underlying NaiveDateTime representation. Returns the day of month starting from 1.

The return value ranges from 1 to 31. (The last day of month differs by months.)

pub fn hour(&self) -> UInt32Chunked

Extract hour from underlying NaiveDateTime representation. Returns the hour number from 0 to 23.

pub fn minute(&self) -> UInt32Chunked

Extract minute from underlying NaiveDateTime representation. Returns the minute number from 0 to 59.

pub fn second(&self) -> UInt32Chunked

Extract second from underlying NaiveDateTime representation. Returns the second number from 0 to 59.

pub fn nanosecond(&self) -> UInt32Chunked

Extract second from underlying NaiveDateTime representation. Returns the number of nanoseconds since the whole non-leap second. The range from 1,000,000,000 to 1,999,999,999 represents the leap second.

pub fn ordinal(&self) -> UInt32Chunked

Returns the day of year starting from 1.

The return value ranges from 1 to 366. (The last day of year differs by years.)

pub fn strftime(&self, fmt: &str) -> Utf8Chunked

Format Datetimewith a fmt rule. See chrono strftime/strptime.

pub fn new_from_naive_datetime(name: &str, v: &[NaiveDateTime]) -> Self

pub fn parse_from_str_slice(name: &str, v: &[&str], fmt: &str) -> Self

impl Logical<DatetimeType, Int64Type>

pub fn buckets(&self, resolution: TimeDelta) -> Self

impl Logical<DateType, Int32Type>

pub fn buckets(&self, resolution: TimeDelta) -> Self