polars_core::datatypes

Type Alias ArrayChunked

pub type ArrayChunked = ChunkedArray<FixedSizeListType>;

Available on crate feature dtype-array only.

Aliased Type

pub struct ArrayChunked { /* private fields */ }

Implementations

impl ArrayChunked

pub fn full_null_with_dtype( name: PlSmallStr, length: usize, inner_dtype: &DataType, width: usize, ) -> ArrayChunked

impl ArrayChunked

pub fn amortized_iter( &self, ) -> AmortizedListIter<'_, impl Iterator<Item = Option<ArrayBox>> + '_>

This is an iterator over a ArrayChunked that save allocations. A Series is: 1. Arc<ChunkedArray> ChunkedArray is: 2. Vec< 3. ArrayRef>

The ArrayRef we indicated with 3. will be updated during iteration. The Series will be pinned in memory, saving an allocation for

Arc<..>

Vec<…>

Warning

Though memory safe in the sense that it will not read unowned memory, UB, or memory leaks this function still needs precautions. The returned should never be cloned or taken longer than a single iteration, as every call on next of the iterator will change the contents of that Series.

Safety

The lifetime of AmortSeries is bound to the iterator. Keeping it alive longer than the iterator is UB.

pub fn amortized_iter_with_name( &self, name: PlSmallStr, ) -> AmortizedListIter<'_, impl Iterator<Item = Option<ArrayBox>> + '_>

This is an iterator over a ArrayChunked that save allocations. A Series is: 1. Arc<ChunkedArray> ChunkedArray is: 2. Vec< 3. ArrayRef>

The ArrayRef we indicated with 3. will be updated during iteration. The Series will be pinned in memory, saving an allocation for

Arc<..>

Vec<…>

If the returned AmortSeries is cloned, the local copy will be replaced and a new container will be set.

pub fn try_apply_amortized_to_list<F>(&self, f: F) -> PolarsResult<ListChunked>
where F: FnMut(AmortSeries) -> PolarsResult<Series>,

pub unsafe fn apply_amortized_same_type<F>(&self, f: F) -> Self
where F: FnMut(AmortSeries) -> Series,

Apply a closure F to each array.

Safety

Return series of F must has the same dtype and number of elements as input.

pub unsafe fn try_apply_amortized_same_type<F>( &self, f: F, ) -> PolarsResult<Self>
where F: FnMut(AmortSeries) -> PolarsResult<Series>,

Try apply a closure F to each array.

Safety

Return series of F must has the same dtype and number of elements as input if it is Ok.

pub unsafe fn zip_and_apply_amortized_same_type<'a, T, F>( &'a self, ca: &'a ChunkedArray<T>, f: F, ) -> Self
where T: PolarsDataType, F: FnMut(Option<AmortSeries>, Option<T::Physical<'a>>) -> Option<Series>,

Zip with a ChunkedArray then apply a binary function F elementwise.

Safety

pub fn apply_amortized_generic<F, K, V>(&self, f: F) -> ChunkedArray<V>
where V: PolarsDataType, F: FnMut(Option<AmortSeries>) -> Option<K> + Copy, V::Array: ArrayFromIter<Option<K>>,

Apply a closure F elementwise.

pub fn try_apply_amortized_generic<F, K, V>( &self, f: F, ) -> PolarsResult<ChunkedArray<V>>
where V: PolarsDataType, F: FnMut(Option<AmortSeries>) -> PolarsResult<Option<K>> + Copy, V::Array: ArrayFromIter<Option<K>>,

Try apply a closure F elementwise.

pub fn for_each_amortized<F>(&self, f: F)
where F: FnMut(Option<AmortSeries>),

impl ArrayChunked

pub fn inner_dtype(&self) -> &DataType

Get the inner data type of the fixed size list.

pub fn set_inner_dtype(&mut self, dtype: DataType)

Panics

Panics if the physical representation of dtype differs the physical representation of the existing inner dtype.

pub fn width(&self) -> usize

pub unsafe fn to_logical(&mut self, inner_dtype: DataType)

Safety

The caller must ensure that the logical type given fits the physical type of the array.

pub fn to_physical_repr(&self) -> Cow<'_, ArrayChunked>

Convert the datatype of the array into the physical datatype.

pub unsafe fn from_physical_unchecked( &self, to_inner_dtype: DataType, ) -> PolarsResult<Self>

Convert a non-logical ArrayChunked back into a logical ArrayChunked without casting.

Safety

This can lead to invalid memory access in downstream code.

pub fn get_inner(&self) -> Series

Get the inner values as Series

pub fn apply_to_inner( &self, func: &dyn Fn(Series) -> PolarsResult<Series>, ) -> PolarsResult<ArrayChunked>

Ignore the list indices and apply func to the inner type as Series.

pub fn get_leaf_array(&self) -> Series

Recurse nested types until we are at the leaf array.

impl ArrayChunked

pub fn get_as_series(&self, idx: usize) -> Option<Series>