polars_core::datatypes

Type Alias BooleanChunked

pub type BooleanChunked = ChunkedArray<BooleanType>;

Aliased Type

pub struct BooleanChunked { /* private fields */ }

Implementations

impl BooleanChunked

Booleans are cast to 1 or 0.

pub fn sum(&self) -> Option<IdxSize>

pub fn min(&self) -> Option<bool>

pub fn max(&self) -> Option<bool>

pub fn mean(&self) -> Option<f64>

impl BooleanChunked

pub fn num_trues(&self) -> usize

pub fn num_falses(&self) -> usize

pub fn first_true_idx(&self) -> Option<usize>

pub fn first_false_idx(&self) -> Option<usize>

impl BooleanChunked

pub fn any(&self) -> bool

Returns whether any of the values in the column are true.

Null values are ignored.

pub fn all(&self) -> bool

Returns whether all values in the array are true.

Null values are ignored.

pub fn any_kleene(&self) -> Option<bool>

Returns whether any of the values in the column are true.

The output is unknown (None) if the array contains any null values and no true values.

pub fn all_kleene(&self) -> Option<bool>

Returns whether all values in the column are true.

The output is unknown (None) if the array contains any null values and no false values.

impl BooleanChunked

pub unsafe fn mmap_slice( name: PlSmallStr, values: &[u8], offset: usize, len: usize, ) -> Self

Create a temporary ChunkedArray from a slice.

Safety

The lifetime will be bound to the lifetime of the slice. This will not be checked by the borrowchecker.

pub fn from_bitmap(name: PlSmallStr, bitmap: Bitmap) -> Self

impl BooleanChunked

pub fn rand_bernoulli( name: PlSmallStr, length: usize, p: f64, ) -> PolarsResult<Self>

Available on crate feature random only.

Create ChunkedArray with samples from a Bernoulli distribution.