Type Definition polars_core::datatypes::ListChunked

pub type ListChunked = ChunkedArray<ListType>;

Implementations

impl ListChunked

pub fn full_null_with_dtype( name: &str, length: usize, inner_dtype: &DataType ) -> ListChunked

impl ListChunked

pub fn par_iter(&self) -> impl ParallelIterator<Item = Option<Series>> + '_

pub fn par_iter_indexed( &mut self ) -> impl IndexedParallelIterator<Item = Option<Series>> + '_

impl ListChunked

pub unsafe fn amortized_iter( &self ) -> AmortizedListIter<'_, impl Iterator<Item = Option<ArrayBox>> + '_>

This is an iterator over a ListChunked that save allocations. A Series is: 1. Arc<ChunkedArray> ChunkedArray is: 2. Vec< 3. ArrayRef>

The ArrayRef we indicated with 3. will be updated during iteration. The Series will be pinned in memory, saving an allocation for

Arc<..>

Vec<…>

Warning

Though memory safe in the sense that it will not read unowned memory, UB, or memory leaks this function still needs precautions. The returned should never be cloned or taken longer than a single iteration, as every call on next of the iterator will change the contents of that Series.

Safety

The lifetime of UnstableSeries is bound to the iterator. Keeping it alive longer than the iterator is UB.

pub unsafe fn amortized_iter_with_name( &self, name: &str ) -> AmortizedListIter<'_, impl Iterator<Item = Option<ArrayBox>> + '_>

Safety

The lifetime of UnstableSeries is bound to the iterator. Keeping it alive longer than the iterator is UB.

pub fn apply_amortized_generic<'a, F, K, V>(&'a self, f: F) -> ChunkedArray<V>where V: PolarsDataType, F: FnMut(Option<UnstableSeries<'a>>) -> Option<K> + Copy, V::Array: ArrayFromIter<Option<K>>,

Apply a closure F elementwise.

pub fn for_each_amortized<'a, F>(&'a self, f: F)where F: FnMut(Option<UnstableSeries<'a>>),

pub fn zip_and_apply_amortized<'a, T, I, F>( &'a self, ca: &'a ChunkedArray<T>, f: F ) -> Selfwhere T: PolarsDataType, &'a ChunkedArray<T>: IntoIterator<IntoIter = I>, I: TrustedLen<Item = Option<T::Physical<'a>>>, F: FnMut(Option<UnstableSeries<'a>>, Option<T::Physical<'a>>) -> Option<Series>,

Zip with a ChunkedArray then apply a binary function F elementwise.

pub fn try_zip_and_apply_amortized<'a, T, I, F>( &'a self, ca: &'a ChunkedArray<T>, f: F ) -> PolarsResult<Self>where T: PolarsDataType, &'a ChunkedArray<T>: IntoIterator<IntoIter = I>, I: TrustedLen<Item = Option<T::Physical<'a>>>, F: FnMut(Option<UnstableSeries<'a>>, Option<T::Physical<'a>>) -> PolarsResult<Option<Series>>,

pub fn apply_amortized<'a, F>(&'a self, f: F) -> Selfwhere F: FnMut(UnstableSeries<'a>) -> Series,

Apply a closure F elementwise.

pub fn try_apply_amortized<'a, F>(&'a self, f: F) -> PolarsResult<Self>where F: FnMut(UnstableSeries<'a>) -> PolarsResult<Series>,

impl ListChunked

pub fn inner_dtype(&self) -> DataType

Get the inner data type of the list.

pub fn set_inner_dtype(&mut self, dtype: DataType)

pub fn set_fast_explode(&mut self)

pub fn _can_fast_explode(&self) -> bool

pub fn to_logical(&mut self, inner_dtype: DataType)

Set the logical type of the ListChunked.

pub fn get_inner(&self) -> Series

Get the inner values as Series, ignoring the list offsets.

pub fn apply_to_inner( &self, func: &dyn Fn(Series) -> PolarsResult<Series> ) -> PolarsResult<ListChunked>

Ignore the list indices and apply func to the inner type as Series.

impl ListChunked

pub fn get_as_series(&self, idx: usize) -> Option<Series>