Type Alias polars_core::datatypes::ListChunked

pub type ListChunked = ChunkedArray<ListType>;

Aliased Type

struct ListChunked { /* private fields */ }

Implementations

impl ListChunked

pub fn full_null_with_dtype( name: &str, length: usize, inner_dtype: &DataType ) -> ListChunked

impl ListChunked

pub fn par_iter(&self) -> impl ParallelIterator<Item = Option<Series>> + '_

pub fn par_iter_indexed( &mut self ) -> impl IndexedParallelIterator<Item = Option<Series>> + '_

impl ListChunked

pub fn to_ndarray<N>(&self) -> PolarsResult<Array2<N::Native>>
where N: PolarsNumericType,

Available on crate feature ndarray only.

If all nested Series have the same length, a 2 dimensional ndarray::Array is returned.

impl ListChunked

pub fn amortized_iter( &self ) -> AmortizedListIter<'_, impl Iterator<Item = Option<ArrayBox>> + '_>

This is an iterator over a ListChunked that saves allocations. A Series is: 1. Arc<ChunkedArray> ChunkedArray is: 2. Vec< 3. ArrayRef>

The ArrayRef we indicated with 3. will be updated during iteration. The Series will be pinned in memory, saving an allocation for

Arc<..>

Vec<…>

If the returned AmortSeries is cloned, the local copy will be replaced and a new container will be set.

pub fn amortized_iter_with_name( &self, name: &str ) -> AmortizedListIter<'_, impl Iterator<Item = Option<ArrayBox>> + '_>

See amortized_iter.

pub fn apply_amortized_generic<F, K, V>(&self, f: F) -> ChunkedArray<V>
where V: PolarsDataType, F: FnMut(Option<AmortSeries>) -> Option<K> + Copy, V::Array: ArrayFromIter<Option<K>>,

Apply a closure F elementwise.

pub fn try_apply_amortized_generic<F, K, V>( &self, f: F ) -> PolarsResult<ChunkedArray<V>>
where V: PolarsDataType, F: FnMut(Option<AmortSeries>) -> PolarsResult<Option<K>> + Copy, V::Array: ArrayFromIter<Option<K>>,

pub fn for_each_amortized<F>(&self, f: F)
where F: FnMut(Option<AmortSeries>),

pub fn zip_and_apply_amortized<'a, T, I, F>( &'a self, ca: &'a ChunkedArray<T>, f: F ) -> Self
where T: PolarsDataType, &'a ChunkedArray<T>: IntoIterator<IntoIter = I>, I: TrustedLen<Item = Option<T::Physical<'a>>>, F: FnMut(Option<AmortSeries>, Option<T::Physical<'a>>) -> Option<Series>,

Zip with a ChunkedArray then apply a binary function F elementwise.

pub fn binary_zip_and_apply_amortized<'a, T, U, F>( &'a self, ca1: &'a ChunkedArray<T>, ca2: &'a ChunkedArray<U>, f: F ) -> Self
where T: PolarsDataType, U: PolarsDataType, F: FnMut(Option<AmortSeries>, Option<T::Physical<'a>>, Option<U::Physical<'a>>) -> Option<Series>,

pub fn try_zip_and_apply_amortized<'a, T, I, F>( &'a self, ca: &'a ChunkedArray<T>, f: F ) -> PolarsResult<Self>
where T: PolarsDataType, &'a ChunkedArray<T>: IntoIterator<IntoIter = I>, I: TrustedLen<Item = Option<T::Physical<'a>>>, F: FnMut(Option<AmortSeries>, Option<T::Physical<'a>>) -> PolarsResult<Option<Series>>,

pub fn apply_amortized<F>(&self, f: F) -> Self
where F: FnMut(AmortSeries) -> Series,

Apply a closure F elementwise.

pub fn try_apply_amortized<F>(&self, f: F) -> PolarsResult<Self>
where F: FnMut(AmortSeries) -> PolarsResult<Series>,

impl ListChunked

pub fn inner_dtype(&self) -> &DataType

Get the inner data type of the list.

pub fn set_inner_dtype(&mut self, dtype: DataType)

pub fn set_fast_explode(&mut self)

pub fn _can_fast_explode(&self) -> bool

pub unsafe fn to_logical(&mut self, inner_dtype: DataType)

Set the logical type of the ListChunked.

Safety

The caller must ensure that the logical type given fits the physical type of the array.

pub fn get_inner(&self) -> Series

Get the inner values as Series, ignoring the list offsets.

pub fn iter_offsets(&self) -> impl Iterator<Item = i64> + '_

Returns an iterator over the offsets of this chunked array.

The offsets are returned as though the array consisted of a single chunk.

pub fn apply_to_inner( &self, func: &dyn Fn(Series) -> PolarsResult<Series> ) -> PolarsResult<ListChunked>

Ignore the list indices and apply func to the inner type as Series.

impl ListChunked

pub fn get_as_series(&self, idx: usize) -> Option<Series>