Type Definition polars_core::datatypes::ListChunked

pub type ListChunked = ChunkedArray<ListType>;

Implementations

impl ListChunked

pub fn append(&mut self, other: &Self)

impl ListChunked

pub fn downcast_iter(
&self
) -> impl Iterator<Item = &ListArray<i64>> + DoubleEndedIterator

pub fn downcast_chunks(&self) -> Chunks<'_, ListArray<i64>>

impl ListChunked

pub fn into_no_null_iter(
&self
) -> impl Iterator<Item = Series> + '_ + Send + Sync + ExactSizeIterator + DoubleEndedIterator + TrustedLen

impl ListChunked

pub fn amortized_iter(
&self
) -> AmortizedListIter<'_, impl Iterator<Item = Option<ArrayBox>> + '_>

This is an iterator over a ListChunked that save allocations. A Series is: 1. Arc ChunkedArray is: 2. Vec< 3. ArrayRef>

The ArrayRef we indicated with 3. will be updated during iteration. The Series will be pinned in memory, saving an allocation for

Arc<..>

Vec<…>

Warning

Though memory safe in the sense that it will not read unowned memory, UB, or memory leaks this function still needs precautions. The returned should never be cloned or taken longer than a single iteration, as every call on next of the iterator will change the contents of that Series.

pub fn apply_amortized<'a, F>(&'a self, f: F) -> Self where
F: FnMut(UnstableSeries<'a>) -> Series,

Apply a closure F elementwise.

pub fn try_apply_amortized<'a, F>(&'a self, f: F) -> Result<Self> where
F: FnMut(UnstableSeries<'a>) -> Result<Series>,

impl ListChunked

pub fn set_fast_explode(&mut self)

impl ListChunked

pub fn inner_dtype(&self) -> DataType

Get the inner data type of the list.