Type Definition polars_core::datatypes::ListChunked

pub type ListChunked = ChunkedArray<ListType>;

Implementations

impl ListChunked

pub fn to_ndarray<N>(&self) -> Result<Array2<N::Native>> where
N: PolarsNumericType,

If all nested Series have the same length, a 2 dimensional ndarray::Array is returned.

impl ListChunked

pub fn amortized_iter(
&self
) -> AmortizedListIter<'_, impl Iterator<Item = Option<ArrayBox>> + '_>

This is an iterator over a ListChunked that save allocations. A Series is: 1. Arc ChunkedArray is: 2. Vec< 3. ArrayRef>

The ArrayRef we indicated with 3. will be updated during iteration. The Series will be pinned in memory, saving an allocation for

Arc<..>

Vec<…>

Warning

Though memory safe in the sense that it will not read unowned memory, UB, or memory leaks this function still needs precautions. The returned should never be cloned or taken longer than a single iteration, as every call on next of the iterator will change the contents of that Series.

pub fn apply_amortized<'a, F>(&'a self, f: F) -> Self where
F: FnMut(UnstableSeries<'a>) -> Series,

Apply a closure F elementwise.

pub fn try_apply_amortized<'a, F>(&'a self, f: F) -> Result<Self> where
F: FnMut(UnstableSeries<'a>) -> Result<Series>,

impl ListChunked

pub fn lst_join(&self, separator: &str) -> Result<Utf8Chunked>

In case the inner dtype DataType::Utf8, the individual items will be joined into a single string separated by separator.

pub fn lst_max(&self) -> Series

pub fn lst_min(&self) -> Series

pub fn lst_sum(&self) -> Series

pub fn lst_mean(&self) -> Float64Chunked

pub fn lst_sort(&self, reverse: bool) -> ListChunked

pub fn lst_reverse(&self) -> ListChunked

pub fn lst_unique(&self) -> Result<ListChunked>

pub fn lst_lengths(&self) -> UInt32Chunked

pub fn lst_get(&self, idx: i64) -> Result<Series>

Get the value by index in the sublists. So index 0 would return the first item of every sublist and index -1 would return the last item of every sublist if an index is out of bounds, it will return a None.

pub fn lst_concat(&self, other: &[Series]) -> Result<ListChunked>

impl ListChunked

pub fn set_fast_explode(&mut self)

pub fn to_logical(&mut self, inner_dtype: DataType)

impl ListChunked

pub fn inner_dtype(&self) -> DataType

Get the inner data type of the list.