Type Definition polars_core::datatypes::ObjectChunked

pub type ObjectChunked<T> = ChunkedArray<ObjectType<T>>;

Available on crate feature object only.

Implementations

impl<T> ObjectChunked<T>where T: PolarsObject,

pub fn new_from_vec(name: &str, v: Vec<T>) -> Self

pub fn new_empty(name: &str) -> Self

impl<T> ObjectChunked<T>where T: PolarsObject,

pub unsafe fn get_object_unchecked( &self, index: usize ) -> Option<&dyn PolarsObjectSafe>

Get a hold to an object that can be formatted or downcasted via the Any trait.

Safety

No bounds checks

pub fn get_object(&self, index: usize) -> Option<&dyn PolarsObjectSafe>

Get a hold to an object that can be formatted or downcasted via the Any trait.