Type Alias polars_core::datatypes::Utf8Chunked

pub type Utf8Chunked = ChunkedArray<Utf8Type>;

Aliased Type

struct Utf8Chunked { /* private fields */ }

Implementations

impl Utf8Chunked

pub fn apply_mut<'a, F>(&'a self, f: F) -> Selfwhere F: FnMut(&'a str) -> &'a str,

pub fn apply_to_buffer<'a, F>(&'a self, f: F) -> Selfwhere F: FnMut(&'a str, &mut String),

Utility that reuses an string buffer to amortize allocations. Prefer this over an apply that returns an owned String.

impl Utf8Chunked

pub fn as_binary(&self) -> BinaryChunked

impl Utf8Chunked

pub fn par_iter_indexed( &self ) -> impl IndexedParallelIterator<Item = Option<&str>>

pub fn par_iter(&self) -> impl ParallelIterator<Item = Option<&str>> + '_

impl<T> ChunkedArray<T>where T: PolarsDataType<Structure = Flat>, for<'a> T::Physical<'a>: TotalOrd,

pub fn append(&mut self, other: &Self)

Append in place. This is done by adding the chunks of other to this ChunkedArray.

See also extend for appends to the underlying memory

impl<T> ChunkedArray<T>where T: PolarsDataType,

pub fn apply_values_generic<'a, U, K, F>(&'a self, op: F) -> ChunkedArray<U>where U: PolarsDataType, F: FnMut(T::Physical<'a>) -> K, U::Array: ArrayFromIter<K>,

pub fn try_apply_values_generic<'a, U, K, F, E>( &'a self, op: F ) -> Result<ChunkedArray<U>, E>where U: PolarsDataType, F: FnMut(T::Physical<'a>) -> Result<K, E>, U::Array: ArrayFromIter<K>,

Applies a function to all elements, regardless of whether they are null or not, after which the null mask is copied from the original array.

pub fn apply_nonnull_values_generic<'a, U, K, F>( &'a self, dtype: DataType, op: F ) -> ChunkedArray<U>where U: PolarsDataType, F: FnMut(T::Physical<'a>) -> K, U::Array: ArrayFromIterDtype<K> + ArrayFromIterDtype<Option<K>>,

Applies a function only to the non-null elements, propagating nulls.

pub fn try_apply_nonnull_values_generic<'a, U, K, F, E>( &'a self, op: F ) -> Result<ChunkedArray<U>, E>where U: PolarsDataType, F: FnMut(T::Physical<'a>) -> Result<K, E>, U::Array: ArrayFromIter<K> + ArrayFromIter<Option<K>>,

Applies a function only to the non-null elements, propagating nulls.

pub fn apply_generic<'a, U, K, F>(&'a self, op: F) -> ChunkedArray<U>where U: PolarsDataType, F: FnMut(Option<T::Physical<'a>>) -> Option<K>, U::Array: ArrayFromIter<Option<K>>,

pub fn try_apply_generic<'a, U, K, F, E>( &'a self, op: F ) -> Result<ChunkedArray<U>, E>where U: PolarsDataType, F: FnMut(Option<T::Physical<'a>>) -> Result<Option<K>, E> + Copy, U::Array: ArrayFromIter<Option<K>>,

impl<T: PolarsNumericType> ChunkedArray<T>

pub fn cast_and_apply_in_place<F, S>(&self, f: F) -> ChunkedArray<S>where F: Fn(S::Native) -> S::Native + Copy, S: PolarsNumericType,

Cast a numeric array to another numeric data type and apply a function in place. This saves an allocation.

pub fn apply_in_place<F>(self, f: F) -> Selfwhere F: Fn(T::Native) -> T::Native + Copy,

Cast a numeric array to another numeric data type and apply a function in place. This saves an allocation.

impl<T: PolarsNumericType> ChunkedArray<T>

pub fn apply_mut<F>(&mut self, f: F)where F: Fn(T::Native) -> T::Native + Copy,

impl<T: PolarsDataType> ChunkedArray<T>

pub fn len(&self) -> usize

Get the length of the ChunkedArray

pub fn null_count(&self) -> usize

Count the null values.

pub fn is_empty(&self) -> bool

Check if ChunkedArray is empty.

pub fn rechunk(&self) -> Self

pub fn slice(&self, offset: i64, length: usize) -> Self

Slice the array. The chunks are reallocated the underlying data slices are zero copy.

When offset is negative it will be counted from the end of the array. This method will never error, and will slice the best match when offset, or length is out of bounds

pub fn limit(&self, num_elements: usize) -> Selfwhere Self: Sized,

Take a view of top n elements

pub fn head(&self, length: Option<usize>) -> Selfwhere Self: Sized,

Get the head of the ChunkedArray

pub fn tail(&self, length: Option<usize>) -> Selfwhere Self: Sized,

Get the tail of the ChunkedArray

impl<T> ChunkedArray<T>where T: PolarsNumericType,

pub fn extend(&mut self, other: &Self)

Extend the memory backed by this array with the values from other.

Different from ChunkedArray::append which adds chunks to this ChunkedArray extend appends the data from other to the underlying PrimitiveArray and thus may cause a reallocation.

However if this does not cause a reallocation, the resulting data structure will not have any extra chunks and thus will yield faster queries.

Prefer extend over append when you want to do a query after a single append. For instance during online operations where you add n rows and rerun a query.

Prefer append over extend when you want to append many times before doing a query. For instance when you read in multiple files and when to store them in a single DataFrame. In the latter case finish the sequence of append operations with a rechunk.

impl<T> ChunkedArray<T>where T: PolarsDataType,

pub fn for_each<'a, F>(&'a self, op: F)where F: FnMut(Option<T::Physical<'a>>),

impl<T: PolarsDataType> ChunkedArray<T>

pub fn is_null(&self) -> BooleanChunked

Get a mask of the null values.

pub fn is_not_null(&self) -> BooleanChunked

Get a mask of the valid values.

impl<T: PolarsNumericType> ChunkedArray<T>

pub fn tile(&self, n: usize) -> Self

impl<T> ChunkedArray<T>where T: PolarsNumericType,

pub fn not_equal_and_validity(&self, rhs: &ChunkedArray<T>) -> BooleanChunked

impl<T> ChunkedArray<T>where T: PolarsFloatType, T::Native: Float,

pub fn is_nan(&self) -> BooleanChunked

pub fn is_not_nan(&self) -> BooleanChunked

pub fn is_finite(&self) -> BooleanChunked

pub fn is_infinite(&self) -> BooleanChunked

pub fn none_to_nan(&self) -> Self

Convert missing values to NaN values.

impl<T> ChunkedArray<T>where T: PolarsDataType,

pub fn with_chunk<A>(name: &str, arr: A) -> Selfwhere A: Array, T: PolarsDataType<Array = A>,

pub fn from_chunk_iter<I>(name: &str, iter: I) -> Selfwhere I: IntoIterator, T: PolarsDataType<Array = <I as IntoIterator>::Item>, <I as IntoIterator>::Item: Array,

pub fn from_chunk_iter_like<I>(ca: &Self, iter: I) -> Selfwhere I: IntoIterator, T: PolarsDataType<Array = <I as IntoIterator>::Item>, <I as IntoIterator>::Item: Array,

pub fn try_from_chunk_iter<I, A, E>(name: &str, iter: I) -> Result<Self, E>where I: IntoIterator<Item = Result<A, E>>, T: PolarsDataType<Array = A>, A: Array,

pub unsafe fn from_chunks(name: &str, chunks: Vec<ArrayRef>) -> Self

Create a new ChunkedArray from existing chunks.

Safety

The Arrow datatype of all chunks must match the PolarsDataType T.

pub unsafe fn with_chunks(&self, chunks: Vec<ArrayRef>) -> Self

Safety

The Arrow datatype of all chunks must match the PolarsDataType T.

pub unsafe fn from_chunks_and_dtype( name: &str, chunks: Vec<ArrayRef>, dtype: DataType ) -> Self

Create a new ChunkedArray from existing chunks.

Safety

The Arrow datatype of all chunks must match the PolarsDataType T.

impl<T> ChunkedArray<T>where T: PolarsNumericType,

pub fn from_vec(name: &str, v: Vec<T::Native>) -> Self

Create a new ChunkedArray by taking ownership of the Vec. This operation is zero copy.

pub fn from_vec_validity( name: &str, values: Vec<T::Native>, buffer: Option<Bitmap> ) -> Self

Create a new ChunkedArray from a Vec and a validity mask.

pub unsafe fn mmap_slice(name: &str, values: &[T::Native]) -> Self

Create a temporary ChunkedArray from a slice.

Safety

The lifetime will be bound to the lifetime of the slice. This will not be checked by the borrowchecker.

impl<T> ChunkedArray<T>where T: PolarsNumericType, Standard: Distribution<T::Native>,

pub fn init_rand(size: usize, null_density: f32, seed: Option<u64>) -> Self

impl<T> ChunkedArray<T>where T: PolarsDataType, ChunkedArray<T>: ChunkTake<IdxCa>,

pub fn sample_n( &self, n: usize, with_replacement: bool, shuffle: bool, seed: Option<u64> ) -> PolarsResult<Self>

Sample n datapoints from this ChunkedArray.

pub fn sample_frac( &self, frac: f64, with_replacement: bool, shuffle: bool, seed: Option<u64> ) -> PolarsResult<Self>

Sample a fraction between 0.0-1.0 of this ChunkedArray.

impl<T> ChunkedArray<T>where T: PolarsNumericType, T::Native: Float,

pub fn rand_normal( name: &str, length: usize, mean: f64, std_dev: f64 ) -> PolarsResult<Self>

Create ChunkedArray with samples from a Normal distribution.

pub fn rand_standard_normal(name: &str, length: usize) -> Self

Create ChunkedArray with samples from a Standard Normal distribution.

pub fn rand_uniform(name: &str, length: usize, low: f64, high: f64) -> Self

Create ChunkedArray with samples from a Uniform distribution.

impl<T: PolarsNumericType> ChunkedArray<T>

pub fn to_vec(&self) -> Vec<Option<T::Native>>

Convert to a Vec of Option<T::Native>.

pub fn to_vec_null_aware( &self ) -> Either<Vec<T::Native>, Vec<Option<T::Native>>>

Convert to a Vec but don’t return Option<T::Native> if there are no null values

impl<T: PolarsDataType> ChunkedArray<T>

pub fn unset_fast_explode_list(&mut self)

pub fn get_flags(&self) -> Settings

pub fn is_sorted_flag(&self) -> IsSorted

pub fn set_sorted_flag(&mut self, sorted: IsSorted)

Set the ‘sorted’ bit meta info.

pub fn first_non_null(&self) -> Option<usize>

Get the index of the first non null value in this ChunkedArray.

pub fn last_non_null(&self) -> Option<usize>

Get the index of the last non null value in this ChunkedArray.

pub fn iter_validities( &self ) -> Map<Iter<'_, ArrayRef>, fn(_: &ArrayRef) -> Option<&Bitmap>>

Get the buffer of bits representing null values

pub fn has_validity(&self) -> bool

Return if any the chunks in this ChunkedArray have a validity bitmap. no bitmap means no null values.

pub fn shrink_to_fit(&mut self)

Shrink the capacity of this array to fit its length.

pub fn clear(&self) -> Self

pub fn unpack_series_matching_type( &self, series: &Series ) -> PolarsResult<&ChunkedArray<T>>

Series to ChunkedArray<T>

pub fn chunk_id(&self) -> ChunkIdIter<'_>

Unique id representing the number of chunks

pub fn chunks(&self) -> &Vec<ArrayRef>

A reference to the chunks

pub unsafe fn chunks_mut(&mut self) -> &mut Vec<ArrayRef>

A mutable reference to the chunks

Safety

The caller must ensure to not change the DataType or length of any of the chunks. And the null_count remains correct.

pub fn is_optimal_aligned(&self) -> bool

Returns true if contains a single chunk and has no null values

pub fn dtype(&self) -> &DataType

Get data type of ChunkedArray.

pub fn name(&self) -> &str

Name of the ChunkedArray.

pub fn ref_field(&self) -> &Field

Get a reference to the field.

pub fn rename(&mut self, name: &str)

Rename this ChunkedArray.

pub fn with_name(self, name: &str) -> Self

Return this ChunkedArray with a new name.

impl<T> ChunkedArray<T>where T: PolarsDataType,

pub fn get(&self, idx: usize) -> Option<T::Physical<'_>>

pub unsafe fn get_unchecked(&self, idx: usize) -> Option<T::Physical<'_>>

Safety

It is the callers responsibility that the idx < self.len().

pub unsafe fn value_unchecked(&self, idx: usize) -> T::Physical<'_>

Safety

It is the callers responsibility that the idx < self.len().

pub fn last(&self) -> Option<T::Physical<'_>>

impl<T> ChunkedArray<T>where T: PolarsDataType,

pub fn layout(&self) -> ChunkedArrayLayout<'_, T>

impl<T> ChunkedArray<T>where T: PolarsNumericType,

pub fn cont_slice(&self) -> PolarsResult<&[T::Native]>

Contiguous slice

pub fn data_views( &self ) -> impl Iterator<Item = &[T::Native]> + DoubleEndedIterator

Get slices of the underlying arrow data. NOTE: null values should be taken into account by the user of these slices as they are handled separately

pub fn into_no_null_iter( &self ) -> impl Iterator<Item = T::Native> + '_ + Send + Sync + ExactSizeIterator + DoubleEndedIterator + TrustedLen

impl<T> ChunkedArray<T>where T: PolarsIntegerType, T::Native: ToPrimitive + FromPrimitive + Debug,

pub fn group_tuples_perfect( &self, max: usize, multithreaded: bool, group_capacity: usize ) -> GroupsProxy

impl<T: PolarsNumericType> ChunkedArray<T>

pub fn new_vec(name: &str, v: Vec<T::Native>) -> Self

Specialization that prevents an allocation prefer this over ChunkedArray::new when you have a Vec<T::Native> and no null values.

impl<T> ChunkedArray<T>where T: PolarsNumericType, ChunkedArray<T>: IntoSeries,

We cannot override the left hand side behaviour. So we create a trait LhsNumOps. This allows for 1.add(&Series)

pub fn lhs_sub<N: Num + NumCast>(&self, lhs: N) -> Self

Apply lhs - self

pub fn lhs_div<N: Num + NumCast>(&self, lhs: N) -> Self

Apply lhs / self

pub fn lhs_rem<N: Num + NumCast>(&self, lhs: N) -> Self

Apply lhs % self