Type Definition polars_core::datatypes::Utf8Chunked

pub type Utf8Chunked = ChunkedArray<Utf8Type>;

Implementations

impl Utf8Chunked

pub fn json_path_match(&self, json_path: &str) -> Result<Utf8Chunked>

Extract json path, first match Refer to https://goessner.net/articles/JsonPath/

impl Utf8Chunked

pub fn hex_decode(&self, strict: Option<bool>) -> Result<Utf8Chunked>

pub fn hex_encode(&self) -> Utf8Chunked

pub fn base64_decode(&self, strict: Option<bool>) -> Result<Utf8Chunked>

pub fn base64_encode(&self) -> Utf8Chunked

impl Utf8Chunked

pub fn str_lengths(&self) -> UInt32Chunked

Get the length of the string values.

pub fn contains(&self, pat: &str) -> Result<BooleanChunked>

Check if strings contain a regex pattern

pub fn replace(&self, pat: &str, val: &str) -> Result<Utf8Chunked>

Replace the leftmost (sub)string by a regex pattern

pub fn replace_all(&self, pat: &str, val: &str) -> Result<Utf8Chunked>

Replace all (sub)strings by a regex pattern

pub fn extract(&self, pat: &str, group_index: usize) -> Result<Utf8Chunked>

Extract the nth capture group from pattern

pub fn to_lowercase(&self) -> Utf8Chunked

Modify the strings to their lowercase equivalent

pub fn to_uppercase(&self) -> Utf8Chunked

Modify the strings to their uppercase equivalent

pub fn concat(&self, other: &Utf8Chunked) -> Self

Concat with the values from a second Utf8Chunked

pub fn str_slice(&self, start: i64, length: Option<u64>) -> Result<Self>

Slice the string values Determines a substring starting from start and with optional length length of each of the elements in array. start can be negative, in which case the start counts from the end of the string.

impl Utf8Chunked

pub fn as_time(&self, fmt: Option<&str>) -> Result<TimeChunked>

pub fn as_date_not_exact(&self, fmt: Option<&str>) -> Result<DateChunked>

pub fn as_datetime_not_exact(
&self,
fmt: Option<&str>,
tu: TimeUnit
) -> Result<DatetimeChunked>

pub fn as_date(&self, fmt: Option<&str>) -> Result<DateChunked>

pub fn as_datetime(
&self,
fmt: Option<&str>,
tu: TimeUnit
) -> Result<DatetimeChunked>