polars_core::error

Module signals

Structs

KeyboardInterrupt

Python hooks SIGINT to instead generate a KeyboardInterrupt exception. So we do the same to try and abort long-running computations and return to Python so that the Python exception can be generated.

Functions

catch_keyboard_interrupt

Runs the passed function, catching any KeyboardInterrupts if they occur while running the function.

register_polars_keyboard_interrupt_hook

try_raise_keyboard_interrupt

Checks if the keyboard interrupt flag is set, and if yes panics as a keyboard interrupt. This function is very cheap.