polars_core::frame::group_by

Trait IntoGroupsType

pub trait IntoGroupsType {
    // Provided method
    fn group_tuples(
        &self,
        _multithreaded: bool,
        _sorted: bool,
    ) -> PolarsResult<GroupsType> { ... }
}

Available on crate feature algorithm_group_by only.

Used to create the tuples for a group_by operation.

Provided Methods

fn group_tuples( &self, _multithreaded: bool, _sorted: bool, ) -> PolarsResult<GroupsType>

Create the tuples need for a group_by operation. * The first value in the tuple is the first index of the group. * The second value in the tuple is the indexes of the groups including the first value.