Module polars_core::prelude

Everything you need to get started with Polars.

Re-exports

pub use crate::chunked_array::arithmetic::Pow;

pub use crate::chunked_array::builder::ChunkedBuilder;

pub use crate::chunked_array::builder::ListBooleanChunkedBuilder;

pub use crate::chunked_array::builder::ListBuilderTrait;

pub use crate::chunked_array::builder::ListPrimitiveChunkedBuilder;

pub use crate::chunked_array::builder::ListUtf8ChunkedBuilder;

pub use crate::chunked_array::builder::NewChunkedArray;

pub use crate::chunked_array::iterator::PolarsIterator;

pub use crate::chunked_array::ops::*;

pub use crate::chunked_array::ChunkedArray;

pub use crate::datatypes;

pub use crate::datatypes::*;

pub use crate::df;

pub use crate::error::PolarsError;

pub use crate::error::Result;

pub use crate::frame::explode::MeltArgs;

pub use crate::frame::hash_join::JoinType;

pub use crate::frame::*;

pub use crate::schema::*;

pub use crate::series::Series;

pub use crate::testing::*;

pub use crate::utils::IntoVec;

pub use crate::chunked_array::temporal::conversion::*;

Modules

default_arrays

list

Structs

Arc

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

ArrowField

Represents Arrow’s metadata of a “column”.

ArrowSchema

An ordered sequence of Fields with associated Metadata.

BooleanChunkedBuilder

CatIter

CategoricalChunked

CategoricalChunkedBuilder

GroupsIdx

Indexes of the groups, the first index is stored separately. this make sorting fast.

PrimitiveChunkedBuilder

Utf8ChunkedBuilder

Enums

GroupsProxy

QuantileInterpolOptions

RevMapping

RevMappingBuilder

Traits

ArrowGetItem

ChunkAggSeries

Aggregations that return Series of unit length. Those can be used in broadcasting operations.

FromData

FromDataUtf8

IndexToUsize

IntoSeries

LhsNumOps

ListFromIter

MutableBitmapExtension

NamedFrom

NamedFromOwned

NumOpsDispatch

PolarsArray

QuantileAggSeries

SeriesTrait

ValueSize

VarAggSeries

VecHash

Type Definitions

GroupsSlice

Every group is indicated by an array where the

IdxArr

IdxSize

The type used by polars to index data.

LargeListArray

LargeStringArray