Module polars_core::prelude

Everything you need to get started with Polars.

Re-exports

pub use crate::chunked_array::arithmetic::ArithmeticChunked;

pub use crate::chunked_array::builder::BinaryChunkedBuilder;

pub use crate::chunked_array::builder::BooleanChunkedBuilder;

pub use crate::chunked_array::builder::ChunkedBuilder;

pub use crate::chunked_array::builder::ListBinaryChunkedBuilder;

pub use crate::chunked_array::builder::ListBooleanChunkedBuilder;

pub use crate::chunked_array::builder::ListBuilderTrait;

pub use crate::chunked_array::builder::ListPrimitiveChunkedBuilder;

pub use crate::chunked_array::builder::ListStringChunkedBuilder;

pub use crate::chunked_array::builder::NewChunkedArray;

pub use crate::chunked_array::builder::PrimitiveChunkedBuilder;

pub use crate::chunked_array::builder::StringChunkedBuilder;

pub use crate::chunked_array::collect::ChunkedCollectInferIterExt;

pub use crate::chunked_array::collect::ChunkedCollectIterExt;

pub use crate::chunked_array::iterator::PolarsIterator;

pub use crate::chunked_array::object::PolarsObject;

object

pub use crate::chunked_array::ChunkedArray;

pub use crate::frame::explode::MeltArgs;

pub use crate::frame::DataFrame;

pub use crate::frame::UniqueKeepStrategy;

pub use crate::hashing::VecHash;

pub use crate::series::arithmetic::LhsNumOps;

pub use crate::series::arithmetic::NumOpsDispatch;

pub use crate::series::IntoSeries;

pub use crate::series::Series;

pub use crate::series::SeriesTrait;

pub use crate::utils::IntoVec;

pub use crate::datatypes;

pub use crate::chunked_array::ops::*;

pub use crate::chunked_array::temporal::conversion::*;

temporal

pub use crate::datatypes::*;

pub use crate::frame::group_by::*;

algorithm_group_by

pub use crate::schema::*;

Modules

default_arrays

list

null

revmapdtype-categorical

slice

string_cachedtype-categorical

utf8

Macros

df

polars_bail

polars_ensure

polars_err

polars_warn

Structs

Arc

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

ArrowField

Represents Arrow’s metadata of a “column”.

ArrowSchema

An ordered sequence of Fields with associated [Metadata].

CatIterdtype-categorical

CategoricalChunkeddtype-categorical

CategoricalChunkedBuilderdtype-categorical

ChunkId

GlobalRevMapMergerdtype-categorical

NullableIdxSize

RollingOptionsFixedWindowrolling_window

RollingQuantileParams

RollingVarParams

StringCacheHolderdtype-categorical

Enable the global string cache as long as the object is alive (RAII).

Enums

Ambiguous

CategoricalOrderingdtype-categorical

IndexOrderndarray

NonExistent

PolarsError

QuantileInterpolOptions

RevMappingdtype-categorical

Traits

ArrayCollectIterExt

CategoricalMergeOperationdtype-categorical

ChunkAggSeries

Aggregations that return Series of unit length. Those can be used in broadcasting operations.

FromData

FromDataBinary

FromDataUtf8

IndexToUsize

ListFromIter

NamedFrom

NamedFromOwned

NumOpsDispatchCheckedchecked_arithmetic

PolarsArray

QuantileAggSeries

SlicedArray

Utility trait to slice concrete arrow arrays whilst keeping their concrete type. E.g. don’t return Box<dyn Array>.

VarAggSeries

Functions

call_categorical_merge_operationdtype-categorical

convert_inner_type

Cast null arrays to inner type and ensure that all offsets remain correct

indexes_to_usizes

make_categoricals_compatibledtype-categorical

make_list_categoricals_compatibledtype-categorical

Type Aliases

ArrayRef

DynArgs

IdxArr

IdxSize

LargeBinaryArray

LargeListArray

LargeStringArray

NullableChunkId

PolarsResult