polars_core

Module prelude

Everything you need to get started with Polars.

Re-exports

pub use crate::chunked_array::ChunkedArray;

pub use crate::chunked_array::arithmetic::ArithmeticChunked;

pub use crate::chunked_array::builder::BinaryChunkedBuilder;

pub use crate::chunked_array::builder::BooleanChunkedBuilder;

pub use crate::chunked_array::builder::ChunkedBuilder;

pub use crate::chunked_array::builder::ListBinaryChunkedBuilder;

pub use crate::chunked_array::builder::ListBooleanChunkedBuilder;

pub use crate::chunked_array::builder::ListBuilderTrait;

pub use crate::chunked_array::builder::ListPrimitiveChunkedBuilder;

pub use crate::chunked_array::builder::ListStringChunkedBuilder;

pub use crate::chunked_array::builder::NewChunkedArray;

pub use crate::chunked_array::builder::PrimitiveChunkedBuilder;

pub use crate::chunked_array::builder::StringChunkedBuilder;

pub use crate::chunked_array::collect::ChunkedCollectInferIterExt;

pub use crate::chunked_array::collect::ChunkedCollectIterExt;

pub use crate::chunked_array::iterator::PolarsIterator;

pub use crate::chunked_array::object::PolarsObject;object

pub use crate::frame::column::Column;

pub use crate::frame::column::IntoColumn;

pub use crate::frame::explode::UnpivotArgsIR;

pub use crate::frame::DataFrame;

pub use crate::frame::UniqueKeepStrategy;

pub use crate::hashing::VecHash;

pub use crate::scalar::Scalar;

pub use crate::series::arithmetic::LhsNumOps;

pub use crate::series::arithmetic::NumOpsDispatch;

pub use crate::series::IntoSeries;

pub use crate::series::Series;

pub use crate::series::SeriesTrait;

pub use crate::utils::IntoVec;

pub use crate::datatypes;

pub use crate::chunked_array::ops::*;

pub use crate::chunked_array::temporal::conversion::*;temporal

pub use crate::datatypes::*;

pub use crate::frame::group_by::*;algorithm_group_by

pub use crate::schema::*;

Modules

default_arrays

fixed_size_list

list

null

slice

utf8

Macros

df

polars_bail

polars_ensure

polars_err

polars_warn

with_match_categorical_physical_type

Structs

Arc

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

ArrowField

Represents Arrow’s metadata of a “column”.

ChunkId

NullableIdxSize

PlSmallStr

String type that inlines small strings.

RollingOptionsFixedWindowrolling_window

RollingVarParams

Enums

Ambiguous

IndexOrderndarray

NonExistent

PolarsError

QuantileMethod

RollingFnParams

Traits

ArrayCollectIterExt

CategoricalPhysicalDtypeExtdtype-categorical

ChunkAggSeries

Aggregations that return Series of unit length. Those can be used in broadcasting operations.

FromData

FromDataBinary

FromDataUtf8

IndexToUsize

InitHashMaps

ListFromIter

NamedFrom

NamedFromOwned

NumOpsDispatchCheckedchecked_arithmetic

QuantileAggSeries

SeedableFromU64SeedExt

SlicedArray

Utility trait to slice concrete arrow arrays whilst keeping their concrete type. E.g. don’t return Box<dyn Array>.

VarAggSeries

Functions

convert_inner_type

Cast null arrays to inner type and ensure that all offsets remain correct

indexes_to_usizes

try_raise_keyboard_interrupt

Checks if the keyboard interrupt flag is set, and if yes panics as a keyboard interrupt. This function is very cheap.

Type Aliases

ArrayRef

ArrowSchema

An ordered sequence of Fields

Categorical8Chunkeddtype-categorical

Categorical16Chunkeddtype-categorical

Categorical32Chunkeddtype-categorical

CategoricalChunkeddtype-categorical

IdxArr

IdxSize

LargeBinaryArray

LargeListArray

LargeStringArray

PlFixedStateQuality

PlHashMap

PlHashSet

PlIndexMap

PlIndexSet

PlRandomState

PlRandomStateQuality

PlSeedableRandomStateQuality

PolarsResult