Struct polars_core::series::Series

pub struct Series(pub Arc<dyn SeriesTrait>);

Series

The columnar data type for a DataFrame.

Most of the available functions are defined in the SeriesTrait trait.

The Series struct consists of typed ChunkedArray’s. To quickly cast a Series to a ChunkedArray you can call the method with the name of the type:

let s: Series = [1, 2, 3].iter().collect();
// Quickly obtain the ChunkedArray wrapped by the Series.
let chunked_array = s.i32().unwrap();

Arithmetic

You can do standard arithmetic on series.

let s = Series::new("a", [1 , 2, 3]);
let out_add = &s + &s;
let out_sub = &s - &s;
let out_div = &s / &s;
let out_mul = &s * &s;

Or with series and numbers.

let s: Series = (1..3).collect();
let out_add_one = &s + 1;
let out_multiply = &s * 10;

// Could not overload left hand side operator.
let out_divide = 1.div(&s);
let out_add = 1.add(&s);
let out_subtract = 1.sub(&s);
let out_multiply = 1.mul(&s);

Comparison

You can obtain boolean mask by comparing series.

let s = Series::new("dollars", &[1, 2, 3]);
let mask = s.equal(1).unwrap();
let valid = [true, false, false].iter();
assert!(mask
    .into_iter()
    .map(|opt_bool| opt_bool.unwrap()) // option, because series can be null
    .zip(valid)
    .all(|(a, b)| a == *b))

See all the comparison operators in the CmpOps trait

Iterators

The Series variants contain differently typed ChunkedArray’s. These structs can be turned into iterators, making it possible to use any function/ closure you want on a Series.

These iterators return an Option<T> because the values of a series may be null.

use polars_core::prelude::*;
let pi = 3.14;
let s = Series::new("angle", [2f32 * pi, pi, 1.5 * pi].as_ref());
let s_cos: Series = s.f32()
                    .expect("series was not an f32 dtype")
                    .into_iter()
                    .map(|opt_angle| opt_angle.map(|angle| angle.cos()))
                    .collect();

Creation

Series can be create from different data structures. Below we’ll show a few ways we can create a Series object.

// Series can be created from Vec's, slices and arrays
Series::new("boolean series", &[true, false, true]);
Series::new("int series", &[1, 2, 3]);
// And can be nullable
Series::new("got nulls", &[Some(1), None, Some(2)]);

// Series can also be collected from iterators
let from_iter: Series = (0..10)
    .into_iter()
    .collect();

Tuple Fields
0: Arc<dyn SeriesTrait>
Implementations

impl Series

pub fn fill_null(&self, strategy: FillNullStrategy) -> PolarsResult<Series>

Replace None values with one of the following strategies:

Forward fill (replace None with the previous value)

Backward fill (replace None with the next value)

Mean fill (replace None with the mean of the whole array)

Min fill (replace None with the minimum of the whole array)

Max fill (replace None with the maximum of the whole array)

Zero fill (replace None with the value zero)

One fill (replace None with the value one)

MinBound fill (replace with the minimum of that data type)

MaxBound fill (replace with the maximum of that data type)

NOTE: If you want to fill the Nones with a value use the fill_null operation on ChunkedArray<T>.

Example

fn example() -> PolarsResult<()> {
    let s = Series::new("some_missing", &[Some(1), None, Some(2)]);

    let filled = s.fill_null(FillNullStrategy::Forward(None))?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(1), Some(2)]);

    let filled = s.fill_null(FillNullStrategy::Backward(None))?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(2), Some(2)]);

    let filled = s.fill_null(FillNullStrategy::Min)?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(1), Some(2)]);

    let filled = s.fill_null(FillNullStrategy::Max)?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(2), Some(2)]);

    let filled = s.fill_null(FillNullStrategy::Mean)?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(1), Some(2)]);

    let filled = s.fill_null(FillNullStrategy::Zero)?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(0), Some(2)]);

    let filled = s.fill_null(FillNullStrategy::One)?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(1), Some(2)]);

    let filled = s.fill_null(FillNullStrategy::MinBound)?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(-2147483648), Some(2)]);

    let filled = s.fill_null(FillNullStrategy::MaxBound)?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(2147483647), Some(2)]);

    Ok(())
}
example();

impl Series

pub fn sample_n( &self, n: usize, with_replacement: bool, shuffle: bool, seed: Option<u64> ) -> PolarsResult<Self>

Available on crate feature random only.

pub fn sample_frac( &self, frac: f64, with_replacement: bool, shuffle: bool, seed: Option<u64> ) -> PolarsResult<Self>

Available on crate feature random only.

Sample a fraction between 0.0-1.0 of this ChunkedArray.

pub fn shuffle(&self, seed: Option<u64>) -> Self

Available on crate feature random only.

impl Series

pub fn fmt_list(&self) -> String

impl Series

pub fn from_any_values( name: &str, values: &[AnyValue<'_>], strict: bool ) -> PolarsResult<Self>

Construct a new Series from a slice of AnyValues.

The data type of the resulting Series is determined by the values and the strict parameter:

If strict is true, the data type is equal to the data type of the first non-null value. If any other non-null values do not match this data type, an error is raised.

If strict is false, the data type is the supertype of the values. An error is returned if no supertype can be determined. WARNING: A full pass over the values is required to determine the supertype.

If no values were passed, the resulting data type is Null.

pub fn from_any_values_and_dtype( name: &str, values: &[AnyValue<'_>], dtype: &DataType, strict: bool ) -> PolarsResult<Self>

Construct a new Series with the given dtype from a slice of AnyValues.

If strict is true, an error is returned if the values do not match the given data type. If strict is false, values that do not match the given data type are cast. If casting is not possible, the values are set to null instead.

impl Series

pub fn try_sub(&self, rhs: &Self) -> PolarsResult<Self>

pub fn try_add(&self, rhs: &Self) -> PolarsResult<Self>

pub fn try_mul(&self, rhs: &Self) -> PolarsResult<Self>

pub fn try_div(&self, rhs: &Self) -> PolarsResult<Self>

pub fn try_rem(&self, rhs: &Self) -> PolarsResult<Self>

impl Series

pub fn wrapping_trunc_div_scalar<T: Num + NumCast>(&self, rhs: T) -> Self

impl Series

pub fn try_add_owned(self, other: Self) -> PolarsResult<Self>

pub fn try_sub_owned(self, other: Self) -> PolarsResult<Self>

pub fn try_mul_owned(self, other: Self) -> PolarsResult<Self>

impl Series

pub unsafe fn from_chunks_and_dtype_unchecked( name: &str, chunks: Vec<ArrayRef>, dtype: &DataType ) -> Self

Takes chunks and a polars datatype and constructs the Series This is faster than creating from chunks and an arrow datatype because there is no casting involved

Safety

The caller must ensure that the given dtype’s physical type matches all the ArrayRef dtypes.

pub unsafe fn _try_from_arrow_unchecked( name: &str, chunks: Vec<ArrayRef>, dtype: &ArrowDataType ) -> PolarsResult<Self>

Safety

The caller must ensure that the given dtype matches all the ArrayRef dtypes.

pub unsafe fn _try_from_arrow_unchecked_with_md( name: &str, chunks: Vec<ArrayRef>, dtype: &ArrowDataType, md: Option<&Metadata> ) -> PolarsResult<Self>

Create a new Series without checking if the inner dtype of the chunks is correct

Safety

The caller must ensure that the given dtype matches all the ArrayRef dtypes.

impl Series

pub fn new_null(name: &str, len: usize) -> Series

impl Series

pub fn array_ref(&self, chunk_idx: usize) -> &ArrayRef

Returns a reference to the Arrow ArrayRef

pub fn to_arrow(&self, chunk_idx: usize, pl_flavor: bool) -> ArrayRef

Convert a chunk in the Series to the correct Arrow type. This conversion is needed because polars doesn’t use a 1 on 1 mapping for logical/ categoricals, etc.

impl Series

pub fn iter(&self) -> SeriesIter<'_>

iterate over Series as AnyValue.

Panics

This will panic if the array is not rechunked first.

pub fn phys_iter(&self) -> SeriesPhysIter<'_>

impl Series

pub fn i8(&self) -> PolarsResult<&Int8Chunked>

Unpack to ChunkedArray of dtype [DataType::Int8]

pub fn i16(&self) -> PolarsResult<&Int16Chunked>

Unpack to ChunkedArray of dtype [DataType::Int16]

pub fn i32(&self) -> PolarsResult<&Int32Chunked>

Unpack to ChunkedArray

let s = Series::new("foo", [1i32 ,2, 3]);
let s_squared: Series = s.i32()
    .unwrap()
    .into_iter()
    .map(|opt_v| {
        match opt_v {
            Some(v) => Some(v * v),
            None => None, // null value
        }
}).collect();

Unpack to ChunkedArray of dtype [DataType::Int32]

pub fn i64(&self) -> PolarsResult<&Int64Chunked>

Unpack to ChunkedArray of dtype [DataType::Int64]

pub fn f32(&self) -> PolarsResult<&Float32Chunked>

Unpack to ChunkedArray of dtype [DataType::Float32]

pub fn f64(&self) -> PolarsResult<&Float64Chunked>

Unpack to ChunkedArray of dtype [DataType::Float64]

pub fn u8(&self) -> PolarsResult<&UInt8Chunked>

Unpack to ChunkedArray of dtype [DataType::UInt8]

pub fn u16(&self) -> PolarsResult<&UInt16Chunked>

Unpack to ChunkedArray of dtype [DataType::UInt16]

pub fn u32(&self) -> PolarsResult<&UInt32Chunked>

Unpack to ChunkedArray of dtype [DataType::UInt32]

pub fn u64(&self) -> PolarsResult<&UInt64Chunked>

Unpack to ChunkedArray of dtype [DataType::UInt64]

pub fn bool(&self) -> PolarsResult<&BooleanChunked>

Unpack to ChunkedArray of dtype [DataType::Boolean]

pub fn str(&self) -> PolarsResult<&StringChunked>

Unpack to ChunkedArray of dtype [DataType::String]

pub fn binary(&self) -> PolarsResult<&BinaryChunked>

Unpack to ChunkedArray of dtype [DataType::Binary]

pub fn binary_offset(&self) -> PolarsResult<&BinaryOffsetChunked>

Unpack to ChunkedArray of dtype [DataType::Binary]

pub fn decimal(&self) -> PolarsResult<&DecimalChunked>

Available on crate feature dtype-decimal only.

Unpack to ChunkedArray of dtype [DataType::Decimal]

pub fn list(&self) -> PolarsResult<&ListChunked>

Unpack to ChunkedArray of dtype list

pub fn categorical(&self) -> PolarsResult<&CategoricalChunked>

Available on crate feature dtype-categorical only.

Unpack to ChunkedArray of dtype [DataType::Categorical]

pub fn null(&self) -> PolarsResult<&NullChunked>

Unpack to ChunkedArray of dtype [DataType::Null]

impl Series

pub fn extend_constant( &self, value: AnyValue<'_>, n: usize ) -> PolarsResult<Self>

Extend with a constant value.

impl Series

pub fn full_null(name: &str, size: usize, dtype: &DataType) -> Self

impl Series

pub fn get_leaf_array(&self) -> Series

Recurse nested types until we are at the leaf array.

pub fn implode(&self) -> PolarsResult<ListChunked>

Convert the values of this Series to a ListChunked with a length of 1, so a Series of [1, 2, 3] becomes [[1, 2, 3]].

pub fn reshape_list(&self, dimensions: &[i64]) -> PolarsResult<Series>

impl Series

pub fn new_empty(name: &str, dtype: &DataType) -> Series

Create a new empty Series.

pub fn clear(&self) -> Series

pub unsafe fn chunks_mut(&mut self) -> &mut Vec<ArrayRef>

Safety

The caller must ensure the length and the data types of ArrayRef does not change. And that the null_count is updated (e.g. with a compute_len())

pub fn select_chunk(&self, i: usize) -> Self

pub fn is_sorted_flag(&self) -> IsSorted

pub fn set_sorted_flag(&mut self, sorted: IsSorted)

pub fn get_flags(&self) -> MetadataFlags

pub fn into_frame(self) -> DataFrame

pub fn rename(&mut self, name: &str) -> &mut Series

Rename series.

pub fn with_name(self, name: &str) -> Series

Return this Series with a new name.

pub fn try_set_metadata<T: PolarsDataType + 'static>( &mut self, metadata: Metadata<T> ) -> bool

Try to set the Metadata for the underlying ChunkedArray

This does not guarantee that the Metadata is always set. It returns whether it was successful.

pub fn from_arrow_chunks( name: &str, arrays: Vec<ArrayRef> ) -> PolarsResult<Series>

pub fn from_arrow(name: &str, array: ArrayRef) -> PolarsResult<Series>

pub fn shrink_to_fit(&mut self)

Shrink the capacity of this array to fit its length.

pub fn append(&mut self, other: &Series) -> PolarsResult<&mut Self>

Append in place. This is done by adding the chunks of other to this Series.

See ChunkedArray::append and ChunkedArray::extend.

pub fn compute_len(&mut self)

Redo a length and null_count compute

pub fn extend(&mut self, other: &Series) -> PolarsResult<&mut Self>

Extend the memory backed by this array with the values from other.

See ChunkedArray::extend and ChunkedArray::append.

pub fn sort(&self, sort_options: SortOptions) -> PolarsResult<Self>

Sort the series with specific options.

Example

let s = Series::new("foo", [2, 1, 3]);
let sorted = s.sort(SortOptions::default())?;
assert_eq!(sorted, Series::new("foo", [1, 2, 3]));
}

See SortOptions for more options.

pub fn as_single_ptr(&mut self) -> PolarsResult<usize>

Only implemented for numeric types

pub fn cast(&self, dtype: &DataType) -> PolarsResult<Self>

pub fn cast_with_options( &self, dtype: &DataType, options: CastOptions ) -> PolarsResult<Self>

Cast [Series] to another [DataType].

pub unsafe fn cast_unchecked(&self, dtype: &DataType) -> PolarsResult<Self>

Cast from physical to logical types without any checks on the validity of the cast.

Safety

This can lead to invalid memory access in downstream code.

pub fn to_float(&self) -> PolarsResult<Series>

Cast numerical types to f64, and keep floats as is.

pub fn sum<T>(&self) -> PolarsResult<T>
where T: NumCast,

Compute the sum of all values in this Series. Returns Some(0) if the array is empty, and None if the array only contains null values.

If the DataType is one of {Int8, UInt8, Int16, UInt16} the Series is first cast to Int64 to prevent overflow issues.

pub fn min<T>(&self) -> PolarsResult<Option<T>>
where T: NumCast,

Returns the minimum value in the array, according to the natural order. Returns an option because the array is nullable.

pub fn max<T>(&self) -> PolarsResult<Option<T>>
where T: NumCast,

Returns the maximum value in the array, according to the natural order. Returns an option because the array is nullable.

pub fn explode(&self) -> PolarsResult<Series>

Explode a list Series. This expands every item to a new row..

pub fn is_nan(&self) -> PolarsResult<BooleanChunked>

Check if float value is NaN (note this is different than missing/ null)

pub fn is_not_nan(&self) -> PolarsResult<BooleanChunked>

Check if float value is NaN (note this is different than missing/ null)

pub fn is_finite(&self) -> PolarsResult<BooleanChunked>

Check if numeric value is finite

pub fn is_infinite(&self) -> PolarsResult<BooleanChunked>

Check if float value is infinite

pub fn zip_with( &self, mask: &BooleanChunked, other: &Series ) -> PolarsResult<Series>

Available on crate feature zip_with only.

Create a new ChunkedArray with values from self where the mask evaluates true and values from other where the mask evaluates false. This function automatically broadcasts unit length inputs.

pub fn to_physical_repr(&self) -> Cow<'_, Series>

Cast a datelike Series to their physical representation. Primitives remain unchanged

Date -> Int32

Datetime-> Int64

Time -> Int64

Categorical -> UInt32

List(inner) -> List(physical of inner)

pub unsafe fn take_unchecked_from_slice(&self, idx: &[IdxSize]) -> Series

Take by index if ChunkedArray contains a single chunk.

Safety

This doesn’t check any bounds. Null validity is checked.

pub fn gather_every(&self, n: usize, offset: usize) -> Series

Traverse and collect every nth element in a new array.

pub fn dot(&self, other: &Series) -> PolarsResult<f64>

Available on crate feature dot_product only.

pub fn sum_reduce(&self) -> PolarsResult<Scalar>

Get the sum of the Series as a new Series of length 1. Returns a Series with a single zeroed entry if self is an empty numeric series.

If the DataType is one of {Int8, UInt8, Int16, UInt16} the Series is first cast to Int64 to prevent overflow issues.

pub fn product(&self) -> PolarsResult<Scalar>

Get the product of an array.

If the DataType is one of {Int8, UInt8, Int16, UInt16} the Series is first cast to Int64 to prevent overflow issues.

pub fn strict_cast(&self, dtype: &DataType) -> PolarsResult<Series>

Cast throws an error if conversion had overflows

pub fn str_value(&self, index: usize) -> PolarsResult<Cow<'_, str>>

pub fn head(&self, length: Option<usize>) -> Series

Get the head of the Series.

pub fn tail(&self, length: Option<usize>) -> Series

Get the tail of the Series.

pub fn mean_reduce(&self) -> Scalar

pub fn unique_stable(&self) -> PolarsResult<Series>

Compute the unique elements, but maintain order. This requires more work than a naive Series::unique.

pub fn idx(&self) -> PolarsResult<&IdxCa>

pub fn estimated_size(&self) -> usize

Returns an estimation of the total (heap) allocated size of the Series in bytes.

Implementation

This estimation is the sum of the size of its buffers, validity, including nested arrays. Multiple arrays may share buffers and bitmaps. Therefore, the size of 2 arrays is not the sum of the sizes computed from this function. In particular, [StructArray]’s size is an upper bound.

When an array is sliced, its allocated size remains constant because the buffer unchanged. However, this function will yield a smaller number. This is because this function returns the visible size of the buffer, not its total capacity.

FFI buffers are included in this estimation.

pub fn as_list(&self) -> ListChunked

Packs every element into a list.

impl Series

pub fn equals(&self, other: &Series) -> bool

Check if series are equal. Note that None == None evaluates to false

pub fn equals_missing(&self, other: &Series) -> bool

Check if all values in series are equal where None == None evaluates to true.

pub fn get_data_ptr(&self) -> usize

Get a pointer to the underlying data of this Series. Can be useful for fast comparisons.