Struct polars_core::series::Series

pub struct Series(pub Arc<dyn SeriesTrait>);

Series

The columnar data type for a DataFrame.

Most of the available functions are defined in the SeriesTrait trait.

The Series struct consists of typed ChunkedArray’s. To quickly cast a Series to a ChunkedArray you can call the method with the name of the type:

let s: Series = [1, 2, 3].iter().collect();
// Quickly obtain the ChunkedArray wrapped by the Series.
let chunked_array = s.i32().unwrap();

Arithmetic

You can do standard arithmetic on series.

let s = Series::new("a", [1 , 2, 3]);
let out_add = &s + &s;
let out_sub = &s - &s;
let out_div = &s / &s;
let out_mul = &s * &s;

Or with series and numbers.

let s: Series = (1..3).collect();
let out_add_one = &s + 1;
let out_multiply = &s * 10;

// Could not overload left hand side operator.
let out_divide = 1.div(&s);
let out_add = 1.add(&s);
let out_subtract = 1.sub(&s);
let out_multiply = 1.mul(&s);

Comparison

You can obtain boolean mask by comparing series.

let s = Series::new("dollars", &[1, 2, 3]);
let mask = s.equal(1);
let valid = [true, false, false].iter();
assert!(mask
    .into_iter()
    .map(|opt_bool| opt_bool.unwrap()) // option, because series can be null
    .zip(valid)
    .all(|(a, b)| a == *b))

See all the comparison operators in the CmpOps trait

Iterators

The Series variants contain differently typed ChunkedArray’s. These structs can be turned into iterators, making it possible to use any function/ closure you want on a Series.

These iterators return an Option<T> because the values of a series may be null.

use polars_core::prelude::*;
let pi = 3.14;
let s = Series::new("angle", [2f32 * pi, pi, 1.5 * pi].as_ref());
let s_cos: Series = s.f32()
                    .expect("series was not an f32 dtype")
                    .into_iter()
                    .map(|opt_angle| opt_angle.map(|angle| angle.cos()))
                    .collect();

Creation

Series can be create from different data structures. Below we’ll show a few ways we can create a Series object.

// Series can be created from Vec's, slices and arrays
Series::new("boolean series", &[true, false, true]);
Series::new("int series", &[1, 2, 3]);
// And can be nullable
Series::new("got nulls", &[Some(1), None, Some(2)]);

// Series can also be collected from iterators
let from_iter: Series = (0..10)
    .into_iter()
    .collect();

Tuple Fields
0: Arc<dyn SeriesTrait>
Implementations

impl Series

pub fn sample_n(
&self,
n: usize,
with_replacement: bool,
seed: u64
) -> Result<Self>

pub fn sample_frac(
&self,
frac: f64,
with_replacement: bool,
seed: u64
) -> Result<Self>

Sample a fraction between 0.0-1.0 of this ChunkedArray.

pub fn shuffle(&self, seed: u64) -> Self

impl Series

pub fn agg_valid_count(&self, groups: &GroupsProxy) -> Option<Series>

pub fn agg_first(&self, groups: &GroupsProxy) -> Series

pub fn agg_n_unique(&self, groups: &GroupsProxy) -> Option<Series>

pub fn agg_last(&self, groups: &GroupsProxy) -> Series

impl Series

pub fn to_arrow(&self, chunk_idx: usize) -> ArrayRef

Convert a chunk in the Series to the correct Arrow type. This conversion is needed because polars doesn’t use a 1 on 1 mapping for logical/ categoricals, etc.

impl Series

pub fn diff(&self, n: usize, null_behavior: NullBehavior) -> Series

impl Series

pub fn extend_constant(&self, value: AnyValue<'_>, n: usize) -> Result<Self>

Extend with a constant value.

impl Series

pub fn full_null(name: &str, size: usize, dtype: &DataType) -> Self

impl Series

pub fn round(&self, decimals: u32) -> Result<Self>

Round underlying floating point array to given decimal.

pub fn floor(&self) -> Result<Self>

Floor underlying floating point array to the lowest integers smaller or equal to the float value.

pub fn ceil(&self) -> Result<Self>

Ceil underlying floating point array to the heighest integers smaller or equal to the float value.

pub fn clip(&self, min: f64, max: f64) -> Result<Self>

Ceil underlying floating point array to the heighest integers smaller or equal to the float value.

impl Series

pub fn to_list(&self) -> Result<ListChunked>

Convert the values of this Series to a ListChunked with a length of 1, So a Series of: [1, 2, 3] becomes [[1, 2, 3]]

pub fn reshape(&self, dims: &[i64]) -> Result<Series>

impl Series

pub fn into_frame(self) -> DataFrame

pub fn rename(&mut self, name: &str) -> &mut Series

Rename series.

pub fn shrink_to_fit(&mut self)

Shrink the capacity of this array to fit it’s length.

pub fn append_array(&mut self, other: ArrayRef) -> Result<&mut Self>

Append arrow array of same datatype.

pub fn append(&mut self, other: &Series) -> Result<&mut Self>

Append in place. This is done by adding the chunks of other to this Series.

See ChunkedArray::append and ChunkedArray::extend.

pub fn extend(&mut self, other: &Series) -> Result<&mut Self>

Extend the memory backed by this array with the values from other.

See ChunkedArray::extend and ChunkedArray::append.

pub fn sort(&self, reverse: bool) -> Self

pub fn as_single_ptr(&mut self) -> Result<usize>

Only implemented for numeric types

pub fn cast(&self, dtype: &DataType) -> Result<Self>

Cast [Series] to another [DataType]

pub fn sum<T>(&self) -> Option<T> where
T: NumCast,

Compute the sum of all values in this Series. Returns None if the array is empty or only contains null values.

If the DataType is one of {Int8, UInt8, Int16, UInt16} the Series is first cast to Int64 to prevent overflow issues.

let s = Series::new("days", &[1, 2, 3]);
assert_eq!(s.sum(), Some(6));

pub fn min<T>(&self) -> Option<T> where
T: NumCast,

Returns the minimum value in the array, according to the natural order. Returns an option because the array is nullable.

let s = Series::new("days", [1, 2, 3].as_ref());
assert_eq!(s.min(), Some(1));

pub fn max<T>(&self) -> Option<T> where
T: NumCast,

Returns the maximum value in the array, according to the natural order. Returns an option because the array is nullable.

let s = Series::new("days", [1, 2, 3].as_ref());
assert_eq!(s.max(), Some(3));

pub fn explode(&self) -> Result<Series>

Explode a list or utf8 Series. This expands every item to a new row..

pub fn is_nan(&self) -> Result<BooleanChunked>

Check if float value is NaN (note this is different than missing/ null)

pub fn is_not_nan(&self) -> Result<BooleanChunked>

Check if float value is NaN (note this is different than missing/ null)

pub fn is_finite(&self) -> Result<BooleanChunked>

Check if float value is finite

pub fn is_infinite(&self) -> Result<BooleanChunked>

Check if float value is finite

pub fn zip_with(&self, mask: &BooleanChunked, other: &Series) -> Result<Series>

Create a new ChunkedArray with values from self where the mask evaluates true and values from other where the mask evaluates false

pub fn to_physical_repr(&self) -> Cow<'_, Series>

Cast a datelike Series to their physical representation. Primitives remain unchanged

Date -> Int32

Datetime-> Int64

Time -> Int64

Categorical -> UInt32

pub unsafe fn take_unchecked_threaded(
&self,
idx: &UInt32Chunked,
rechunk: bool
) -> Result<Series>

Take by index if ChunkedArray contains a single chunk.

Safety

This doesn’t check any bounds. Null validity is checked.

pub fn take_threaded(
&self,
idx: &UInt32Chunked,
rechunk: bool
) -> Result<Series>

Take by index. This operation is clone.

Safety

Out of bounds access doesn’t Error but will return a Null value

pub fn filter_threaded(
&self,
filter: &BooleanChunked,
rechunk: bool
) -> Result<Series>

Filter by boolean mask. This operation clones data.

pub fn sum_as_series(&self) -> Series

Get the sum of the Series as a new Series of length 1.

If the DataType is one of {Int8, UInt8, Int16, UInt16} the Series is first cast to Int64 to prevent overflow issues.

pub fn cummax(&self, _reverse: bool) -> Series

Get an array with the cumulative max computed at every element

pub fn cummin(&self, _reverse: bool) -> Series

Get an array with the cumulative min computed at every element

pub fn cumsum(&self, reverse: bool) -> Series

Get an array with the cumulative sum computed at every element

If the DataType is one of {Int8, UInt8, Int16, UInt16} the Series is first cast to Int64 to prevent overflow issues.

pub fn cumprod(&self, reverse: bool) -> Series

Get an array with the cumulative product computed at every element

If the DataType is one of {Int8, UInt8, Int16, UInt16, Int32, UInt32} the Series is first cast to Int64 to prevent overflow issues.

pub fn product(&self) -> Series

Get the product of an array.

If the DataType is one of {Int8, UInt8, Int16, UInt16} the Series is first cast to Int64 to prevent overflow issues.

pub fn rolling_var(&self, _options: RollingOptions) -> Result<Series>

Apply a rolling variance to a Series. See:

pub fn rolling_std(&self, _options: RollingOptions) -> Result<Series>

Apply a rolling std to a Series. See:

pub fn rolling_mean(&self, _options: RollingOptions) -> Result<Series>

Apply a rolling mean to a Series. See: ChunkedArray::rolling_mean

pub fn rolling_sum(&self, _options: RollingOptions) -> Result<Series>

Apply a rolling sum to a Series. See: ChunkedArray::rolling_sum

pub fn rolling_median(&self, _options: RollingOptions) -> Result<Series>

Apply a rolling median to a Series. See: ChunkedArray::rolling_median.

pub fn rolling_quantile(
&self,
_quantile: f64,
_interpolation: QuantileInterpolOptions,
_options: RollingOptions
) -> Result<Series>

Apply a rolling quantile to a Series. See: ChunkedArray::rolling_quantile.

pub fn rolling_min(&self, _options: RollingOptions) -> Result<Series>

Apply a rolling min to a Series. See: ChunkedArray::rolling_min

pub fn rolling_max(&self, _options: RollingOptions) -> Result<Series>

Apply a rolling max to a Series. See: ChunkedArray::rolling_max

pub fn rank(&self, options: RankOptions) -> Series

pub fn strict_cast(&self, data_type: &DataType) -> Result<Series>

Cast throws an error if conversion had overflows

pub fn is_logical(&self) -> bool

Check if the underlying data is a logical type.

pub fn is_numeric_physical(&self) -> bool

Check if underlying physical data is numeric.

Date types and Categoricals are also considered numeric.

pub fn is_numeric(&self) -> bool

Check if underlying data is numeric

pub fn abs(&self) -> Result<Series>

convert numerical values to their absolute value

pub fn str_value(&self, index: usize) -> Cow<'_, str>

pub fn head(&self, length: Option<usize>) -> Series

Get the head of the Series.

pub fn tail(&self, length: Option<usize>) -> Series

Get the tail of the Series.

pub fn mean_as_series(&self) -> Series

impl Series

pub fn series_equal(&self, other: &Series) -> bool

Check if series are equal. Note that None == None evaluates to false

pub fn series_equal_missing(&self, other: &Series) -> bool

Check if all values in series are equal where None == None evaluates to true.

pub fn get_data_ptr(&self) -> usize

Get a pointer to the underlying data of this Series. Can be useful for fast comparisons.