Trait polars_core::series::SeriesTrait

pub trait SeriesTrait: Send + Sync + PrivateSeries {
    fn rename(&mut self, name: &str);

    fn take_every(&self, n: usize) -> Series;

    fn sample_n(&self, n: usize, with_replacement: bool) -> Result<Series>;

    fn sample_frac(&self, frac: f64, with_replacement: bool) -> Result<Series>;

    fn cum_max(&self, _reverse: bool) -> Series { ... }

    fn cum_min(&self, _reverse: bool) -> Series { ... }

    fn cum_sum(&self, _reverse: bool) -> Series { ... }

    fn array_data(&self) -> Vec<ArrayDataRef> { ... }

    fn chunk_lengths(&self) -> &Vec<usize> { ... }

    fn name(&self) -> &str { ... }

    fn field(&self) -> &Field { ... }

    fn dtype(&self) -> &DataType { ... }

    fn chunks(&self) -> &Vec<ArrayRef> { ... }

    fn n_chunks(&self) -> usize { ... }

    fn i8(&self) -> Result<&Int8Chunked> { ... }

    fn i16(&self) -> Result<&Int16Chunked> { ... }

    fn i32(&self) -> Result<&Int32Chunked> { ... }

    fn i64(&self) -> Result<&Int64Chunked> { ... }

    fn f32(&self) -> Result<&Float32Chunked> { ... }

    fn f64(&self) -> Result<&Float64Chunked> { ... }

    fn u8(&self) -> Result<&UInt8Chunked> { ... }

    fn u16(&self) -> Result<&UInt16Chunked> { ... }

    fn u32(&self) -> Result<&UInt32Chunked> { ... }

    fn u64(&self) -> Result<&UInt64Chunked> { ... }

    fn bool(&self) -> Result<&BooleanChunked> { ... }

    fn utf8(&self) -> Result<&Utf8Chunked> { ... }

    fn date32(&self) -> Result<&Date32Chunked> { ... }

    fn date64(&self) -> Result<&Date64Chunked> { ... }

    fn time64_nanosecond(&self) -> Result<&Time64NanosecondChunked> { ... }

    fn duration_nanosecond(&self) -> Result<&DurationNanosecondChunked> { ... }

    fn duration_millisecond(&self) -> Result<&DurationMillisecondChunked> { ... }

    fn list(&self) -> Result<&ListChunked> { ... }

    fn categorical(&self) -> Result<&CategoricalChunked> { ... }

    fn append_array(&mut self, _other: ArrayRef) -> Result<()> { ... }

    fn limit(&self, num_elements: usize) -> Result<Series> { ... }

    fn slice(&self, _offset: usize, _length: usize) -> Result<Series> { ... }

    fn append(&mut self, _other: &Series) -> Result<()> { ... }

    fn filter(&self, _filter: &BooleanChunked) -> Result<Series> { ... }

    fn take_iter(&self, _iter: &mut dyn Iterator<Item = usize>) -> Series { ... }

    unsafe fn take_iter_unchecked(
        &self, 
        _iter: &mut dyn Iterator<Item = usize>
    ) -> Series { ... }

    unsafe fn take_unchecked(&self, _idx: &UInt32Chunked) -> Result<Series> { ... }

    unsafe fn take_opt_iter_unchecked(
        &self, 
        _iter: &mut dyn Iterator<Item = Option<usize>>
    ) -> Series { ... }

    fn take_opt_iter(
        &self, 
        _iter: &mut dyn Iterator<Item = Option<usize>>
    ) -> Series { ... }

    fn take(&self, _indices: &UInt32Chunked) -> Series { ... }

    fn len(&self) -> usize { ... }

    fn is_empty(&self) -> bool { ... }

    fn rechunk(&self) -> Series { ... }

    fn head(&self, _length: Option<usize>) -> Series { ... }

    fn tail(&self, _length: Option<usize>) -> Series { ... }

    fn drop_nulls(&self) -> Series { ... }

    fn expand_at_index(&self, _index: usize, _length: usize) -> Series { ... }

    fn cast_with_datatype(&self, _data_type: &DataType) -> Result<Series> { ... }

    fn to_dummies(&self) -> Result<DataFrame> { ... }

    fn value_counts(&self) -> Result<DataFrame> { ... }

    fn get(&self, _index: usize) -> AnyValue<'_> { ... }

    unsafe fn get_unchecked(&self, _index: usize) -> AnyValue<'_> { ... }

    fn sort_in_place(&mut self, _reverse: bool) { ... }

    fn sort(&self, _reverse: bool) -> Series { ... }

    fn argsort(&self, _reverse: bool) -> UInt32Chunked { ... }

    fn null_count(&self) -> usize { ... }

    fn unique(&self) -> Result<Series> { ... }

    fn n_unique(&self) -> Result<usize> { ... }

    fn arg_unique(&self) -> Result<Vec<u32>> { ... }

    fn arg_true(&self) -> Result<UInt32Chunked> { ... }

    fn is_null(&self) -> BooleanChunked { ... }

    fn is_not_null(&self) -> BooleanChunked { ... }

    fn is_unique(&self) -> Result<BooleanChunked> { ... }

    fn is_duplicated(&self) -> Result<BooleanChunked> { ... }

    fn null_bits(&self) -> Vec<(usize, Option<Buffer>)> { ... }

    fn reverse(&self) -> Series { ... }

    fn as_single_ptr(&mut self) -> Result<usize> { ... }

    fn shift(&self, _periods: i64) -> Series { ... }

    fn fill_none(&self, _strategy: FillNoneStrategy) -> Result<Series> { ... }

    fn zip_with(
        &self, 
        _mask: &BooleanChunked, 
        _other: &Series
    ) -> Result<Series> { ... }

    fn sum_as_series(&self) -> Series { ... }

    fn max_as_series(&self) -> Series { ... }

    fn min_as_series(&self) -> Series { ... }

    fn mean_as_series(&self) -> Series { ... }

    fn median_as_series(&self) -> Series { ... }

    fn var_as_series(&self) -> Series { ... }

    fn std_as_series(&self) -> Series { ... }

    fn quantile_as_series(&self, _quantile: f64) -> Result<Series> { ... }

    fn rolling_mean(
        &self, 
        _window_size: usize, 
        _weight: Option<&[f64]>, 
        _ignore_null: bool
    ) -> Result<Series> { ... }

    fn rolling_sum(
        &self, 
        _window_size: usize, 
        _weight: Option<&[f64]>, 
        _ignore_null: bool
    ) -> Result<Series> { ... }

    fn rolling_min(
        &self, 
        _window_size: usize, 
        _weight: Option<&[f64]>, 
        _ignore_null: bool
    ) -> Result<Series> { ... }

    fn rolling_max(
        &self, 
        _window_size: usize, 
        _weight: Option<&[f64]>, 
        _ignore_null: bool
    ) -> Result<Series> { ... }

    fn fmt_list(&self) -> String { ... }

    fn hour(&self) -> Result<Series> { ... }

    fn minute(&self) -> Result<Series> { ... }

    fn second(&self) -> Result<Series> { ... }

    fn nanosecond(&self) -> Result<Series> { ... }

    fn day(&self) -> Result<Series> { ... }

    fn ordinal_day(&self) -> Result<Series> { ... }

    fn month(&self) -> Result<Series> { ... }

    fn year(&self) -> Result<Series> { ... }

    fn datetime_str_fmt(&self, fmt: &str) -> Result<Series> { ... }

    fn clone_inner(&self) -> Arc<dyn SeriesTrait> { ... }

    fn get_as_any(&self, _index: usize) -> &dyn Any { ... }

    fn pow(&self, _exponent: f64) -> Result<Series> { ... }
}

Required methods

fn rename(&mut self, name: &str)

Rename the Series.

fn take_every(&self, n: usize) -> Series

Take every nth value as a new Series

fn sample_n(&self, n: usize, with_replacement: bool) -> Result<Series>

Sample n datapoints from this Series.

fn sample_frac(&self, frac: f64, with_replacement: bool) -> Result<Series>

Sample a fraction between 0.0-1.0 of this ChunkedArray.

Loading content...
Provided methods

fn cum_max(&self, _reverse: bool) -> Series

Get an array with the cumulative max computed at every element

fn cum_min(&self, _reverse: bool) -> Series

Get an array with the cumulative min computed at every element

fn cum_sum(&self, _reverse: bool) -> Series

Get an array with the cumulative sum computed at every element

fn array_data(&self) -> Vec<ArrayDataRef>

Get Arrow ArrayData

fn chunk_lengths(&self) -> &Vec<usize>

Get the lengths of the underlying chunks

fn name(&self) -> &str

Name of series.

fn field(&self) -> &Field

Get field (used in schema)

fn dtype(&self) -> &DataType

Get datatype of series.

fn chunks(&self) -> &Vec<ArrayRef>

Underlying chunks.

fn n_chunks(&self) -> usize

Number of chunks in this Series

fn i8(&self) -> Result<&Int8Chunked>

Unpack to ChunkedArray of dtype i8

fn i16(&self) -> Result<&Int16Chunked>

Unpack to ChunkedArray i16

fn i32(&self) -> Result<&Int32Chunked>

Unpack to ChunkedArray

let s: Series = [1, 2, 3].iter().collect();
let s_squared: Series = s.i32()
    .unwrap()
    .into_iter()
    .map(|opt_v| {
        match opt_v {
            Some(v) => Some(v * v),
            None => None, // null value
        }
}).collect();

fn i64(&self) -> Result<&Int64Chunked>

Unpack to ChunkedArray of dtype i64

fn f32(&self) -> Result<&Float32Chunked>

Unpack to ChunkedArray of dtype f32

fn f64(&self) -> Result<&Float64Chunked>

Unpack to ChunkedArray of dtype f64

fn u8(&self) -> Result<&UInt8Chunked>

Unpack to ChunkedArray of dtype u8

fn u16(&self) -> Result<&UInt16Chunked>

Unpack to ChunkedArray of dtype u16

fn u32(&self) -> Result<&UInt32Chunked>

Unpack to ChunkedArray of dtype u32

fn u64(&self) -> Result<&UInt64Chunked>

Unpack to ChunkedArray of dtype u64

fn bool(&self) -> Result<&BooleanChunked>

Unpack to ChunkedArray of dtype bool

fn utf8(&self) -> Result<&Utf8Chunked>

Unpack to ChunkedArray of dtype utf8

fn date32(&self) -> Result<&Date32Chunked>

Unpack to ChunkedArray of dtype date32

fn date64(&self) -> Result<&Date64Chunked>

Unpack to ChunkedArray of dtype date64

fn time64_nanosecond(&self) -> Result<&Time64NanosecondChunked>

Unpack to ChunkedArray of dtype time64_nanosecond

fn duration_nanosecond(&self) -> Result<&DurationNanosecondChunked>

Unpack to ChunkedArray of dtype duration_nanosecond

fn duration_millisecond(&self) -> Result<&DurationMillisecondChunked>

Unpack to ChunkedArray of dtype duration_millisecond

fn list(&self) -> Result<&ListChunked>

Unpack to ChunkedArray of dtype list

fn categorical(&self) -> Result<&CategoricalChunked>

Unpack to ChunkedArray of dtype categorical

fn append_array(&mut self, _other: ArrayRef) -> Result<()>

Append Arrow array of same dtype to this Series.

fn limit(&self, num_elements: usize) -> Result<Series>

Take num_elements from the top as a zero copy view.

fn slice(&self, _offset: usize, _length: usize) -> Result<Series>

Get a zero copy view of the data.

fn append(&mut self, _other: &Series) -> Result<()>

Append a Series of the same type in place.

fn filter(&self, _filter: &BooleanChunked) -> Result<Series>

Filter by boolean mask. This operation clones data.

fn take_iter(&self, _iter: &mut dyn Iterator<Item = usize>) -> Series

Take by index from an iterator. This operation clones the data.

Safety

Out of bounds access doesn't Error but will return a Null value

unsafe fn take_iter_unchecked(
&self,
_iter: &mut dyn Iterator<Item = usize>
) -> Series

Take by index from an iterator. This operation clones the data.

Safety

This doesn't check any bounds or null validity.

unsafe fn take_unchecked(&self, _idx: &UInt32Chunked) -> Result<Series>

Take by index if ChunkedArray contains a single chunk.

Safety

This doesn't check any bounds. Null validity is checked.

unsafe fn take_opt_iter_unchecked(
&self,
_iter: &mut dyn Iterator<Item = Option<usize>>
) -> Series

Take by index from an iterator. This operation clones the data.

Safety

This doesn't check any bounds or null validity.

fn take_opt_iter(
&self,
_iter: &mut dyn Iterator<Item = Option<usize>>
) -> Series

Take by index from an iterator. This operation clones the data.

Safety

Out of bounds access doesn't Error but will return a Null value

fn take(&self, _indices: &UInt32Chunked) -> Series

Take by index. This operation is clone.

Safety

Out of bounds access doesn't Error but will return a Null value

fn len(&self) -> usize

Get length of series.

fn is_empty(&self) -> bool

Check if Series is empty.

fn rechunk(&self) -> Series

Aggregate all chunks to a contiguous array of memory.

fn head(&self, _length: Option<usize>) -> Series

Get the head of the Series.

fn tail(&self, _length: Option<usize>) -> Series

Get the tail of the Series.

fn drop_nulls(&self) -> Series

Drop all null values and return a new Series.

fn expand_at_index(&self, _index: usize, _length: usize) -> Series

Create a new Series filled with values at that index.

Example

use polars_core::prelude::*;
let s = Series::new("a", [0i32, 1, 8]);
let expanded = s.expand_at_index(2, 4);
assert_eq!(Vec::from(expanded.i32().unwrap()), &[Some(8), Some(8), Some(8), Some(8)])

fn cast_with_datatype(&self, _data_type: &DataType) -> Result<Series>

fn to_dummies(&self) -> Result<DataFrame>

Create dummy variables. See DataFrame

fn value_counts(&self) -> Result<DataFrame>

fn get(&self, _index: usize) -> AnyValue<'_>

Get a single value by index. Don't use this operation for loops as a runtime cast is needed for every iteration.

unsafe fn get_unchecked(&self, _index: usize) -> AnyValue<'_>

Get a single value by index. Don't use this operation for loops as a runtime cast is needed for every iteration.

Safety

Does not do any bounds checking

fn sort_in_place(&mut self, _reverse: bool)

Sort in place.

fn sort(&self, _reverse: bool) -> Series

fn argsort(&self, _reverse: bool) -> UInt32Chunked

Retrieve the indexes needed for a sort.

fn null_count(&self) -> usize

Count the null values.

fn unique(&self) -> Result<Series>

Get unique values in the Series.

fn n_unique(&self) -> Result<usize>

Get unique values in the Series.

fn arg_unique(&self) -> Result<Vec<u32>>

Get first indexes of unique values.

fn arg_true(&self) -> Result<UInt32Chunked>

Get indexes that evaluate true

fn is_null(&self) -> BooleanChunked

Get a mask of the null values.

fn is_not_null(&self) -> BooleanChunked

Get a mask of the non-null values.

fn is_unique(&self) -> Result<BooleanChunked>

Get a mask of all the unique values.

fn is_duplicated(&self) -> Result<BooleanChunked>

Get a mask of all the duplicated values.

fn null_bits(&self) -> Vec<(usize, Option<Buffer>)>

Get the bits that represent the null values of the underlying ChunkedArray

fn reverse(&self) -> Series

return a Series in reversed order

fn as_single_ptr(&mut self) -> Result<usize>

Rechunk and return a pointer to the start of the Series. Only implemented for numeric types

fn shift(&self, _periods: i64) -> Series

Shift the values by a given period and fill the parts that will be empty due to this operation with Nones.

NOTE: If you want to fill the Nones with a value use the shift operation on ChunkedArray<T>.

Example

fn example() -> Result<()> {
    let s = Series::new("series", &[1, 2, 3]);

    let shifted = s.shift(1);
    assert_eq!(Vec::from(shifted.i32()?), &[None, Some(1), Some(2)]);

    let shifted = s.shift(-1);
    assert_eq!(Vec::from(shifted.i32()?), &[Some(2), Some(3), None]);

    let shifted = s.shift(2);
    assert_eq!(Vec::from(shifted.i32()?), &[None, None, Some(1)]);

    Ok(())
}
example();

fn fill_none(&self, _strategy: FillNoneStrategy) -> Result<Series>

Replace None values with one of the following strategies:

Forward fill (replace None with the previous value)

Backward fill (replace None with the next value)

Mean fill (replace None with the mean of the whole array)

Min fill (replace None with the minimum of the whole array)

Max fill (replace None with the maximum of the whole array)

NOTE: If you want to fill the Nones with a value use the fill_none operation on ChunkedArray<T>.

Example

fn example() -> Result<()> {
    let s = Series::new("some_missing", &[Some(1), None, Some(2)]);

    let filled = s.fill_none(FillNoneStrategy::Forward)?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(1), Some(2)]);

    let filled = s.fill_none(FillNoneStrategy::Backward)?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(2), Some(2)]);

    let filled = s.fill_none(FillNoneStrategy::Min)?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(1), Some(2)]);

    let filled = s.fill_none(FillNoneStrategy::Max)?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(2), Some(2)]);

    let filled = s.fill_none(FillNoneStrategy::Mean)?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(1), Some(2)]);

    Ok(())
}
example();

fn zip_with(&self, _mask: &BooleanChunked, _other: &Series) -> Result<Series>

Create a new ChunkedArray with values from self where the mask evaluates true and values from other where the mask evaluates false

fn sum_as_series(&self) -> Series

Get the sum of the Series as a new Series of length 1.

fn max_as_series(&self) -> Series

Get the max of the Series as a new Series of length 1.

fn min_as_series(&self) -> Series

Get the min of the Series as a new Series of length 1.

fn mean_as_series(&self) -> Series

Get the mean of the Series as a new Series of length 1.

fn median_as_series(&self) -> Series

Get the median of the Series as a new Series of length 1.

fn var_as_series(&self) -> Series

Get the variance of the Series as a new Series of length 1.

fn std_as_series(&self) -> Series

Get the standard deviation of the Series as a new Series of length 1.

fn quantile_as_series(&self, _quantile: f64) -> Result<Series>

Get the quantile of the ChunkedArray as a new Series of length 1.

fn rolling_mean(
&self,
_window_size: usize,
_weight: Option<&[f64]>,
_ignore_null: bool
) -> Result<Series>

Apply a rolling mean to a Series. See: ChunkedArray::rolling_mean.

fn rolling_sum(
&self,
_window_size: usize,
_weight: Option<&[f64]>,
_ignore_null: bool
) -> Result<Series>

Apply a rolling sum to a Series. See: ChunkedArray::rolling_mean.

fn rolling_min(
&self,
_window_size: usize,
_weight: Option<&[f64]>,
_ignore_null: bool
) -> Result<Series>

Apply a rolling min to a Series. See: ChunkedArray::rolling_mean.

fn rolling_max(
&self,
_window_size: usize,
_weight: Option<&[f64]>,
_ignore_null: bool
) -> Result<Series>

Apply a rolling max to a Series. See: ChunkedArray::rolling_mean.

fn fmt_list(&self) -> String

fn hour(&self) -> Result<Series>

Extract hour from underlying NaiveDateTime representation. Returns the hour number from 0 to 23.

fn minute(&self) -> Result<Series>

Extract minute from underlying NaiveDateTime representation. Returns the minute number from 0 to 59.

fn second(&self) -> Result<Series>

Extract second from underlying NaiveDateTime representation. Returns the second number from 0 to 59.

fn nanosecond(&self) -> Result<Series>

Extract second from underlying NaiveDateTime representation. Returns the number of nanoseconds since the whole non-leap second. The range from 1,000,000,000 to 1,999,999,999 represents the leap second.

fn day(&self) -> Result<Series>

Extract day from underlying NaiveDateTime representation. Returns the day of month starting from 1.

The return value ranges from 1 to 31. (The last day of month differs by months.)

fn ordinal_day(&self) -> Result<Series>

Returns the day of year starting from 1.

The return value ranges from 1 to 366. (The last day of year differs by years.)

fn month(&self) -> Result<Series>

Extract month from underlying NaiveDateTime representation. Returns the month number starting from 1.

The return value ranges from 1 to 12.

fn year(&self) -> Result<Series>

Extract month from underlying NaiveDateTime representation. Returns the year number in the calendar date.

fn datetime_str_fmt(&self, fmt: &str) -> Result<Series>

Format Date32/Date64 with a fmt rule. See chrono strftime/strptime.

fn clone_inner(&self) -> Arc<dyn SeriesTrait>

Clone inner ChunkedArray and wrap in a new Arc

fn get_as_any(&self, _index: usize) -> &dyn Any

Get the value at this index as a downcastable Any trait ref.

fn pow(&self, _exponent: f64) -> Result<Series>

Raise a numeric series to the power of exponent.

Loading content...
Implementations

impl<'a> dyn SeriesTrait + 'a

pub fn unpack<N: 'static>(&self) -> Result<&ChunkedArray<N>> where
N: PolarsDataType,