Module polars_core::utils

Re-exports

pub use arrow;

pub use rayon;

Modules

flatten

Structs

NoNull

Just a wrapper structure. Useful for certain impl specializations This is for instance use to implement impl<T> FromIterator<T::Native> for NoNull<ChunkedArray<T>> as Option<T::Native> was already implemented: impl<T> FromIterator<Option<T::Native>> for ChunkedArray<T>

TrustMyLength

Wrap

Traits

Arg

CustomIterTools

CustomIterToolsSized

FromTrustedLenIterator

IntoVec

Functions

_set_partition_size

_to_physical_and_bit_repr

Transform to physical type and coerce similar sized integer to a bit representation to reduce compiler bloat

accumulate_dataframes_horizontal

accumulate_dataframes_vertical

This takes ownership of the DataFrame so that drop is called earlier.

accumulate_dataframes_vertical_unchecked

This takes ownership of the DataFrame so that drop is called earlier. Does not check if schema is correct

accumulate_dataframes_vertical_unchecked_optional

align_chunks_binary

Ensure the chunks in both ChunkedArrays have the same length.

align_chunks_ternary

Panics

any_values_to_supertype

Determine the supertype of a collection of AnyValue.

any_values_to_supertype_and_n_dtypes

Determine the supertype and the number of unique data types of a collection of AnyValue.

coalesce_nulls

ensure that nulls are propagated to both arrays

coalesce_nulls_series

concat_df

Concat the DataFrames to a single DataFrame.

concat_df_unchecked

Concat the DataFrames to a single DataFrame.

dtypes_to_schema

Convert a collection of DataType into a schema.

dtypes_to_supertype

Given multiple data types, determine the data type that all types can safely be cast to.

ensure_sorted_arg

get_supertype

Given two data types, determine the data type that both types can safely be cast to.

get_time_units

handle_casting_failures

slice_slice

split_ca

split_df_as_ref

try_get_supertype

Given two data types, determine the data type that both types can safely be cast to.

with_unstable_series

A utility that allocates an UnstableSeries. The applied function can then use that series container to save heap allocations and swap arrow arrays.