polars_core

Module utils

Re-exports

pub use arrow;

pub use rayon;

Modules

flatten

Structs

NoNull

Just a wrapper structure which is useful for certain impl specializations.

SuperTypeFlags

SuperTypeOptions

TrustMyLength

Wrap

Traits

Container

CustomIterTools

CustomIterToolsSized

FromTrustedLenIterator

IntoVec

Convenience for x.into_iter().map(Into::into).collect() using an into_vec() function.

Functions

_set_partition_size

accumulate_dataframes_horizontal

accumulate_dataframes_vertical

This takes ownership of the DataFrame so that drop is called earlier.

accumulate_dataframes_vertical_unchecked

This takes ownership of the DataFrame so that drop is called earlier. Does not check if schema is correct

accumulate_dataframes_vertical_unchecked_optional

align_chunks_binary

Ensure the chunks in both ChunkedArrays have the same length.

align_chunks_binary_ca_series

Ensure the chunks in ChunkedArray and Series have the same length.

align_chunks_ternary

Panics

any_values_to_supertype

Determine the supertype of a collection of AnyValue.

any_values_to_supertype_and_n_dtypes

Determine the supertype and the number of unique data types of a collection of AnyValue.

binary_concatenate_validities

coalesce_nulls

ensure that nulls are propagated to both arrays

coalesce_nulls_columns

concat_df

Concat the DataFrames to a single DataFrame.

concat_df_unchecked

Concat the DataFrames to a single DataFrame.

dtypes_to_schema

Convert a collection of DataType into a schema.

dtypes_to_supertype

Given multiple data types, determine the data type that all types can safely be cast to.

first_non_null

get_numeric_upcast_supertype_lossless

Returns a numeric supertype that l and r can be safely upcasted to if it exists.

get_supertype

get_supertype_with_options

Given two data types, determine the data type that both types can safely be cast to.

get_time_units

handle_casting_failures

is_deprecated_cast

last_non_null

materialize_dyn_int

merge_dtypes_many

operation_exceeded_idxsize_msg

slice_offsets

slice_slice

split

Splits, but doesn’t flatten chunks. E.g. a container can still have multiple chunks.

split_and_flatten

Split a Container in target elements. The target doesn’t have to be respected if not Deviation of the target might be done to create more equal size chunks.

split_df_as_ref

Split a DataFrame in target elements. The target doesn’t have to be respected if not strict. Deviation of the target might be done to create more equal size chunks.

try_get_supertype

Given two data types, determine the data type that both types can safely be cast to.

try_get_supertype_with_options

with_unstable_series

A utility that allocates an AmortSeries. The applied function can then use that series container to save heap allocations and swap arrow arrays.